==> 2_Taxon_Check.Rmd <==
---
title: "Generate taxon groups"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---

# 1 LOAD
```{r clear, echo=F}
rm(list = ls())
```



```{r, echo=F, warning=F, message=FALSE}
#Load required packages
library(readr)
library(tidyr)
library(dplyr)
library(magrittr)
library(ggplot2)
library(stringr)
library(plantlist)
library(vegdata)
```


# 2 DATA
Here we load the data and chose which dataset we want to use.  
We later remove all rows without a plot ID and check which taxa are present in the dataset
```{r Load data, message= F, warning=FALSE}
fast <- T
if(fast) {
  eva <- read_delim("../EVA Data/onePromilleEva.csv", "\t")
} else {
  eva <- read_delim("../EVA Data/171_NeophyteInvasions20230216_notJUICE_species.csv", "\t")
}

# Filter out all rows which don't have a plot ID
eva <- eva[!is.na(eva$PlotObservationID),]

# Which taxon groups are there in EVA?
unique(eva$`Taxon group`)
```

# 3 KINGDOM
## 3.1 Genus
Make a new column with only the genus' name.  
```{r Matched concept,  message= F, warning=FALSE}
# Fast method to extract the first string from the 'Matched concept' column e.g. Genus
# \\1 says to use the first part (ie word) as input for the new column
# []+* is used to take all subsequent letters (with capitals and small letters)
# gsub is the function to replace the text (designated in the first part) with the second part (unless it is like here, then we just take that part) from the column designated last
eva$Genus <- gsub("([A-Za-z]+).*", "\\1", eva$`Matched concept`)

# This worked, all plant observations have a genus
# sum(is.na(eva$Genus))
```


## 3.2 Functions
We first create three functions required further in the computation. They do the following:  

1.    Get the ID for each taxon group
2.    Get the name of the taxon group, if it is unique it renders the correct one, if not it displays NA. We make sure that the NA's and Unknowns are not given as unique classes
3.    Give the number of options for the taxon

```{r Functions, message= F, warning=FALSE}
# get ID
taxonGroupToId <- function(taxonGroup) {
  eva$`Taxon group ID`[match(taxonGroup, eva$`Taxon group`)]
}

# get name kingdom group, if there are multiple return NA
computeTaxonGroup <- function(possibleTaxonGroups) {
  # get all possible groups for that specific genus
  uniqueTaxonGroups <- unique(possibleTaxonGroups[!(is.na(possibleTaxonGroups) | possibleTaxonGroups == "Unknown")])
  # if more than 1 return NA because one must be incorrect
  if(length(uniqueTaxonGroups) == 1) {
    uniqueTaxonGroups[1]
  } else {
    NA
  }
}

# get number of different taxa
numberOfDifferentTaxonGroups <- function(possibleTaxonGroups) {
  # count number of possible groups
  length(unique(possibleTaxonGroups[!(is.na(possibleTaxonGroups) | possibleTaxonGroups == "Unknown")]))
}
```


## 3.3 Wrong
We create a matrix with the taxon group for each genus.  
We later show all geni with multiple taxa, i.e. there is a wrong taxon in Eva. We combine all erroneous inputs in a dataframe per genus and taxon group.
```{r, message= F, warning=FALSE }
# For every genus: if all taxa with that genus have the same taxon group, give it that taxon group. Else give it NA. 
# Furthermore: for every genus count the number of different taxon groups found in EVA for that genus.
taxonGroupForGeni <- eva |> 
  # group eva per genus
  group_by(Genus) |> 
  # summarize by computing for each genus the group, if there are multiple return NA and give number
  summarise(
    TaxonGroup = computeTaxonGroup(`Taxon group`), # Unique =name, else= NA
    NumberOfDifferentTaxonGroups = numberOfDifferentTaxonGroups(`Taxon group`)) # Number of options per genus

# Show the geni for which we have multiple taxon groups. E.g. there must be erroneous taxon groups in EVA.
taxonGroupForGeni[taxonGroupForGeni$NumberOfDifferentTaxonGroups > 1,]

# All geni with more than one taxon group assignment
# Checked for correctness
genusTaxonGroupCorrections <- data.frame(Genus = c("Hepatica", "Lamprothamnium", "Leptorhaphis", "Mycoblastus", "Peltaria"),
                                       TaxonGroup = c("Vascular plant", "Alga", "Lichen", "Lichen", "Vascular plant"))
```


## 3.4 Assign wrong
We combine all genus that need a new taxon group.  
First we make a function to obtain a vector with TRUE if taxon starts with a specified genus.  
We then make a new dataframe and store in this one the matched concept and old and new taxon.
```{r, message= F, warning=FALSE}
# Function to obtain a vector with TRUE if taxon starts with a specified genus
hasGenus <- function(taxa, genus) {
 startsWith(taxa, paste(genusTaxonGroupCorrections$Genus[i], " ", sep = ""))
  # check if the taxa (given) stars with the same words as the name of the genus in the correction file
  # paste is used to ensure that the genus name is one word
}

# Compute all taxa which need to be reassigned to a different taxon group and store the result to file
# fist an empty file
taxaTaxonGroupCorrections <- data.frame(`Matched concept` = character(0), NewTaxonGroup = character(0), OldTaxonGroup = character(0))

for(i in seq(1,length(genusTaxonGroupCorrections$Genus))) {
  #make vector with T when the genus in eva corresponds to a genus that must be corrected and has the wrong taxon group
  selection <- eva$Genus == genusTaxonGroupCorrections$Genus[i] & eva$`Taxon group` != genusTaxonGroupCorrections$TaxonGroup[i] 
  # count number of erroneous observations per genus and taxon group
  # give the number of species per group of the same matched concept and taxon group
  erroneousTaxa <- eva[selection,] |> group_by(`Matched concept`, `Taxon group`) |> summarise(n = n())
  # Add all taxa with correction to the final result
  # This is a file that we can use, containing the name, the taxon group we just assigned and the old one
  taxaTaxonGroupCorrections <- add_row(taxaTaxonGroupCorrections, #Dataframe
          Matched.concept = erroneousTaxa$`Matched concept`,  #genus
          NewTaxonGroup = rep(genusTaxonGroupCorrections$TaxonGroup[i],times=length(erroneousTaxa$`Matched concept`)),
          OldTaxonGroup = erroneousTaxa$`Taxon group`) 
}
```


We store the results in a new file
```{r}
#write_csv(taxaTaxonGroupCorrections, "taxa_taxon_group_corrections.csv")
```



## 3.5 Missing
We first load the data and check uniqueness of taxa
```{r, message= F, warning=FALSE}
# Unknown kingdom
taxonGroupForGeni <- read_csv("../Intermediate Data/taxon_group_for_unassigned_geni.csv")

# check what we will actually change
x<-eva[which(eva$Genus%in%taxonGroupForGeni$Genus),c(1,3,6,12)]
# which taxa are in our dataset
unique(taxonGroupForGeni$TaxonGroup)
unique(x$`Taxon group`)
```



Check all taxa that need to be assigned and save to file.
First create empty dataset.
Then we compute the names of all species for which the genus did not have a taxon and later assign it based on our dataset
```{r, message= F, warning=FALSE}
# Compute all taxa which need to be assigned to a taxon group and store the result to file
taxaTaxonGroupAssignments <- data.frame(`Matched concept` = character(0), TaxonGroup = character(0))

for(i in seq(1,length(taxonGroupForGeni$Genus))) {
  # Select all geni equal to the ones in the dataset
  selection <- eva$Genus == taxonGroupForGeni$Genus[i]
  # Take names of all species of a specific genus that did not have the taxon yet
  taxaToBeAssigned <- unique(eva$`Matched concept`[selection])
  # Add all taxa which need to be assigned to the final result
  taxaTaxonGroupAssignments <- add_row(taxaTaxonGroupAssignments,
                                       Matched.concept = taxaToBeAssigned,
                                       TaxonGroup = taxonGroupForGeni$TaxonGroup[i])
}
```

We store the results in a new file
```{r}
#write_csv(taxaTaxonGroupAssignments, "taxa_taxon_group_assignments.csv")
```


## 3.6 Eva assign
```{r}
# WRONG DATA
# What is the location of the species in eva that are specified incorrectly
index <- eva$`Matched concept` %in% taxaTaxonGroupCorrections$Matched.concept

# Change to correct for each species
eva$`Taxon group`[index] <- taxaTaxonGroupCorrections$NewTaxonGroup[match(eva$`Matched concept`[index], taxaTaxonGroupCorrections$Matched.concept)]

# MISSING DATA
index <- eva$`Matched concept` %in% taxaTaxonGroupAssignments$Matched.concept
eva$`Taxon group`[index] <- taxaTaxonGroupAssignments$TaxonGroup[match(eva$`Matched concept`[index], taxaTaxonGroupAssignments$Matched.concept)]
# Some taxon names seem incorrect
taxonNameCorrections <- read_csv("../Neophyte Assignments/UniqueTaxaEurope-2023-04-23.csv", show_col_types = FALSE)

# give the species name from taxon name corrections to eva in the species column
# to lower makes sure all data is written without capitals
# match gives vector with position of argument eva in taxonNameCorrections, and does this for the total length of eva
eva$species <- taxonNameCorrections$species[match(tolower(eva$`Matched concept`), tolower(taxonNameCorrections$Matched.concept))]
```


# 4 TPL
## 4.1 Prepare data
We load the data
```{r , message=FALSE, warning=FALSE}
fast <- F
if(fast) {
  eva <- read_delim("../EVA Data/onePromilleEva.csv", "\t")
} else {
  eva <- read_delim("../EVA Data/171_NeophyteInvasions20230216_notJUICE_species.csv", "\t")
}

# Filter out all rows which don't have a plot ID
eva <- eva[!is.na(eva$PlotObservationID),]
```


We update the taxa data
```{r, warning=F, message=FALSE}
# Update taxon group for some species where it was specified incorrectly
taxonGroupCorrections <- read_csv("../Intermediate Data/taxa_taxon_group_corrections.csv")

# What is the location of the species in eva that are specified incorrectly
index <- eva$`Matched concept` %in% taxonGroupCorrections$Matched.concept

# Change to correct for each species
eva$`Taxon group`[index] <- taxonGroupCorrections$NewTaxonGroup[match(eva$`Matched concept`[index], taxonGroupCorrections$Matched.concept)]


# Update taxon group for species that did not have one in EVA
taxonGroupAssignments <- read_csv("../Intermediate Data/taxa_taxon_group_assignments.csv")
index <- eva$`Matched concept` %in% taxonGroupAssignments$Matched.concept
eva$`Taxon group`[index] <- taxonGroupAssignments$TaxonGroup[match(eva$`Matched concept`[index], taxonGroupAssignments$Matched.concept)]

taxonGroupAssignments <- taxonGroupAssignments[taxonGroupAssignments$TaxonGroup=="Unknown",]


# Some taxon names seem incorrect
taxonNameCorrections <- read_csv("../Neophyte Assignments/UniqueTaxaEurope-2023-04-23.csv")

# give the species name from taxon name corrections to eva in the species column
# to lower makes sure all data is written without capitals
# match gives vector with position of argument eva in taxonNameCorrections, and does this for the total length of eva
eva$species <- taxonNameCorrections$species[match(tolower(eva$`Matched concept`), tolower(taxonNameCorrections$Matched.concept))]

# some data of species is.na, we assign for these the values of the matched concept
eva$species[is.na(eva$species)] <- eva$`Matched concept`[is.na(eva$species)] 
```


```{r}
unknown <- eva[eva$`Taxon group`=="Unknown",]
plant_list <- TPL(unknown$species)
unique(plant_list$GROUP)
plant_list$GROUP <- gsub("Angiosperms","Vascular plant",plant_list$GROUP)
plant_list$GROUP <- gsub("Gymnosperms","Vascular plant",plant_list$GROUP)

eva$`Taxon group`[eva$`Taxon group`=="Unknown"] <- plant_list$GROUP[match(eva$species[eva$`Taxon group`=="Unknown"], plant_list$YOUR_SEARCH)]
unique(eva$`Taxon group`)
```




We only keep vascular plants and remove the columns in which this data is designated.
```{r}
eva <- eva[eva$`Taxon group` == "Vascular plant", names(eva) %in% c("PlotObservationID","Matched concept", "Layer","Cover %","species")]
eva <- eva[!is.na(eva$species),]
```


## 4.2 Status 
We first get the data for genus level interpretation
```{r}
# get genus --> very slow
# eva$genus <- gsub("([A-Za-z]+).*", "\\1", eva$species)

# get all unique names in eva
eva_names <- unique(eva$species)
# make dataframe and give name
eva_names <- as.data.frame(eva_names)
colnames(eva_names)[1] <- "species"
# get genus data
eva_names$genus <- gsub("([A-Za-z]+).*", "\\1", eva_names$species)

# assign to eva
eva$genus <- eva_names$genus[match(eva$species, eva_names$species)]

# get new list of unique species
eva_names <- unique(eva$species)

# we remove the aggr part
#eva_names <- gsub(" aggr.", "", eva_names)
#eva$species <- gsub(" aggr.", "", eva$species)
# some duplicates now --> remove
eva_names <- unique(eva_names)


# we will check names here quickly
status_already <- read.csv("TPL_status_all_new.csv")
status_already <- status_already[, -1]
status_already <- status_already[!duplicated(status_already),]

# names to calculate
eva_status <- eva_names[!eva_names %in% status_already$SCIENTIFIC_NAME]

# status calculation
status <- status(eva_status)
# remove dubble entries
status <- status[!duplicated(status),]

status <- rbind(status_already, status)
status <- status[!duplicated(status),]

#status<- status_already

if(fast){
  status<- status[status$SCIENTIFIC_NAME %in% eva_names,]
}

rm(list=c("status_already"))
#write.csv(status, "TPL_status_all.csv")
```


## 4.3 Remove duplicates
```{r}
# change names to accepted
accept_dup<- data.frame(species= character(), family= character(), old_name= character())
# run for all species
for(i in 1: length(eva_names)){ 
  # first check if they exist in our status data, if not keep only old name
  if(any(is.na(status$STATUS[status$SCIENTIFIC_NAME== eva_names[i]]))){
    accept_dup <- add_row(accept_dup,
                      species= NA, 
                      family= NA,
                      old_name= eva_names[i])
  # then check if they exist in our status data with an accepted name
  } else if(any(status$STATUS[status$SCIENTIFIC_NAME==eva_names[i]]=="Accepted")){
    accept_dup <- add_row(accept_dup,
                      species= status$ACCEPTED_SPECIES[status$STATUS=="Accepted" & status$SCIENTIFIC_NAME==eva_names[i]], 
                      family= status$FAMILY[status$STATUS=="Accepted" & status$SCIENTIFIC_NAME==eva_names[i]],
                      old_name= eva_names[i])
  # check for species that are present only once and are designated as 'synonym'
  } else if(all(!any(status$STATUS[status$SCIENTIFIC_NAME==eva_names[i]]=="Accepted"),
            length(status$STATUS[status$SCIENTIFIC_NAME==eva_names[i]])==1 &
            status$STATUS[status$SCIENTIFIC_NAME==eva_names[i]]=="Synonym")){
    accept_dup <- add_row(accept_dup,
                      species= status$ACCEPTED_SPECIES[status$STATUS=="Synonym" & status$SCIENTIFIC_NAME==eva_names[i]], 
                      family= status$FAMILY[status$STATUS=="Synonym" & status$SCIENTIFIC_NAME==eva_names[i]],
                      old_name= eva_names[i]) 
  # if no other options, give NA  
  }  else {
   accept_dup <- add_row(accept_dup,
                      species= NA, 
                      family= NA,
                      old_name= eva_names[i])
  }
}

# make data frame
accept_dup<- as.data.frame(accept_dup)
# check if there are still duplicated species in the old name list --> no
dup<- accept_dup[accept_dup$old_name %in% accept_dup$species[duplicated(accept_dup$old_name)],]

# count number of species without a species name and retrieve them
sum(is.na(accept_dup$species))
accept_dup_na <- accept_dup[is.na(accept_dup$species),]
# check how much observations these are
length(eva$PlotObservationID[eva$species %in% accept_dup_na$old_name])/ length(eva$PlotObservationID)


# also retrieve all status values for these species
status_na <- status[status$SCIENTIFIC_NAME %in% accept_dup_na$old_name, ]

# look which species are in eva but not in the accept_dup list
eva_names[!eva_names %in% accept_dup$old_name]
# this list are species that are also not present in the status list
eva_names[!eva_names %in% status$SCIENTIFIC_NAME]
# and there are also species in status that are not in eva but in status
status[!status$SCIENTIFIC_NAME %in% eva_names,]

# we create a list with crop species that can be removed
crops <- c("crop rye","crop spelt","crop barley","crop potato", "crop wheat", "crop vineyard", "crop radish","crop lettuce","crop gladiolus",
           "crop salsify","crop zucchini","crop parsley","crop clover","crop tobacco", "crop maize","crop rye","crop canola","crop wheat", 
           "crop potato", "crop oat","crop pea","crop onion","crop asparagus" ,"crop tomato","crop strawberry" ,"crop carrot" ,"crop cabbage",
           "crop spelt" ,"crop rhubarb", "crop hops","crop beet","crop flax","crop bean","crop digitalis", "crop celery" ,"crop currant",
           "crop buckwheat","crop millet")
# eva_names <- eva_names[!eva_names %in% crops]
# eva <- eva[!eva$species %in% crops,]

# check length of number of occurences of these species
length(eva$PlotObservationID[eva$species %in% eva_names[!eva_names %in% status$SCIENTIFIC_NAME]])/ length(eva$PlotObservationID)

# we will just add them to the existing list
accept_dup <- add_row(accept_dup, 
                      species= NA, 
                      family= NA,
                      old_name= eva_names[!eva_names %in% status$SCIENTIFIC_NAME])


# we check whether we now have every species in eva in our new list
all.equal(sort(accept_dup$old_name), sort(eva_names))
# count number of species without a species name
sum(is.na(accept_dup$family))

# add accepted species names and family
eva$new_species <- accept_dup$species[match(eva$species,accept_dup$old_name)]
eva$family <- accept_dup$family[match(eva$species,accept_dup$old_name)]

# list of all species in dataset
eva_family <- eva[!duplicated(eva$species),]
eva_family$species[!eva_family$species %in% eva_names]
# remove all species without a family name
genus_family <- eva_family[!is.na(eva_family$family),]
# get all unique geni with a family
genus_family <- genus_family[!duplicated(genus_family$genus),]


# match geni with families
eva$family[is.na(eva$family)] <- genus_family$family[match(eva$genus[is.na(eva$family)], genus_family$genus)]
```


## 4.4 TPL and own
Now we will use TPL
```{r}
# we still have some species left
eva_family <- eva[is.na(eva$family), ]
# we get all unique species
eva_family <- eva_family[!duplicated(eva_family$species),]


# TPL
plant_list <- TPL(eva_family$species)

# link with data
eva$family[is.na(eva$family)] <- plant_list$FAMILY[match(eva$species[is.na(eva$family)], plant_list$YOUR_SEARCH)]

# check species without family name
# we still have some species left
eva_family <- eva[is.na(eva$family),]
eva_family <- eva_family[!duplicated(eva_family$species),]


# we make a dataframe for the species
own_class <- data.frame(species= c("Pseudopodospermum hispanicum", 
                                   "Calamagrostis × baltica", 
                                   "Archanthemis marschalliana",
                                   "Pseudopodospermum elatum",
                                   "Willemetia stipitata",
                                   "Parthenocissus quinquefolia",
                                   "Elaeagnus angustifolia",
                                   "Astragalus angustifolius subsp. angustifolius",
                                   "Cyathophylla chlorifolia",
                                   "Anthemis cretica subsp. cretica",
                                   "Teucrium chamaedrys subsp. tauricola",
                                   "Arenaria pamphylica subsp. alpestris",
                                   "Achillea formosa subsp. formosa",
                                   "Pilosella piloselloides subsp. magyarica",
                                   "Centaurea idaea",
                                   "Torminalis glaberrima",
                                   "Ajuga",
                                   "Achillea sivasica",
                                   "Melanortocarya obtusifolia",
                                   "Achillea formosa subsp. amanica",
                                   "Archanthemis marschalliana subsp. pectinata",
                                   "Malus florentina",
                                   "Phrynella ortegioides",
                                   "Chaloupkaea chrysantha",
                                   "Dorystoechas hastata",
                                   "Archanthemis marschalliana subsp. sosnovskyana",
                                   "Xatardia scabra","Lolium",
                                   "Hormuzakia aggregata",
                                   "× Agropogon lutosus",
                                   "Resetnikia triquetra",
                                   "Lolium × elongatum",
                                   "Vahlodea atropurpurea",
                                   "Vogtia annua",
                                   "Aizoanthemopsis hispanica",
                                   "Arthrocaulon meridionalis",
                                   "Myriolimon ferulaceum",
                                   "Aloiampelos ciliaris",
                                   "Vogtia microphylla",
                                   "Lolium × holmbergii",
                                   "Parapholis cylindrica",
                                   "Anchusa variegata",
                                   "Phyllolepidum cyclocarpum",
                                   "Sinapis pubescens",
                                   "Lonicera kabylica",
                                   "Noccaea cypria",
                                   "Smyrnium olusatrum",
                                   "Anthericum",
                                   "Astartoseris triquetra",
                                   "Archanthemis fruticulosa",
                                   "Micromeria maderensis",
                                   "Malva phoenicea",
                                   "Heliotropium crispum",
                                   "Mairetis microsperma",
                                   "× Pseudorhiza nieschalkii",
                                   "Achillea millefolium subsp. collina",
                                   "× Dactylodenia st-quintini",
                                   "Limnanthes douglasii",
                                   "Myriolimon diffusum",
                                   "× Pucciphippsia vacillans",
                                   "Sphenopus divaricatus",
                                   "Hormuzakia aggregata"),
                           family= c("Asteraceae", "Poaceae",
                                     "Asteraceae","Asteraceae",
                                     "Asteraceae", "Vitaceae",
                                     "Elaeagnaceae","Fabaceae",
                                     "Caryophyllaceae","Asteraceae",
                                     "Lamiaceae","Caryophyllaceae",
                                     "Asteraceae","Asteraceae",
                                     "Asteraceae","Rosaceae",
                                     "Lamiaceae","Asteraceae",
                                     "Boraginaceae","Asteraceae",
                                     "Asteraceae","Rosaceae",
                                     "Caryophyllaceae","Crassulaceae",
                                     "Lamiaceae","Asteraceae",
                                     "Apiaceae","Poaceae",
                                     "Boraginaceae","Poaceae",
                                     "Brassicaceae","Poaceae",
                                     "Poaceae","Asteraceae",
                                     "Aizoaceae","Amaranthaceae",
                                     "Plumbaginaceae","Asphodelaceae",
                                     "Asteraceae","Poaceae",
                                     "Poaceae","Boraginaceae",
                                     "Brassicaceae","Brassicaceae",
                                     "Caprifoliaceae", "Brassicaceae",
                                     "Apiaceae","Asparagaceae",
                                     "Asteraceae","Asteraceae",
                                     "Lamiaceae","Malvaceae",
                                     "Boraginaceae","Boraginaceae",
                                     "Orchidaceae","Asteraceae",
                                     "Orchidaceae","Limnanthaceae",
                                     "Plumbaginaceae","Poaceae",
                                     "Poaceae","Boraginaceae"),
                           old_name=c("Pseudopodospermum hispanicum",
                                      "x_Calammophila baltica", 
                                      "Archanthemis marschalliana",
                                      "Pseudopodospermum elatum",
                                      "Calycocorsusástipitatus* species",
                                      "Partenocissus quinquefolia",
                                      "Elaeagnos angustifolia",
                                      "Astragals angustifolius subsp. angustifolius",
                                      "Saponasia chlorifolia",
                                      "Antemis cretica subsp. pontica",
                                      "Teuchrium chamaedrys subsp. tauricola",
                                      "Arenari pamphylica subsp. alpestris",
                                      "Leucocylus formosus subsp. formosus",
                                      "Pllosella piloselloides subsp. megalomastix",
                                      "Cenaturea idaea",
                                      "Srbus torminalis var. orientalis",
                                      "Chamaepitys",
                                      "Chrysocamela noeana",
                                      "Melanortocarya obtusifolia",
                                      "Leucocyclus formosus subsp. amanicus",
                                      "Archanthemis marschalliana subsp. pectinata",
                                      "x_Malosorbus florentina",
                                      "Phrynella ortegioides",
                                      "Chaloupkaea chrysantha",
                                      "Dorystoechas hastata",
                                      "Archanthemis marschalliana subsp. sosnovskyana",
                                      "Xatartia scabra",
                                      "x_Schedolium",
                                      "Hormuzakia aggregata",
                                      "x_Agropogon littoralis",
                                      "Resetnikia triquetra",
                                      "x_Schedolium loliaceum",
                                      "Vahlodea atropurpurea",
                                      "Vogtia annua",
                                      "Aizoanthemopsis hispanica",
                                      "Arthrocaulon meridionalis",
                                      "Myriolimon ferulaceum",
                                      "Aloiampelos ciliaris",
                                      "Vogtia microphylla",
                                      "x_Schedolium holmbergii",
                                      "Haynardia cylindrica",
                                      "Anchusella variegata",
                                      "Phyllolepidum cyclocarpum subsp. cyclocarpum",
                                      "Sinapsis pubescens",
                                      "Lonciera kabylica",
                                      "Thlapsi cyprium",
                                      "Smirnium oluzatrum",
                                      "Phalangium species",
                                      "Astartoseris triquetra",
                                      "Archanthemis fruticulosa",
                                      "Mycromeria thymoides",
                                      "Navaea phoenicia",
                                      "Heiotropium erosum var. prostratum",
                                      "Mairetis microsperma",
                                      "x-Pseudorhiza nieschalkii",
                                      "Achilea millefolium subsp. collina f. rubriflora",
                                      "x-Dactylodenia st-quintini",
                                      "Limnanthus douglasii",
                                      "Myriolimon diffusum",
                                      "x-Pucciphippsia vacillans",
                                      "Sphaenopus divaricatus",
                                      "Hormuzakiagata"
                                      ))



remove <- c("Eontodon concinnus","Guadinopsis macra subsp. macra","Chelianthes fragrans","Avenochloa cycladica",
            "Dycot spec","Keimling species 1","Keimling species 2","Chinaschilf* species","Sturmia species",
            "Magnoliophyta species", "Tracheophyta species","Planta indeterminate","Pteridophyta species",
            "Polypodiopsida species","Chenopodiaceae","Spermatophyta species","Liliopsida","Magnoliopsida species",
            "Gefaesspflanze species","Hypanthium glabre","Kiemplant species","Dipsacaceae","Filicopsida species",
            "Liliopsida species", "Cephalophysis species", "Dicot sp.1","Dicot sp.2","Gras species", 
            "Hakvrucht cultuurgewas", "Graan cultuurgewas","Crucifer species","Rubiales species",
            "Dicotyledone species","Monocotyledonae* species","Balansae glaberrima","Poales species",
            "Microspermae species","Pinopsida species","Pteropsida species","crop millet","Liana indet.",
            "Rubiacee species","Ombellifere species","Myriophorum insulare","Decumbens species","Seedlings of",
            "Cyclotella comta","Dichodonti pellucidum var. flavescens","Teilingia granulata","Kale bodem",
            "Dicotyledones species","Thorea ramossissima","Plant onbekend","Lygopus species","Onchophorus virens"
            )

unclear <- c("Cardueae","x_Sorbaronia mitschurinii","x_Triticosecale","× Triticosecale rimpaui",
             "Mycelis tinctoria","x_Triticosecale rimpaui","Allopecurus littoralis","Echilops ovata",
             "Cirsiun acanthoides","Nardurus myosuroides","Thymelae garganica",
             "Acantholimum pulverulentum var. pulverulentum","Poligonum piperitum","x-Festulpia hubbardii",
             "Phyllirea","Reamuria vermiculata")


# we add our own classification
eva$new_species[eva$species %in% own_class$old_name] <- 
  own_class$species[match(eva$species[eva$species %in% own_class$old_name], own_class$old_name)]
# family
eva$family[eva$species %in% own_class$old_name] <- 
  own_class$family[match(eva$species[eva$species %in% own_class$old_name], own_class$old_name)]

# remove the species
eva<- eva[!eva$species %in% remove,]
eva<- eva[!eva$species %in% unclear,]


# we check whether we still have family data left
eva[is.na(eva$family),]
unique(eva$species[is.na(eva$family)])
any(is.na(eva$family))
```





## 4.5 Official names
```{r}
# get all unique names in eva
eva_names <- unique(eva$new_species)
# make dataframe and give name
eva_names <- as.data.frame(eva_names)
colnames(eva_names)[1] <- "species"
# get genus data
eva_names$genus <- gsub("([A-Za-z]+).*", "\\1", eva_names$species)

# assign to eva
eva$new_genus <- eva_names$genus[match(eva$new_species, eva_names$species)]

# get new list of unique species
eva_names <- unique(eva$species)


# make official species list with also genus and family
eva$official <- ifelse(!is.na(eva$new_species), eva$new_species, eva$species)
eva$official_genus <- ifelse(!is.na(eva$new_species), eva$new_genus, eva$genus)

# new data set with different order
new_eva <- eva[, c(1,10,11, 8, 3,4 , 5)]
colnames(new_eva)[1:7]<- c("PlotObservationID","species","genus","family","layer","cover","old name")

# check difference in number of species
length(unique(new_eva$species))
length(unique(new_eva$`old name`))

new_eva

# write csv
# write_csv(new_eva, "eva_accepted_names_all.csv")
```


# 5 EURO+MED
## 5.1 Load database
```{r}
library(RSQLite)

con <- dbConnect(drv=RSQLite::SQLite(), dbname="C:/Users/u0166342/Documents/Boeren/Impact/eva_neophytes/EIVE Data/EuroSL.sqlite")

## list all tables
tables <- dbListTables(con)
tables <- tables[tables != "sqlite_sequence"]

lDataFrames <- vector("list", length=length(tables))
for (i in seq(along=tables)) {
  lDataFrames[[i]] <- dbGetQuery(conn=con, statement=paste("SELECT * FROM '", tables[[i]], "'", sep=""))
}

Euro <- lDataFrames[[1]]
```


## 5.2 Prepare data
We load the data
```{r , message=FALSE, warning=FALSE}
fast <- F
if(fast) {
  eva <- read_delim("../EVA Data/onePromilleEva.csv", "\t", show_col_types = FALSE)
} else {
  eva <- read_delim("../EVA Data/171_NeophyteInvasions20230216_notJUICE_species.csv", "\t", show_col_types = FALSE)
}

# Filter out all rows which don't have a plot ID
eva <- eva[!is.na(eva$PlotObservationID),]
```


We update the taxa data
```{r, warning=F, message=FALSE}
# Update taxon group for some species where it was specified incorrectly
taxonGroupCorrections <- read_csv("../Intermediate Data/taxa_taxon_group_corrections.csv")

# What is the location of the species in eva that are specified incorrectly
index <- eva$`Matched concept` %in% taxonGroupCorrections$Matched.concept

# Change to correct for each species
eva$`Taxon group`[index] <- taxonGroupCorrections$NewTaxonGroup[match(eva$`Matched concept`[index], taxonGroupCorrections$Matched.concept)]


# Update taxon group for species that did not have one in EVA
taxonGroupAssignments <- read_csv("../Intermediate Data/taxa_taxon_group_assignments.csv")
index <- eva$`Matched concept` %in% taxonGroupAssignments$Matched.concept
eva$`Taxon group`[index] <- taxonGroupAssignments$TaxonGroup[match(eva$`Matched concept`[index], taxonGroupAssignments$Matched.concept)]

#taxonGroupAssignments <- taxonGroupAssignments[taxonGroupAssignments$TaxonGroup=="Unknown",]
```


```{r, warning=F, message=FALSE}
# Some taxon names seem incorrect
taxonNameCorrections <- read_csv("../Neophyte-Impact/Neophyte Assignments/UniqueTaxaEurope-2023-04-23.csv")

# give the species name from taxon name corrections to eva in the species column
# to lower makes sure all data is written without capitals
# match gives vector with position of argument eva in taxonNameCorrections, and does this for the total length of eva
eva$species <- taxonNameCorrections$species[match(tolower(eva$`Matched concept`), tolower(taxonNameCorrections$Matched.concept))]

# get all unique species eva data
test <- eva[!is.na(eva$species),]
# remove duplicates of the matched concept
test <- test[!duplicated(test$`Matched concept`),]

# get names eva
eva_names <- as.data.frame(unique(eva$`Matched concept`))
colnames(eva_names)<- "species"

# add taxon names and euro matches
eva_names$taxon <- eva$`Taxon group`[match( eva_names$species, eva$`Matched concept`)]
eva_names$euro <- Euro$TaxonConcept[match(vegdata::taxname.abbr(eva_names$species), vegdata::taxname.abbr(Euro$TaxonName))]
length(unique(eva_names$euro))
#eva_names <- eva_names[!(eva_names$taxon=="Moss"|eva_names$taxon=="Lichen"), ]
eva_names <- eva_names[!is.na(eva_names$euro),]

# make new dataframes, one with species in both Irena (test) and euro (eva_names) databases
test2 <-test$species[test$`Matched concept` %in% eva_names$species]
# one with data in euro and Irena
test3 <-eva_names[eva_names$species %in% test$`Matched concept`,]

# give all the data of euromed
control <- test3[(test2 != test3$euro),]

# change in eva
eva$species[eva$`Matched concept` %in% control$species]<-
  control$euro[match(eva$`Matched concept`[eva$`Matched concept` %in% control$species],control$species)]


# add other data from euro+med
eva_names <- as.data.frame(unique(eva$`Matched concept`[is.na(eva$species)]))
colnames(eva_names)<- "species"
eva_names$euro <- Euro$TaxonConcept[match(vegdata::taxname.abbr(eva_names$species), vegdata::taxname.abbr(Euro$TaxonName))]
eva_names$taxon <- eva$`Taxon group`[match( eva_names$species, eva$`Matched concept`)]
sum(!is.na(eva_names$euro[eva_names$taxon=="Unknown"]))

# add to eva
eva$species[eva$`Matched concept` %in% eva_names$species] <-
  eva_names$euro[match(eva$`Matched concept`[eva$`Matched concept` %in% eva_names$species], eva_names$species)]
```

`

```{r}
# remove all lichens etc
unique(eva$`Taxon group`)
eva <- eva[!eva$`Taxon group` %in% c("Lichen", "Moss", "Alga","Stonewort","Mushroom"),]

# which species still not linked
eva_names <- as.data.frame(unique(eva$`Matched concept`[is.na(eva$species)]))
colnames(eva_names)<- "species"
eva_names$taxon <- eva$`Taxon group`[match( eva_names$species, eva$`Matched concept`)]
```


## 5.3 TPL
```{r}
# check taxon group
plant_list <- TPL(vegdata::taxname.abbr(eva_names$species))

# sort (makes easier removing)
eva_names$species <- sort(eva_names$species)
remove <- eva_names$species[!plant_list$GROUP[order(plant_list$YOUR_SEARCH)] %in% c("Angiosperms", "Gymnosperms")]

eva_names <- as.data.frame(eva_names$species[!eva_names$species %in% remove])
colnames(eva_names)<- "species"

# also remove these from the eva database
eva <- eva[!eva$`Matched concept` %in% remove,]
```



```{r}
status <- status(vegdata::taxname.abbr(eva_names$species))
status <- status[!duplicated(status),]

#write.csv(status, "TPL_status_euro.csv")

status_names <- unique(status$SCIENTIFIC_NAME)

# change names to accepted
accept_dup<- data.frame(species= character(), family= character(), old_name= character())
# run for all species
for(i in 1: length(eva_names$species)){ 
  # first check if they exist in our status data, if not keep only old name
  if(any(is.na(status$STATUS[status$SCIENTIFIC_NAME %in% status_names[i]]))){
    accept_dup <- add_row(accept_dup,
                      species= NA, 
                      family= NA,
                      old_name= eva_names$species[i])
  # then check if they exist in our status data with an accepted name
  } else if(any(status$STATUS[status$SCIENTIFIC_NAME %in% taxname.abbr(status_names[i])]=="Accepted")){
    accept_dup <- add_row(accept_dup,
                      species= status$ACCEPTED_SPECIES[status$STATUS=="Accepted" & status$SCIENTIFIC_NAME==status_names[i]], 
                      family= status$FAMILY[status$STATUS=="Accepted" & status$SCIENTIFIC_NAME==status_names[i]],
                      old_name= eva_names$species[i])
  # check for species that are present only once and are designated as 'synonym'
  } else if(all(!any(status$STATUS[status$SCIENTIFIC_NAME==status_names[i]]=="Accepted"),
            length(status$STATUS[status$SCIENTIFIC_NAME==status_names[i]])==1 &
            status$STATUS[status$SCIENTIFIC_NAME==status_names[i]]=="Synonym")){
    accept_dup <- add_row(accept_dup,
                      species= status$ACCEPTED_SPECIES[status$STATUS=="Synonym" & status$SCIENTIFIC_NAME==status_names[i]], 
                      family= status$FAMILY[status$STATUS=="Synonym" & status$SCIENTIFIC_NAME==status_names[i]],
                      old_name= eva_names$species[i]) 
  # if no other options, give NA  
  }  else {
   accept_dup <- add_row(accept_dup,
                      species= NA, 
                      family= NA,
                      old_name= eva_names$species[i])
  }
}
```



```{r}
# give new names from TPL
eva$species[eva$`Matched concept` %in% accept_dup$old_name]<-
  accept_dup$species[match(eva$`Matched concept`[eva$`Matched concept` %in% accept_dup$old_name], accept_dup$old_name)] 

eva <- eva[eva$`Taxon group` %in% c("Vascular plant", "Unknown"),]

#which still missing
eva_names <- as.data.frame(unique(eva$`Matched concept`[is.na(eva$species)]))
colnames(eva_names)<- "species"
eva_names$taxon <- eva$`Taxon group`[match( eva_names$species, eva$`Matched concept`)]

sum(eva$`Taxon group`[!is.na(eva$species)]=="Unknown")
```


## 5.4 Check
```{r}
# all plants (originally, now also on unknown species)
eva_plants <-  eva_names

# load kew data link
library(kewr)

# search powo and store results in list if we find a direct link
mybiglist <- list()
for(i in 1: length(eva_plants$species)){
  # powo
  r<- search_powo(eva_plants$species[i])
  # only if there is a match store result in data, otherwise put NA
  if(length(r$results)>0){
    # make dataframe
    tmp <- tidy(r)
    # add to list
    mybiglist[[i]] <- tmp
  } else {
    mybiglist[[i]] <- NA
  }
}

i=8
# only retrieve accepted names
accepted <- data.frame(new_name = character(), old_name = character(), number= numeric(), taxon= character())
for(i in 1: length(eva_plants$species)){
  if(!all(is.na(mybiglist[[i]]))){
    if(any(mybiglist[[i]]$accepted==T)){
      accepted <- add_row(accepted, new_name= mybiglist[[i]]$name[mybiglist[[i]]$accepted==T], old_name =eva_plants$species[i],  
                          number= i, taxon= eva_plants$taxon[i])
    } else {
      accepted <- add_row(accepted, new_name= NA, old_name =eva_plants$species[i] ,  number= i, taxon=eva_plants$taxon[i])
    }
  } else {
      accepted <- add_row(accepted, new_name= NA,old_name = eva_plants$species[i],  number= i, taxon= eva_plants$taxon[i] )
  } 
}

# some species have multiple accepted names, we cannot be sure which species it is so remove
accepted_dup <- accepted[accepted$old_name %in% accepted$old_name[duplicated(accepted$old_name)],]
length(unique(accepted_dup$number))

# some species were not found
accepted_na <- accepted[is.na(accepted$new_name),]

# only retain unique new names
accepted<- accepted[!(accepted$old_name %in% accepted$old_name[duplicated(accepted$old_name)] | is.na(accepted$new_name)),]

# add to eva
eva$species[eva$`Matched concept` %in% accepted$old_name] <- 
  accepted$new_name[match(eva$`Matched concept`[eva$`Matched concept` %in% accepted$old_name], accepted$old_name)]

# check how many observations in eva are not assigned
unknown <- eva[is.na(eva$species),]
# make unique --> 1143 = 1136 + 7
unknown_sp <- unknown[!duplicated(unknown$`Matched concept`),]


unknown_plants <- unknown_sp[unknown_sp$`Taxon group`=="Vascular plant",]
unknown_plants <- unknown_plants[!unknown_plants$`Matched concept` %in% accepted_dup$old_name,]
```


We will only check vascular plants
```{r}
own_class <- data.frame(species= c("Centaurea bimorpha",
                                   "Aconitum anthora subsp. anthora",
                                   "Ostericum maximowiczii",
                                   "Saccharum ravennae",
                                   "Monanthes brachycaulos",
                                   "Sinapidendron gymnocalyx",
                                   "Betula pendula subsp. pendula",
                                   "Bunium bulbocastanum",
                                   "Asphodelus cerasiferus",
                                   "Rosmarinus eriocalyx",
                                   "Muehlenbeckia platyclada"),
                        old_name=c("Centaurea dimorpha", 
                                  "Aconitum anthoroidetum",
                                  "Angelica maximoviczii",
                                  "Saccharum ravenae",
                                  "Monanthes brachycaulon",
                                  "Sinapidendron gymnocalix",
                                  "Betula xpubescens subsp. xverrucosa",
                                  "Bunium bulbostanum",
                                  "Asphodelus cerasifera",
                                  "Rosmarinus eriocalix",
                                  "Muehlenbeckia platyclados"))

eva$species[eva$`Matched concept` %in% own_class$old_name]<-
  own_class$species[match(eva$`Matched concept`[eva$`Matched concept` %in% own_class$old_name], own_class$old_name)]

unknown_plants <- unknown_plants[!unknown_plants$`Matched concept` %in% own_class$old_name,]

eva$species[eva$`Matched concept` %in% unknown_plants$`Matched concept`]<-
  eva$`Matched concept`[eva$`Matched concept` %in% unknown_plants$`Matched concept`]

eva$species[eva$`Matched concept` %in% accepted_dup$old_name[accepted_dup$taxon=="Vascular plant"]]<-
  eva$`Matched concept`[eva$`Matched concept` %in% accepted_dup$old_name[accepted_dup$taxon=="Vascular plant"]]

own_class <- data.frame(species= c("Muehlenbeckia platyclada",
                                   "Betula petraea",
                                   "Campanula lingulata"),
                        old_name=c("Muehlenbeckia platyclados",
                                   "Betula \"petraea\"",
                                   "Campanula ligulata"))

eva$species[eva$`Matched concept` %in% own_class$old_name]<-
  own_class$species[match(eva$`Matched concept`[eva$`Matched concept` %in% own_class$old_name], own_class$old_name)]

# check how many observations in eva are not assigned
unknown <- eva[is.na(eva$species),]
# make unique
unknown_sp <- unknown[!duplicated(unknown$`Matched concept`),]

eva <- eva[!eva$`Matched concept` %in% unknown_sp$`Matched concept`,]
length(unique(eva$species[eva$`Taxon group`[!is.na(eva$species)]=="Unknown"]))
```


```{r}
unknown <- eva[eva$`Taxon group`=="Unknown",]
eva_names <- as.data.frame(unique(eva$species[eva$`Taxon group`=="Unknown"]))
colnames(eva_names)<- "species"

# add taxon names and euro matches
eva_names$taxon <- eva$`Taxon group`[match( eva_names$species, eva$`Matched concept`)]


plant_list <-TPL(eva_names$species)

# sort (makes easier removing)
eva_names$species <- sort(eva_names$species)
remove <- eva_names$species[!is.na(plant_list$GROUP[order(plant_list$YOUR_SEARCH)]) & 
                              !plant_list$GROUP[order(plant_list$YOUR_SEARCH)] %in% c("Angiosperms", "Gymnosperms")]

eva_names <- as.data.frame(eva_names$species[!eva_names$species %in% remove])
colnames(eva_names)<- "species"

# also remove these from the eva database
eva <- eva[!eva$`Matched concept` %in% remove,]

unknown <- eva[eva$`Taxon group`=="Unknown",]
#write_csv(eva, "eva_euro_med.csv")
```


# 6 FILE 4
## 6.1 Analysis
This has become wrong, as we have some intra-european neophytes that are present in one part and not in another 
Compute which species occurs how often in eva:
```{r}

# For speed optimization we extract only the necessary information from EVA
reducedEva <- eva |> select(PlotObservationID, species, old.name, `Cover %`)

uniqueSpecies <- unique(reducedEva$species)

# Empty dataset
plotsPerSpecies <- data.frame(names = character(), numberOfPlots = integer())

# Prepare parallel
parallel::detectCores()
n.cores <- parallel::detectCores() - 2
my.cluster <- parallel::makeCluster(
  n.cores, 
  type = "PSOCK"
)
print(my.cluster)
doParallel::registerDoParallel(cl = my.cluster)
foreach::getDoParRegistered()
foreach::getDoParWorkers()


# Count in how many plots a species occurs
begin<- Sys.time()
plotsPerSpecies<-foreach(i= 1:length(uniqueSpecies), .combine='rbind', .packages=c("dplyr","mgcv")) %dopar% {
  tmp <- reducedEva$PlotObservationID[reducedEva$species == uniqueSpecies[i]]
  names=uniqueSpecies[i]
  numberOfPlots= length(unique(tmp))
  plotsPerSpecies[i,]<- c(names, numberOfPlots)
}
plotsPerSpecies<- as.data.frame(plotsPerSpecies)
plotsPerSpecies<-setNames(plotsPerSpecies, c("names", "numberOfPlots"))
row.names(plotsPerSpecies)<- c(1: length(uniqueSpecies))
end<- Sys.time()
round(end-begin)

parallel::stopCluster(cl = my.cluster)
#write_csv(plotsPerSpecies, "plotsPerSpeciesInEurope.csv")
```


## 6.2 Downsample
Determine which plant species have more than 10 plots with cover > 50% and cover < 1% for analysis:
```{r, message=F}
# Load data previously made with the number of plots per species across the dataset

# plotsPerSpecies <- read_csv("plotsPerSpeciesInEurope.csv")

parallel::detectCores()
n.cores <- parallel::detectCores() - 2
my.cluster <- parallel::makeCluster(
  n.cores, 
  type = "PSOCK"
)
print(my.cluster)
doParallel::registerDoParallel(cl = my.cluster)
foreach::getDoParRegistered()
foreach::getDoParWorkers()

# Get the names of the species which have at least 20 plots (since we want 10 smaller and 10 bigger)
candidatesForCoverClassAnalysis <- plotsPerSpecies$names[plotsPerSpecies$numberOfPlots >= 20]

# create empty vector
taxaForCoverClassAnalysis <- c()
# run for all candidates
begin<- Sys.time()
taxaForCoverAnalysis<- foreach(i= 1: length(candidatesForCoverClassAnalysis), .combine='rbind', .packages=c("dplyr","mgcv")) %dopar%  {
  # if the sum of all species with a cover of 50% for a species and less than 1% each are more than 10 
  # && makes sure to only look at the second if it is already T
  taxon<- candidatesForCoverClassAnalysis[i]
  if(sum(eva$species == taxon & eva$`Cover %` >= 50) > 10 && sum(eva$species == taxon & eva$`Cover %` < 1) > 10) {
    # write the species name to the list
    taxaForCoverClassAnalysis[i] <- c(taxon)
  }
}
end<- Sys.time()
round(end-begin)

parallel::stopCluster(cl = my.cluster)
##write_csv(data.frame(taxa = taxaForCoverClassAnalysis), "taxaForCoverClassAnalysis.csv")
```

## 6.3 New method
```{r}
plotsPerSpecies <- read_csv("speciesDominance_new_categorization_10.csv", show_col_types = FALSE)

taxa_new<- plotsPerSpecies[plotsPerSpecies$numberOfPlots >=50,]

#write_csv(data.frame(taxa = taxa_new), "taxa_new.csv")
```
