==> 3_Region_Check.Rmd <==
---
title: "Plots per region"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---

```{r, echo=FALSE}
rm(list=ls())
```


# 1 Prepare
## 1.1 Load
```{r, echo=F, warning=F, message=F}
library(readr)
library(tidyverse)
library(sf)
library(sp)
#library(CoordinateCleaner)
library(doParallel)
library(units)
#library(lwgeom)
```


## 1.2 Data
Read data from header
```{r, warning= F, message= F}
header <- read_delim("../EVA Data/171_NeophyteInvasions20230216_notJUICE_header.csv", "\t")
fast <- T
if(fast) {
  header <- header[runif(length(header$PlotObservationID)) > 0.999,]
}
```

# 2 Country check
## 2.1 No coordinates
Check how many plots have a country defined but no coordinates and later remove those plots
```{r, warning=F, message=F}
# Count plots without coordinates but assigned country
numberOfPlotWithCountryAndNoCoordinates <- sum(!is.na(header$Country) & (is.na(header$Latitude) | is.na(header$Longitude)))
# give relative number of plots with country but no coordinates
percentOfPlotsWithCountryButNoCoordinates <- numberOfPlotWithCountryAndNoCoordinates / length(header$Country) * 100
percentOfPlotsWithCountryButNoCoordinates

#remove data without coordinates
header <- header[!(is.na(header$Latitude) | is.na(header$Longitude)),]
```

## 2.2 Country join
Make header a simple feature object and give it the WGS84 CRS
```{r, warning=F, message=F}
# make geometry with long and lat
plotLocations <- st_as_sf(header, coords = c("Longitude","Latitude"), remove = FALSE)
st_crs(plotLocations) <- CRS("+proj=longlat +datum=WGS84")
```


Load the Euro+Med European region shapefiles and give the same CRS
```{r, warning=F, message=F}
medRegions <- read_sf("../Extra data/EURO+MED/Regions", "Emed_regions")
medRegions <- st_transform(medRegions, CRS("+proj=longlat +datum=WGS84"))
```


Return all records that match spatially except the 20th row, which are the Netherlands, these are subsequently run seperately because there seems to be some problems in the shapefile
```{r}
# Assign all plots which are within one of the regions to the given region
sf_use_s2(T) #S2 can provide better performance for certain types of spatial operations, especially when dealing with large spatial datasets.
joinedData <- st_join(plotLocations, medRegions[-20,], join = st_within)

# There is something awry with the geometry of the Netherlands, this is why we treat it specially here
sf_use_s2(F)
res <- st_join(plotLocations, medRegions[20,], join = st_within)
# Merge the Netherlands back into the overall data
# use the in data to give the index for joinedData (because we use res for the region and replace totally with the subset of data being from the Netherlands)
joinedData[(res$Region == "Netherlands") %in% TRUE,] <- res[(res$Region == "Netherlands") %in% TRUE,]
```


# 3 Unassigned
Not all data was assigned to a region, check percentage of plots not assigned. We then look to assign the rest to a region when they are within 10 km of the closest region. 

## 3.1 General
First we compute the distances to a bounding box, which facilitates the calculations. If the plot is within the threshold distance to the box, we compute the real distance for every region untill we obtain the smallest distance. This data is stored in region, together with the distance.
```{r}
# Percentage not assigned
percentNotAssignedPlots <- sum(is.na(joinedData$Region))/length(joinedData$PlotObservationID)*100
percentNotAssignedPlots
```

## 3.2 Sea 
```{r}
# Create plot with all na data
remainingPlots <-header[is.na(joinedData$Region),]

# Check which plots are not located in the sea
not_sea<- cc_sea(remainingPlots, lon= "Longitude", lat= "Latitude", buffer= 25000)

# Check which plots are removed and count them
sea <- remainingPlots[!remainingPlots$PlotObservationID %in% not_sea$PlotObservationID,]
sum(sea$PlotObservationID %in% remainingPlots$PlotObservationID)

plot<- F
if(plot){
Europe<- st_combine(medRegions)
Europe_GG<- fortify(Europe)
sea<- fortify(sea)
sea <-  st_as_sf(sea, coords = c("Longitude","Latitude"), remove = FALSE)
st_crs(sea) <- CRS("+proj=longlat +datum=WGS84")

ggplot()+
  geom_sf(data=Europe_GG)+
  geom_sf(data= sea, color="red", size=4)
}
```

## 3.3 Region assignment
```{r}
# Prepare parallel
parallel::detectCores()
n.cores <- parallel::detectCores() - 2
my.cluster <- parallel::makeCluster(n.cores, type = "PSOCK")
print(my.cluster)
doParallel::registerDoParallel(cl = my.cluster)
foreach::getDoParRegistered()
foreach::getDoParWorkers()

# Create threshold to use
distanceThreshold <- set_units(10000,m)

# Dataset of points that were not assigned with sf_within
remainingPlots <-plotLocations[is.na(joinedData$Region),]

# make smaller to make calculation quicker
remainingPlots<- remainingPlots[, c("PlotObservationID","Country", "geometry")]

# Compute bounding boxes for the regions, this allows to decide faster, if a plot is outside the distanceThreshold from a region
boundingBoxes <- medRegions

# Create box for every country
for(i in 1:length(boundingBoxes$country)) {
  boundingBoxes$geometry[i] <- st_as_sfc(st_bbox(boundingBoxes$geometry[i]))
}

# Begin parameters
remainingPlots$Region <- NA
remainingPlots$Distance <- -1

# Start clock
begin<-Sys.time() 

# Perform parallel loop and add results to remaining plots
remainingPlots[,4:5]<-foreach(i = 1:length(remainingPlots$PlotObservationID), 
                              .combine='rbind', .packages=c("dplyr","mgcv", "sf","units")) %dopar% {
  
  # set region to no region and distance to large value
  region <- NA
  distance <- set_units(1000000,m)
  
  # check for every country individually if distance to the box is more than is allowed
  for(j in 1:length(boundingBoxes$country)) {
    sf_use_s2(F)
    distBoundingBox <- st_distance(boundingBoxes[j,], remainingPlots[i,]) #Calculate the distance
    # Only if the distance to the bounding box is smaller than distanceThreshold we need to compute the real distance to the region
    if(distBoundingBox < distanceThreshold){
      dist <- st_distance(medRegions[j,], remainingPlots[i,])
      # if this distance is smaller than the previously computed distance and the threshold, save this one
      if(dist < distanceThreshold & dist < distance) {
        region <- medRegions$Region[j]
        distance <- dist
      }
    }
  }
  # add these to the remaining plots
  remainingPlots$Region[i] <- c(region, distance)
}
# End time and calculation time needed
end<-Sys.time()
round(end-begin, 2)  

# Stop cluster
parallel::stopCluster(cl = my.cluster)

# Change type of distance in plot
remainingPlots$Distance<- as.double(remainingPlots$Distance)
```


We assign the remaining plots to the region obtained (if applicable)
```{r}
# reduce dataset with country and region data
allPlotsWithRegion <- joinedData[,c("PlotObservationID", "Region", "Country")]
# calculate number of plots without a region
sum(is.na(allPlotsWithRegion$Region))

# merge the plots which lie within a region with those which are in proximity
allPlotsWithRegion[is.na(allPlotsWithRegion$Region),] <- remainingPlots[,c("PlotObservationID", "Region", "Country")]
# Look at how many plots have no region after the buffer method
sum(is.na(allPlotsWithRegion$Region))


# align with sea
sea_removed <- allPlotsWithRegion[allPlotsWithRegion$PlotObservationID %in% sea$PlotObservationID & is.na(allPlotsWithRegion$Region),]
# 5254 of 12806 removed plots was located in the sea
not_sea_removed <- allPlotsWithRegion[!allPlotsWithRegion$PlotObservationID %in% sea$PlotObservationID & is.na(allPlotsWithRegion$Region),]
# 7552 sites not in sea
# check up gives that these are mostly from Algeria (1346), Marocco (1110), Tunisia (610), Turkey (450), Cyprus (508) or Israel (279)
not_sea_removed |> group_by(Country)|> summarise(n=n())
```

## 3.4 Save
We make the dataset a normal dataframe (remove the geometry) and save the dataset
```{r}
# remove geometry
st_geometry(allPlotsWithRegion) <- NULL

# save file
#write_csv(allPlotsWithRegion, "plot_to_region_assignment_new.csv")
#write_csv(sea, "sea_3.csv")
```


# 4 Extra
## 4.1 Still unassigned
Some countries (from eva) still posses plots that are not assigned to an Euro+Med region. We first check how many and then check for which countries this applies.
```{r}
# percent still unassigned
percentOfPlotsStillUnassigned <- sum(is.na(allPlotsWithRegion$Region))/length(allPlotsWithRegion$Region)*100

# from which countries do we have plots which are still unassigned?
unique(allPlotsWithRegion$Country[is.na(allPlotsWithRegion$Region)])

# quite some data from North-Africa (Tunisia, Algeria), Islands or coastal countries
```

## 4.2 Check sea
```{r}
# get all plots which are not assigned
remainingPlots <- remainingPlots[is.na(remainingPlots$Region),]

# how many of these plots were located in the sea
percentSea<- sum(remainingPlots$PlotObservationID %in% sea$PlotObservationID)/ length(sea$PlotObservationID)
# 56.1% in sea are in remaining --> 44% accepted but originally in sea
percentRegionSea<- sum(remainingPlots$PlotObservationID %in% sea$PlotObservationID)/ length(remainingPlots$PlotObservationID)
# 41.0% in remaining also in sea
```



## 4.3 check assigned
Check whether the number of plots which are in a region match number of plots in a country
```{r}
compareCountryLabelWithRegion <- function(country, region) {
  message("Number of plots in region: ", sum(allPlotsWithRegion$Region == region, na.rm = T))
  message("Number of plots in country: ", sum(allPlotsWithRegion$Country == country, na.rm = T))
}

# Compare the number of plots in a region with the number of plots which are in the "same" country, does it make sense?
compareCountryLabelWithRegion("Albania", "Albania")
compareCountryLabelWithRegion("Austria", "Austria")
compareCountryLabelWithRegion("Netherlands", "Netherlands")
compareCountryLabelWithRegion("Italy", "Italy")
compareCountryLabelWithRegion("Italy", "Sicily")
compareCountryLabelWithRegion("Italy", "Sardinia")

```


# 5 Plot
We now plot for a region
```{r}
plot= F
if(plot){
# Which regions are there?
sort(medRegions$Region)

#allPlotsWithRegion<- read.csv("plot_to_region_assignment_small.csv")

# Code for plotting the plots and regions
regionToLookAt <- ("France") # With this line you can select which region to plot
regionShape<- medRegions[medRegions$Region==regionToLookAt,]
boundingBox <- boundingBoxes$geometry[match(regionToLookAt, medRegions$Region)]
plotsInRegion <- plotLocations[allPlotsWithRegion$Region == regionToLookAt & runif(length(header$Country)) > 0,] # Here you can reduce the number of plots shown

medRegions.df <- fortify(medRegions)
regionShape <- fortify(regionShape)
boundingBox <- fortify(boundingBox)

ggplot() +   
  #geom_sf(data = boundingBox, color = "red") + 
  geom_sf(data = regionShape, color = "black", fill = "cyan1") + 
  geom_sf(data = plotsInRegion, color = "red", size = 1) + 
  coord_sf()
}
```

Plot Europe with plots
```{r}
if(plot){
Europe<- st_combine(medRegions)

Europe_Plots<- fortify(allPlotsWithRegion[!is.na(allPlotsWithRegion$Region),])

Europe_GG<- fortify(Europe)
ggplot()+
  geom_sf(data=Europe_GG)+
  geom_sf(data= Europe_Plots, color="red", size=0.5)

allPlotsWithRegion[(is.na(allPlotsWithRegion$Region)),]
}
```
