==> 4_Data_Preparation.Rmd <==
---
title: "Data preparation"
subtitle: "based on the work of Christian Ledergerber"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---
# 1 LOAD
## 1.1 prepare
We first clean the environment
```{r, echo=T}
#rm(list=ls())
eval <- F
rm(list=ls())
```

We load the required libraries
```{r, warning=F, message=F}
library(readr)
library(tidyverse)
```


## 1.2 Eva Data

## 1.3 New data
```{r}
eva <- read_csv("../EVA data/eva_ESy.csv", show_col_types = FALSE)
```



Here we decide on which layers to keep
```{r}
# sum all species over layers if present in multiple layers
numberOfRowsTreeLayer<- length(eva$Layer[eva$Layer %in% c(1:3)])

head(eva)
Layerremove=F
if(Layerremove){
  numberOfRowsTreeLayer<- length(eva$Layer[eva$Layer %in% c(1:3)])
  eva <- eva |> group_by(PlotObservationID, species) |> filter(Layer %in% c(0,4:9))
}

test <- eva[duplicated(eva[, c("PlotObservationID","name","Layer")]),]
length(unique(test$PlotObservationID))

# Which rows contain duplicates
numberOfRowsDuplicates<- sum(test$duplicated>0)
Layersum=T
if(Layersum){
  eva <- eva |> group_by(PlotObservationID, species,`Turboveg2 concept`, `Matched concept`, irena, name, source, Layer) |>
    summarise(`Cover %`= sum(`Cover %`))
} else {
  eva <- eva |> select(PlotObservationID, species,`Turboveg2 concept`, `Matched concept`, irena, name, source,`Cover %`, Layer)
}

# remove species with 0 cover
numberOfZeroCover <- length(eva$PlotObservationID[eva$`Cover %`=="0"])
# look at from which datasets these come --> "Belgium-forest" "Belgium-INBOVEG" "GrassVeg.DE" "Turkey NFVDT"--> wrong entering --> remove
# unique(header$Dataset[header$PlotObservationID %in% (eva$PlotObservationID[eva$`Cover %`=="0"])])
coverzero=T
if(coverzero){
  eva<- eva[!eva$`Cover %`=="0", ]
}
```


Compute total cover for each plot.
```{r}
# fisher calculation
reducedEva <- eva[, c(1,6,9,8)]
reducedEva <- reducedEva |> group_by(PlotObservationID, name, Layer) |> summarise(cover= sum(`Cover %`))
reducedEva <- reducedEva|> group_by(PlotObservationID, name) |> 
  summarise(cover = (1-prod((1-ifelse(cover/100>1, 1, cover/100 ))))*100)

totalCover <- 
  reducedEva %>% 
  group_by(PlotObservationID) %>% 
  summarise(totalCover = sum(cover))

# Check distribution (we max on 2000, but total cover goes up to 25000)
hist(totalCover$totalCover[totalCover$totalCover<2000 ])
hist(totalCover$totalCover[totalCover$totalCover> 900 & totalCover$totalCover<2000 ])
# We calculate the number of plots in which this occurs
numberOfTooLargeCover<- length(totalCover$totalCover[totalCover$totalCover> 300])

# We check the datasets --> many datasets have more than 900 (64 out of 320)
# unique(header$Dataset[header$PlotObservationID %in% (totalCover$PlotObservationID[totalCover$totalCover> 900])])

# We check the plots (only largest subset)
eva[eva$PlotObservationID %in% totalCover$PlotObservationID[totalCover$totalCover> 9000],]

maxcover=T
if(maxcover){
  eva<- eva[!eva$PlotObservationID %in% totalCover$PlotObservationID[totalCover$totalCover> 500],] 
}
```


## 1.4 Header Data
Read in the header information. Look at and remove plots for which:

-   Area is NA or smaller than 0
-   Area is larger than 1000 m²
-   The longitude or latitude data is not known
-   The plot is not assigned to a region
-   The plot only contains presence/absence data

```{r, message=F, warning=F}
header <- read_delim("../EVA Data/171_NeophyteInvasions20230216_notJUICE_header.csv", "\t")
totalNumberOfPlots <- length(header$PlotObservationID)

# Plot sizes too small/ inexistent
plotsWithoutPlotSize <- is.na(header$`Relevé area (m²)`) | header$`Relevé area (m²)` <= 1
numberOfPlotsWithoutPlotSize <- sum(plotsWithoutPlotSize)

# Plot sizes too large (https://onlinelibrary.wiley.com/doi/full/10.1002/ece3.8965)
tooLargePlots <- header$`Relevé area (m²)` > 1024
numberOfTooLargePlots<- sum(tooLargePlots, na.rm=T)

# Plots without coordinates
plotsWithoutLocation <- is.na(header$Latitude) | is.na(header$Longitude)
numberOfPlotsWithoutLocation <- sum(plotsWithoutLocation)

# Check whether the amount of observation plots correspond
# This is the data made in 3_Region_check (here we also removed all data without coordinates)
plotsRegionAssignments <- read_csv("../Extra data/Intermediate/plot_to_region_assignment.csv")
# Error message if the two are not the same
if( sum(header$PlotObservationID[!plotsWithoutLocation] != plotsRegionAssignments$PlotObservationID) > 0) {
  message("Error: plotsRegionAssignments expected to contain the same plots as contained in header data.")
}

# Count number of plots unassigned
# create logical vector of header length
plotsWithoutRegionAssignment <- vector(mode = "logical", length = length(header$PlotObservationID))
# make true when the plot has no region, taking into consideration that some plots have no coordinates
plotsWithoutRegionAssignment[!plotsWithoutLocation] <- is.na(plotsRegionAssignments$Region)
numberOfPlotsWithoutRegionAssignment = sum(plotsWithoutRegionAssignment)

# Remove only Presence/ Absence data and NA data
namesForPresenceAbsence <- c("Presence/Absence", "Presentie/Absentie", NA)
# logical vector with length header 
plotsWithPresenceAbsenceOnly <- header$`Cover abundance scale` %in% namesForPresenceAbsence
numberOfPlotsWithoutCoverInformation <- sum(plotsWithPresenceAbsenceOnly)

# sum all plots to be removed
allPlotsToBeRemoved <- plotsWithoutPlotSize | tooLargePlots | plotsWithoutLocation | plotsWithPresenceAbsenceOnly | plotsWithoutRegionAssignment
totalNumberRemoved <- sum(allPlotsToBeRemoved)
```


`r numberOfPlotsWithoutPlotSize` of the plots do not contain a size, `r numberOfTooLargePlots` were larger than 1000m² and therefore most likely no plots, `r numberOfPlotsWithoutLocation` of the plots do not have a location, `r numberOfPlotsWithoutRegionAssignment` cannot be assigned to one of the neophyte regions and `r numberOfPlotsWithoutCoverInformation` of the plots do not contain cover data. All these plots (`r totalNumberRemoved`) are removed:
```{r}
header <- header[!allPlotsToBeRemoved,]
```

We also add the region to the data for analysis and change the region and dataset to a factor.
```{r}
header <- left_join(header, plotsRegionAssignments, by = "PlotObservationID")
header$Region <- as.factor(header$Region)
header$Dataset <- as.factor(header$Dataset)
```

Finally, we will keep only the plots that are still available in both Eva and Header
```{r}
plotsInEvaAndHeader <- intersect(eva$PlotObservationID, header$PlotObservationID)
eva <- eva[eva$PlotObservationID %in% plotsInEvaAndHeader,]
header <- header[header$PlotObservationID %in% plotsInEvaAndHeader,]
length(header$PlotObservationID)
```
After filtering `r length(header$PlotObservationID)` plots remain.


# 2 INDICATORS
```{r}
# Create the dataframe
species_data <- data.frame(old = c( "crop vineyard", "crop barley", "crop maize", "crop wheat", "crop potato","crop hops", "crop pea", "crop digitalis", 
                                    "crop tomato", 
                                    "crop currant","crop cabbage", "crop bean", "crop onion", "crop asparagus", "crop spelt","crop carrot", "crop rhubarb", 
                                    "crop clover", 
                                    "crop tobacco", "crop zucchini", "crop parsley", "crop gladiolus", "crop salsify", "crop radish", "crop lettuce",
                                    "crop buckwheat"),
                          new = c("Vitis vinifera", "Hordeum vulgare", "Zea mays", "Triticum aestivum", "Solanum tuberosum", "Humulus lupulus", "Pisum sativum", 
                                  "Digitalis purpurea", "Solanum lycopersicum", "Ribes rubrum","Brassica oleracea", "Phaseolus vulgaris", "Allium cepa", 
                                  "Asparagus officinalis", "Triticum aestivum","Daucus carota", "Rheum rhabarbarum", "Trifolium repens", "Nicotiana tabacum", 
                                  "Cucurbita pepo", "Petroselinum crispum", "Gladiolus grandiflorus", "Tragopogon porrifolius", "Raphanus sativus", 
                                  "Lactuca sativa","Fagopyrum esculentum"))


eva$name[eva$name %in% species_data$old] <- species_data$new[match(eva$name[eva$name %in% species_data$old], species_data$old)]

length(unique(eva$PlotObservationID[eva$name %in% species_data$new]))
```


## 2.1 EIVE
```{r}
eive <- read_delim("../Extra data/ENV/EIVE_Paper_1.0_SM_08.csv", ",")

eva_names <- eva |> group_by(name, irena, `Matched concept`, species, `Turboveg2 concept`) |> summarise(n=n())

# name
eive$name <- eva_names$name[match(eive$TaxonConcept, eva_names$name)]
sum(!is.na(eive$name))
# remove aggr.
eive$name[is.na(eive$name)] <- eva_names$name[match(eive$TaxonConcept[is.na(eive$name)], gsub(" aggr\\.", "", eva_names$name))]
sum(!is.na(eive$name))
# vegdata
eive$name[is.na(eive$name)] <- eva_names$name[match(eive$TaxonConcept[is.na(eive$name)], vegdata::taxname.abbr((eva_names$name)))]
sum(!is.na(eive$name))
eive$name[is.na(eive$name)] <- eva_names$name[match(eive$TaxonConcept[is.na(eive$name)], vegdata::taxname.simplify((eva_names$name)))]
sum(!is.na(eive$name))

# check species
eive$name[is.na(eive$name)] <- eva_names$name[match(eive$TaxonConcept[is.na(eive$name)], eva_names$species)]
sum(!is.na(eive$name))
eive$name[is.na(eive$name)] <- eva_names$name[match(eive$TaxonConcept[is.na(eive$name)], eva_names$irena)]
sum(!is.na(eive$name))
eive$name[is.na(eive$name)] <- eva_names$name[match(eive$TaxonConcept[is.na(eive$name)], eva_names$`Matched concept`)]
sum(!is.na(eive$name))
eive$name[is.na(eive$name)] <- eva_names$name[match(eive$TaxonConcept[is.na(eive$name)], eva_names$`Turboveg2 concept`)]
sum(!is.na(eive$name))

eive <- eive |> group_by(name) |> summarise(EIVEresL= mean(`EIVEres-L`), EIVEresM = mean(`EIVEres-M`), EIVEresN = mean(`EIVEres-N`),
                                            EIVEresT= mean(`EIVEres-T`), EIVEresR = mean(`EIVEres-R`), 
                                            eive_name =  names(sort(table((TaxonConcept)), decreasing=T)[1]))

```


```{r}
# assign to eva_names
eva_names <- left_join(eva_names, eive, by= c("name"="name"))

# assign to eva
eva$eive_name <- eva_names$eive_name[match(eva$name, eva_names$name)]
eva$EIVEresL <- eva_names$EIVEresL[match(eva$name, eva_names$name)]
eva$EIVEresT <- eva_names$EIVEresT[match(eva$name, eva_names$name)]
eva$EIVEresM <- eva_names$EIVEresM[match(eva$name, eva_names$name)]
eva$EIVEresN <- eva_names$EIVEresN[match(eva$name, eva_names$name)]
eva$EIVEresR <- eva_names$EIVEresR[match(eva$name, eva_names$name)]

sum(!is.na(eva$eive_name))/length(eva$PlotObservationID)
sum(!is.na(eva_names$eive_name))/length(eva_names$name)
```

## 2.2 DIV
```{r}
div <- read_delim("../Extra data/DIV/disturbance_indicator_values.csv", ",")

eva_names <- eva |> group_by(name, irena, `Matched concept`, species, `Turboveg2 concept`) |> summarise(n=n())

# name
div$name <- eva_names$name[match(div$species, eva_names$name)]
sum(!is.na(div$name))
# remove aggr.
div$name[is.na(div$name)] <- eva_names$name[match(div$species[is.na(div$name)], gsub(" aggr\\.", "", eva_names$name))]
sum(!is.na(div$name))
# vegdata
div$name[is.na(div$name)] <- eva_names$name[match(div$species[is.na(div$name)], vegdata::taxname.abbr((eva_names$name)))]
sum(!is.na(div$name))
div$name[is.na(div$name)] <- eva_names$name[match(div$species[is.na(div$name)], vegdata::taxname.simplify((eva_names$name)))]
sum(!is.na(div$name))

# check species
div$name[is.na(div$name)] <- eva_names$name[match(div$species[is.na(div$name)], eva_names$species)]
sum(!is.na(div$name))
div$name[is.na(div$name)] <- eva_names$name[match(div$species[is.na(div$name)], eva_names$irena)]
sum(!is.na(div$name))
div$name[is.na(div$name)] <- eva_names$name[match(div$species[is.na(div$name)], eva_names$`Matched concept`)]
sum(!is.na(div$name))
div$name[is.na(div$name)] <- eva_names$name[match(div$species[is.na(div$name)], eva_names$`Turboveg2 concept`)]
sum(!is.na(div$name))

div <- div |> group_by(name) |> summarise(Disturbance.Severity = mean(`Disturbance.Severity`), 
                                          Disturbance.Severity.herblayer = mean(`Disturbance.Severity.herblayer`), 
                                          Disturbance.Frequency = mean(`Disturbance.Frequency`),
                                          Disturbance.Frequency.herblayer= mean(`Disturbance.Frequency.herblayer`), 
                                          Grazing.Pressure = mean(`Grazing.Pressure`), 
                                          Mowing.Frequency= mean(`Mowing.Frequency`), 
                                          Soil.Disturbance = mean(`Soil.Disturbance`), 
                                          div_name = names(sort(table((species)), decreasing=T)[1]))
```


```{r}
# assign to eva_names
eva_names <- left_join(eva_names, div, by= c("name"="name"))

# assign to eva
eva$div_name <- eva_names$div_name[match(eva$name, eva_names$name)]
eva$Disturbance.Severity  <- eva_names$Disturbance.Severity[match(eva$name, eva_names$name)]
eva$Disturbance.Severity.herblayer <- eva_names$Disturbance.Severity.herblayer[match(eva$name, eva_names$name)]
eva$Disturbance.Frequency <- eva_names$Disturbance.Frequency[match(eva$name, eva_names$name)]
eva$Disturbance.Frequency.herblayer <- eva_names$Disturbance.Frequency.herblayer[match(eva$name, eva_names$name)]
eva$Grazing.Pressure <- eva_names$Grazing.Pressure[match(eva$name, eva_names$name)]
eva$Mowing.Frequency <- eva_names$Mowing.Frequency[match(eva$name, eva_names$name)]
eva$Soil.Disturbance <- eva_names$Soil.Disturbance[match(eva$name, eva_names$name)]

# calculate percentages
sum(!is.na(eva$div_name))/length(eva$PlotObservationID)
sum(!is.na(eva_names$div_name))/length(eva_names$name)
```


# 3 ANALYSIS
## 3.1 Aggregate
Aggregating EVA (#plant per plot and average indicator values) and calculating all kinds of weighting methods for the Eive values.

We have only four options (first three are weighted means)
-   square root of coverage* IV
-   1/niche width * IV
-   square root of coverage* 1/niche width * IV
-   normal mean IV
Following Scherrer et al (2019), Kafer et al (2004) and Descombes et al (2020), the mean indicator value is best correlated with reality but criticized by Tolygesi (2014) 
Following Saatkamp et al (2022), 1/ niche width very good

```{r}
# Function to compute the square root of coverage weighted indicator value 
# Calculate the weight for every species in the plot, then take the indicator value and finally compute the weighted mean
computeSquareRootWeightedMean <- function(indicatorValue, coverage) {
  weights <- sqrt(coverage[!is.na(indicatorValue)])
  indicators <- indicatorValue[!is.na(indicatorValue)]
  weighted.mean(indicators, weights)
}

# Function to compute 1/niche-width weighted mean
NWwmean <- function(indicatorValue, nicheWidth) {
  weights <- 1 / nicheWidth[!is.na(indicatorValue)]
  indicators <- indicatorValue[!is.na(indicatorValue)]
  weighted.mean(indicators, weights)
}

# Function to compute sqrt(cover)/niche-width weighted mean
cNWwmean <- function(indicatorValue, cover, nicheWidth) {
  weights <- sqrt(cover[!is.na(indicatorValue)]) / nicheWidth[!is.na(indicatorValue)]
  indicators <- indicatorValue[!is.na(indicatorValue)]
  weighted.mean(indicators, weights)
}
```


Run aggregation
```{r}
begin<- Sys.time()

aggregatedEva <- eva |> 
  group_by(PlotObservationID) |> 
  summarise(
    numberOfVascularPlantSpecies = length(unique(name)),
    
    EIVEresM.sqrt = computeSquareRootWeightedMean(EIVEresM, `Cover %`), 
    EIVEresN.sqrt = computeSquareRootWeightedMean(EIVEresN, `Cover %`), 
    EIVEresR.sqrt = computeSquareRootWeightedMean(EIVEresR, `Cover %`), 
    EIVEresL.sqrt = computeSquareRootWeightedMean(EIVEresL, `Cover %`),
    EIVEresT.sqrt = computeSquareRootWeightedMean(EIVEresT, `Cover %`),

    EIVEresM = mean(EIVEresM, na.rm = T), 
    EIVEresN = mean(EIVEresN, na.rm = T), 
    EIVEresR = mean(EIVEresR, na.rm = T), 
    EIVEresL = mean(EIVEresL, na.rm = T), 
    EIVEresT = mean(EIVEresT, na.rm = T), 

    DistSeverity.sqrt = computeSquareRootWeightedMean(Disturbance.Severity, `Cover %`), 
    #DistSeverityHerblayer.sqrt = computeSquareRootWeightedMean(Disturbance.Severity.herblayer, `Cover %`),
    DistFrequency.sqrt = computeSquareRootWeightedMean(Disturbance.Frequency, `Cover %`),
    #DistFrequencyHerblayer.sqrt = computeSquareRootWeightedMean(Disturbance.Frequency.herblayer, `Cover %`),
    Grazing.Pressure.sqrt = computeSquareRootWeightedMean(Grazing.Pressure, `Cover %`),
    Mowing.Frequency.sqrt = computeSquareRootWeightedMean(Mowing.Frequency, `Cover %`),
    Soil.Disturbance.sqrt = computeSquareRootWeightedMean(Soil.Disturbance, `Cover %`),
    
    DistSeverity = mean(Disturbance.Severity, na.rm = T),
    #DistSeverityHerblayer = mean(Disturbance.Severity.herblayer, na.rm = T),
    DistFrequency = mean(Disturbance.Frequency, na.rm = T),
    #DistFrequencyHerblayer = mean(Disturbance.Frequency.herblayer, na.rm = T),
    Grazing.Pressure=mean(Grazing.Pressure, na.rm=T),
    Mowing.Frequency= mean(Mowing.Frequency, na.rm=T),
    Soil.Disturbance=mean(Soil.Disturbance, na.rm=T),
 
    logDistSeverity = mean(log(Disturbance.Severity), na.rm=T),
    logDistFrequency = mean(log(Disturbance.Frequency), na.rm = T),
    logGrazing.Pressure=mean(log(Grazing.Pressure), na.rm=T),
    logMowingFrequency=mean(log(Mowing.Frequency), na.rm=T),
    logSoil.Disturbance=mean(log(Soil.Disturbance), na.rm=T)
    
    # Here two previous option I do not know why this method
    # Nevermind, otherwise you create -inf values when it was 0
    #logDistSeverity = mean(log(Disturbance.Severity + 1/2*min(div$Disturbance.Severity[div$Disturbance.Severity>0])), na.rm = T),
    #logDistFrequency = mean(log(Disturbance.Frequency + 1/2*min(div$Disturbance.Frequency[div$Disturbance.Frequency>0])), na.rm = T)
    )
end<- Sys.time()
round(end-begin)
```


Join the result with the header data:
```{r}
fullPlotData<- subset(header, select= c(PlotObservationID, Country,Longitude, Latitude, Region, Dataset, `Relevé area (m²)`,`Date of recording`))
fullPlotData <- full_join(fullPlotData, aggregatedEva, by = "PlotObservationID")
names(fullPlotData)[names(fullPlotData) == 'Relevé area (m²)'] <- 'Area'
names(fullPlotData)[names(fullPlotData) == 'Date of recording'] <- 'Date'
```



## 3.2 Filter SR
Check which plots are extremely species rich:
```{r}
# Which plots SR>100
highDiversityPlots <- fullPlotData[fullPlotData$numberOfVascularPlantSpecies > 100,]

# Jurgen plots suspiciously high
highDiversityPlotsToBeRemoved <- c(842839, 843192, 842921) # plots published by Juergen which are for sure incorrect

revision_removal <- highDiversityPlotsToBeRemoved

highDiversityPlots$Dataset<- as.character(highDiversityPlots$Dataset)
# Look at datasets containing the plots with SR> 100
highDiversityDatasets <- unique(highDiversityPlots$Dataset)

# Look when SR is high for more than two plots in a dataset and obtain their observation ID
for(dataset in highDiversityDatasets) {
  # take all plots for a specific dataset
  plotsInDataset <- highDiversityPlots[highDiversityPlots$Dataset == dataset,]
  # take per amount of SR the frequency this specific SR is present
  multiplicityOfRichnessOccurrence <- as.data.frame(table(plotsInDataset$numberOfVascularPlantSpecies))
  # take all times the same SR is present multiple times in a plot
  multiplicityOfRichnessOccurrence <- multiplicityOfRichnessOccurrence[multiplicityOfRichnessOccurrence$Freq >= 2,] # we want to remove only the plots which > 100 species and suspiciously frequent in the database
  for(richness in multiplicityOfRichnessOccurrence$Var1) {
    # take the index of all plots in the dataset with this species richness
    plotsToBeRemoved <- plotsInDataset[plotsInDataset$numberOfVascularPlantSpecies == richness,]
    highDiversityPlotsToBeRemoved <- append(highDiversityPlotsToBeRemoved, plotsToBeRemoved$PlotObservationID)
  }
}
highDiversityPlots[highDiversityPlots$PlotObservationID %in% highDiversityPlotsToBeRemoved,]

numberOfHighDiversityPlotsToBeRemoved <- length(highDiversityPlotsToBeRemoved)

# remove the too high diversiy plots from the dataset
fullPlotData <- fullPlotData[!(fullPlotData$PlotObservationID %in% revision_removal),]

# make a new highdiversityplots dataset for all true diverse plots (according to the method)
highDiversityPlots <- fullPlotData[fullPlotData$numberOfVascularPlantSpecies > 100,]
# plot to have an idea how the SR is distributed
# hist(fullPlotData$numberOfVascularPlantSpecies)
```


```{r}
dup_check <- highDiversityPlots[duplicated(highDiversityPlots[, c(5,9)]) | duplicated(highDiversityPlots[, c(5,9)], fromLast=T), ]

check <- eva[eva$PlotObservationID %in% dup_check$PlotObservationID, ]

list_species <-  check |> group_by(PlotObservationID) %>%
  summarise(species_set = paste(sort(unique(name)), collapse = ","), .groups = "drop")

remove <- which(duplicated(list_species$species_set))
length(remove)
fullPlotData <- fullPlotData[!fullPlotData$PlotObservationID %in%
                               list_species$PlotObservationID[remove],]
```

`r numberOfHighDiversityPlotsToBeRemoved` of the plots were removed because they have suspiciously high, non unique richness in the same database or are known to be incorrect.

## 3.3 Filter IV
Filtering out plots for which some of the indicator values could not be estimated:
```{r}
# Cells with NA in fullPlot
fullPlotDataNas <- is.na(fullPlotData)
# Get the plot observations
plotsForWhichNotAllIndicatorValuesWereEstimated <- fullPlotDataNas[,1]
# 9 is number of vascular plants
for(i in 9:dim(fullPlotDataNas)[2]) {
  plotsForWhichNotAllIndicatorValuesWereEstimated <- plotsForWhichNotAllIndicatorValuesWereEstimated | fullPlotDataNas[,i]
}
numberOfPlotsWithMissingIndicatorValues <- sum(plotsForWhichNotAllIndicatorValuesWereEstimated)

# remove the plots without an indicator value
fullPlotData <- fullPlotData[!plotsForWhichNotAllIndicatorValuesWereEstimated,]
```


At least 80% species
```{r}
# remove plots with less than 80% cover EIVE
test <- eva |> group_by(PlotObservationID) |> summarise(n = sum(!is.na(eive_name))/ length(eive_name))
sum(test$n >=0.80)/ length(test$n)
sum(test$n <0.80)
fullPlotData <- fullPlotData[fullPlotData$PlotObservationID %in% test$PlotObservationID[test$n >= 0.80],]

# remove plots with less than 80% cover DIV
test <- eva |> group_by(PlotObservationID) |> summarise(n = sum(!is.na(div_name))/ length(name))
sum(test$n >=0.80)/ length(test$n)
sum(test$n < 0.80)
fullPlotData <- fullPlotData[fullPlotData$PlotObservationID %in% test$PlotObservationID[test$n >= 0.80],]
hist(test$n[test$n<0.95])
```




## 3.4 Filter date
```{r}
fullPlotData$Date<- as.Date(fullPlotData$Date, format = "%d.%m.%Y")
numberOfPlotsWithoutDate<-length(fullPlotData$PlotObservationID[is.na(fullPlotData$Date)])
fullPlotData<- fullPlotData[!is.na(fullPlotData$Date),]
fullPlotData <- fullPlotData[fullPlotData$Date >= "1980-01-01",]
```

## 3.5 Duplicates
```{r}
# new dataset to adapt
data <- fullPlotData
data$all <- paste(data$Longitude, data$Latitude)

# dataset duplicated plots
dup <- data[duplicated(data[, c(9,10:19, 25:29,35)]) | 
              duplicated(data[, c(9,10:19, 25:29,35)], fromLast=T),]

# remove those with even equal size
remove <- dup$PlotObservationID[duplicated(dup[, c(7,8,9,10:19, 25:29,35)])]
length(remove)
fullPlotData <- fullPlotData[!fullPlotData$PlotObservationID %in% remove,]
data <- data[!data$PlotObservationID %in% remove,]

# remove already these plots
dup <- dup[!dup$PlotObservationID %in% remove,]
dup <- dup[duplicated(dup[, c(8,9,10:19, 25:29,35)]) |
           duplicated(dup[, c(8,9,10:19, 25:29,35)], fromLast=T),]

# Define the columns used to identify duplicates
dup_colnames <- colnames(dup)[ c(8, 9, 10:19, 25:29, 35)]
# Keep only duplicates
dup_filtered <- dup %>% group_by(across(all_of(dup_colnames))) %>% filter(n() > 1) %>% slice_max(order_by = Area, n = 1) %>% ungroup()
remove <- dup$PlotObservationID[!dup$PlotObservationID %in% dup_filtered$PlotObservationID]
length(remove)
fullPlotData <- fullPlotData[!fullPlotData$PlotObservationID %in% remove,]
data <- data[!data$PlotObservationID %in% remove,]

# check also for plots with duplicate coordinates, date and species richnesses
dup <- data[duplicated(data[, c(8,35,9)]) | 
              duplicated(data[, c(8,35,9)], fromLast=T),]
# this is not feasible to do
dup <- dup[duplicated(dup[, c(9,10:19, 25:29,35)]) | 
              duplicated(dup[, c(9,10:19, 25:29,35)], fromLast=T),]
remove <- dup$PlotObservationID[duplicated(dup[, c(9, 10:19, 25:29,35)])]
length(remove)
fullPlotData <- fullPlotData[!fullPlotData$PlotObservationID %in% remove,]
data <- data[!data$PlotObservationID %in% remove,]


# check also for plots with duplicate coordinates, date and species richnesses
dup <- data[duplicated(data[, c(8,35,9)]) | 
              duplicated(data[, c(8,35,9)], fromLast=T),]

remove <- dup$PlotObservationID[duplicated(dup[, c(8,35,9)])| duplicated(dup[, c(8,35,9)], fromLast=T)]
length(remove)

test <- eva[eva$PlotObservationID %in% remove, c(1,6)]
test$all <- data$all[match(test$PlotObservationID, data$PlotObservationID)]


# Group by plot and collect species as sorted concatenated string
species_by_plot <- test %>%
  group_by(PlotObservationID, all) %>%
  summarise(species_set = paste(sort(unique(name)), collapse = ","), .groups = "drop")

# Find plots with identical species sets
duplicate_species_sets <- species_by_plot %>%
  group_by(species_set, all) %>%
  filter(n() > 1) %>%
  summarise(plots_with_same_species = list(PlotObservationID), .groups = "drop")

# Optional: see number of matching plots per group
duplicate_species_sets$plot_count <- sapply(duplicate_species_sets$plots_with_same_species, length)

# View result


# Step 3: Flatten the list and identify which PlotObservationIDs to remove
to_remove <- unlist(lapply(duplicate_species_sets$plots_with_same_species,
                           function(ids) ids[-1]))  # keep the first, remove the rest

# Step 4: Remove duplicates from the original `test` dataset
test <- test[!test$PlotObservationID %in% to_remove,]

dup <- dup[!dup$PlotObservationID %in% to_remove,]

length(to_remove)

fullPlotData <- fullPlotData[!fullPlotData$PlotObservationID %in% to_remove,]
```


Remove all from data
```{r}
# remove the removed plots also from eva, header and aggregated eva
aggregatedEva <- aggregatedEva[aggregatedEva$PlotObservationID %in% fullPlotData$PlotObservationID,]
header <- header[header$PlotObservationID %in% fullPlotData$PlotObservationID,]
eva <- eva[eva$PlotObservationID %in% fullPlotData$PlotObservationID,]
```

`r numberOfPlotsWithMissingIndicatorValues` of the plots were removed because not all indicator values could be estimated.

## 3.6 Uncertainty
```{r}
summary(header$`Location uncertainty (m)`)
numberUncertain <- header$PlotObservationID[header$`Location uncertainty (m)`>1000]
length(numberUncertain)

header <- header[!header$PlotObservationID %in% numberUncertain,]
fullPlotData <- fullPlotData[!fullPlotData$PlotObservationID  %in% numberUncertain,]
```


# 6 EXPORT
## 6.1 EVA + HEAD
```{r}
# vector with eva names
eva_names<- as.data.frame(unique(eva$`Matched concept`))
colnames(eva_names) <- "Matched concept"
eva_names$name <- eva$name[match( eva_names$`Matched concept`, eva$`Matched concept`)]
eva_names$species <- eva$species[match( eva_names$`Matched concept`, eva$`Matched concept`)]
eva_names$eive <- eva$eive_name[match( eva_names$`Matched concept`, eva$`Matched concept`)]
eva_names$div <- eva$div_name[match( eva_names$`Matched concept`, eva$`Matched concept`)]


# Make sure all plot data is in eva
fullPlotData <- fullPlotData[fullPlotData$PlotObservationID %in% eva$PlotObservationID,]
fullPlotData$Dataset <- as.factor(as.character(fullPlotData$Dataset))
write_csv(fullPlotData, "../EVA data/fullPlotData_4.csv")

# Make sure all eva data has a plot 
eva <- eva[eva$PlotObservationID %in% fullPlotData$PlotObservationID,]
fullPlotEva <- eva
head(eva)
write_csv(eva, "../EVA data/fullPlotEva_4.csv")
```




## 6.2 Check neophytes
```{r}
# File for Irena
# Eva
eva2<- fullPlotEva[,c("PlotObservationID","species","irena", "Matched concept", "name","source")]
fullPlot2<- fullPlotData[,c("PlotObservationID","Region")]

# eva names
eva<- fullPlotEva
eva_names <- unique(eva2[, c(2:6)])


# Join and summarise
eva_country<- left_join(eva2, fullPlot2)
species_country<- eva_country |> group_by(Region, species, irena, `Matched concept`, name, source) |> summarize(n=n())

# load neophyte data
neophyte<-readxl::read_excel("../Extra data/Species names/Irena_taxa_update.xlsx")

# relate species
neophyte$name <- eva_names$name[match(neophyte$species, eva_names$name)]
neophyte$source <- NA
neophyte$source <- eva_names$source[match(neophyte$species, eva_names$name)]
# 23995
sum(is.na(neophyte$name))
neophyte$name[is.na(neophyte$name)] <- eva_names$name[match(neophyte$species[is.na(neophyte$name)], eva_names$species)]
neophyte$source[is.na(neophyte$source)] <- eva_names$source[match(neophyte$species[is.na(neophyte$source)], eva_names$species)]
# 2300
sum(is.na(neophyte$name))
neophyte$name[is.na(neophyte$name)] <- eva_names$name[match(neophyte$species[is.na(neophyte$name)], eva_names$irena)]
neophyte$source[is.na(neophyte$source)] <- eva_names$source[match(neophyte$species[is.na(neophyte$source)], eva_names$irena)]
# 1838
sum(is.na(neophyte$name))
neophyte$name[is.na(neophyte$name)] <- eva_names$name[match(neophyte$species[is.na(neophyte$name)], eva_names$`Matched concept`)]
neophyte$source[is.na(neophyte$source)] <- eva_names$source[match(neophyte$species[is.na(neophyte$source)], eva_names$`Matched concept`)]
# 1649
sum(is.na(neophyte$name))

# relate matched concept
neophyte$name[is.na(neophyte$name)] <- eva_names$name[match(neophyte$`Matched concept`[is.na(neophyte$name)], eva_names$`Matched concept`)]
neophyte$source[is.na(neophyte$source)] <- eva_names$source[match(neophyte$`Matched concept`[is.na(neophyte$source)], eva_names$`Matched concept`)]
# 897
sum(is.na(neophyte$name))



# unknown species previously in database and not anymore
unknown <- neophyte[is.na(neophyte$name),]

# remove the 'plants' that were just used for the species richness
species_country <- species_country[!species_country$source=="Plant",]

# some are duplicated --> if name, region and status are equal can be removed
x <- neophyte[duplicated(neophyte[, c(2, 10,8)]) | duplicated(neophyte[, c(2, 10,8)], fromLast=T) ,]
# remove duplicates in neophyte --> 97518 --> 97504
neophyte <- neophyte[!duplicated(neophyte[, c(2, 10,8)]),]

# some are duplicated --> look at them --> 230
dup <-  neophyte[duplicated(neophyte[, c(10,2)]) | duplicated(neophyte[, c(10,2)], fromLast = T) ,]
# remove the status column
dup <- dup[,-c(8)]
#write_csv(dup, "merging_problems_status.csv")
# give status of species (so not subspecies or varieties)
change <- neophyte[neophyte$species %in% stringr::str_remove(dup$name," aggr."),]
change <- change[, c(2,10,8)]


# check species for which this was not possible
check <- change[duplicated(change[,c(1,2)]) | duplicated(change[,c(1,2)], fromLast = T) ,]
# one species --> give correct 
change$statusNew[duplicated(change[,c(1,2)]) | duplicated(change[,c(1,2)], fromLast = T)] <- "native"
# end with removing all duplicates (this species)
change <- change[!duplicated(change),]
# merge
dup<- left_join(dup, change, by= c("name"="name", "Region"="Region"))



# check which are still na
unique(dup$name[is.na(dup$statusNew)])
# give correct species classifications
change <- data.frame( name= c("Vicia sativa","Symphyotrichum novi-belgii aggr.","Rubus fruticosus aggr."),
                      status= c("native","extra","native"))
# change their status
dup$statusNew[is.na(dup$statusNew)] <- 
  change$status[match(dup$name[is.na(dup$statusNew)], change$name)]
dup <- dup |> relocate(statusNew, .after= n)
# change these
neophyte[duplicated(neophyte[, c(10,2)]) | duplicated(neophyte[, c(10,2)], fromLast = T),] <- dup
neophyte <-neophyte[!duplicated(neophyte[, c(10,2,8)]),]

# also remove duplicates in the species country list
species_country <- species_country[!duplicated(species_country[,c(5,1)]),]


# merge --> 97727 --> 97591 --> 97558 --> 97391 (97424)
species_country <- full_join(species_country, neophyte[, -c(1,3:5,6,7,9,11)], by= c("name"="name", "Region"="Region"))
species_country <- species_country[!is.na(species_country$species),]
head(species_country)

neophyte <- neophyte[, -c(1,6,7,9)]

head(neophyte)

# save
```



```{r}
# check na's
x <- species_country[is.na(species_country$statusNew),]
x$statusNew <- neophyte$statusNew[match(x$irena, neophyte$irena)]
species_country$statusNew[is.na(species_country$statusNew)] <-
  neophyte$statusNew[match(species_country$irena[is.na(species_country$statusNew)], neophyte$irena)]

x <- species_country[is.na(species_country$statusNew),]
# majority is error in ESy --> change vicia sativa
species_country$name[species_country$name=="Vicia ramuliflora"] <- "Vicia sativa"

# last change ourselves
change <- data.frame(Region =c("Turkey", "Turkey","Switzerland",
                               "Ukraine","France","Switzerland",
                               "France","Italy", "France","Albania",
                               "Lithuania","Turkey","RF.CS","Turkey"),
                     species = c("Alyssum bornmuelleri",
                                 "Astragalus xylorrhizus",
                                 "Crepis pygmaea",
                                 "Cytisus wulfii",
                                 "Festuca christianii-bernardii",
                                 "Limnobium palustre",
                                 "Panicum hillmanii",
                                 "Phyllolepidum rupestre",
                                 "Sagina subulata var. subulata",
                                 "Sideritis raeseri",
                                 "Ranunculus penicillatus",
                                 "Rosa dumalis subsp. antalyensis",
                                 "Suaeda crassifolia",
                                 "Veronica orientalis subsp. carduchorum"),
                     status= c("native","native","native",
                               "native","native","native",
                               "native","native", "native",
                               "native","native","native",
                               "native","native"))

x$statusNew <- "native"

species_country$statusNew[is.na(species_country$statusNew)]<- "native"

name_change <- data.frame(old.name= c("Limnobium palustre",
                                      "Sagina subulata var. subulata",
                                      "Rosa dumalis subsp. antalyensis",
                                      "Veronica orientalis subsp. carduchorum"),
                          new.name= c("Hygrohypnum luridum",
                                      "Sagina subulata",
                                      "Rosa boissieri",
                                      "Veronica orientalis"))

species_country$species[species_country$species %in% name_change$old.name] <-
  name_change$new.name[match(species_country$species[species_country$species %in% name_change$old.name], name_change$old.name)]
species_country$name[species_country$name %in% name_change$old.name] <-
  name_change$new.name[match(species_country$name[species_country$name %in% name_change$old.name], name_change$old.name)]

eva$species[eva$species %in% name_change$old.name] <-
  name_change$new.name[match(eva$species[eva$species %in% name_change$old.name], name_change$old.name)]
eva$name[eva$name %in% name_change$old.name] <-
  name_change$new.name[match(eva$name[eva$name %in% name_change$old.name], name_change$old.name)]

write_csv(species_country,"../EVA data/country_species_new.csv")
write_csv(eva, "../EVA data/fullPlotEva_4.csv")
```
