==> 4.2_ENS_Calculation.Rmd <==
---
title: "Species diversity and eveness"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---

# 1 DATA
## 1.1 Preparation
We first clean the environment.
```{r, echo=T}
rm(list=ls())
```

Subsequently we load the required packages.
```{r, warnings=F, message=F}
library(tidyverse)
library(doParallel)
library(vegan)
```

## 1.2 Data loaded
Then, we load data of EVA
```{r, warnings=F, message=F}
# Eva data for which plots are available and after correcting for incorrect or weird plot observations
eva <- read_csv("../EVA data/fullPlotEva_new.csv")
```

Load filtered Data containing the richness and indicator values for each plot. In case we want to save computing time: down sample it.
```{r, warnings=F, message=F}
# header like data, filtered and appended with EIVE and DIV
fullPlotData <- read_csv("../EVA data/fullPlotData_new.csv")

fast <- F
if(fast) {
  fullPlotData <- fullPlotData[runif(length(fullPlotData$PlotObservationID)) > 0.90,]
  eva <- eva[eva$PlotObservationID %in% fullPlotData$PlotObservationID,]
}
```


Load the neophyte definitions:
```{r, message=F}
# Data on which species are neophytes
native_intra_analysis=F
if(native_intra_analysis){
  species_country_status<- read_csv("../EVA data/country_species_new.csv", show_col_types = FALSE)
} else{
  species_country_status<- read_csv("../EVA data/country_species_new.csv", show_col_types = FALSE)
  species_country_status$Neophyte[species_country_status$Neophyte=="native_intra"] <- "native"
}

# Assign these names to the eva list
extra_EU <- unique(species_country_status$species[species_country_status$Neophyte=="extra"])
intra_EU <- unique(species_country_status$species[species_country_status$Neophyte=="intra"])
native_intra <- unique(species_country_status$species[species_country_status$Neophyte=="native_intra"])
native <-  unique(species_country_status$species[species_country_status$Neophyte=="native"])
```


Assign classification to species
```{r}
# Only observation ID and Region
fullPlot2<- fullPlotData[,c("PlotObservationID","Region")]
# Right join to keep only species present in fullplot (otherwise a lot of NAs)
eva<- right_join(eva, fullPlot2, by = c("PlotObservationID"="PlotObservationID"))
# Join eva and classification
eva_country_neophyte<- left_join(eva, species_country_status[, -c(2:4,6:8)], by= c("Region"= "Region", "name"= "name"))
eva<- eva_country_neophyte
# Look at how much from every type are present
tt<- as.data.frame(table(eva_country_neophyte$Neophyte))
tt

rm(fullPlot2, eva_country_neophyte, tt)
```


# 2 ENS
## 2.1 Prepare
```{r}
# check duplicates species
check <- unique(eva[, c(6,10,21)])
length(unique(eva$name))
y <-check[duplicated(check$name) | duplicated(check$name, fromLast = T),]
unique(y$name)


# Calculate fisher
# not necessary here yet, as doing so would merge some species that are different subspecies per plot. It is important to note that this file takes this into account by calculating the ENSO only on species names (so not subspecies removed through EUNIS-ESy)


# reduce size eva
eva<- eva[, c(1,6,9)]
# change name
colnames(eva)[3]<- "cover"

plant <- eva[eva$name=="Plant",]
number <- c("one","two","three","four","five","six","seven","eight","nine","ten","eleven","twelve","thirteen",
            "fourteen","fifteen","sixteen")

for(i in unique(plant$PlotObservationID)) {
 x<- plant[plant$PlotObservationID==i,]
  if(length(x>1)){
    for(j in 1:nrow(x)){
        x$name[j] <- paste("Plant",number[j], sep = " ")
    }
    plant[plant$PlotObservationID==i,] <- x
  }
}

eva[eva$name=="Plant",] <- plant

# check data layer
# eva$Layer <- as.factor(eva$Layer)
# We now have some trees, which can possibly affect our calculations
```


## 2.2 Hill numbers
```{r}
# make cluster
n.cores<- parallel::detectCores()-6
my.cluster <- parallel::makeCluster(n.cores, type = "PSOCK" )
print(my.cluster)
doParallel::registerDoParallel(cl = my.cluster)
foreach::getDoParRegistered()
foreach::getDoParWorkers()

# get all unique plot IDs required to split our database
plotID<- unique(eva$PlotObservationID)
which(plotID == 1522409)
# run for all and time
i=99
begin<-Sys.time()
x<- foreach(i= 1:99, .combine = "rbind", .packages = c("dplyr","mgcv","tidyr","vegan")) %dopar% {
  # create the length of the to be calculated data chunk
  j <- ceiling(length(plotID)/99)
  # create begin and end values for the plot observation ID
  begin<- ((i-1)*j+1)
  end<- (i*j)
  # make the vegetation dataframe for this chunk
  matrix<- pivot_wider(eva[eva$PlotObservationID %in% plotID[begin:end],], names_from= c("name"),
            values_from =("cover" ), values_fill = 0, values_fn = sum)
  # get the plot values
  plots<- matrix[,1]
  # remove the plot values (required to calculate the hill values)
  matrix<-matrix[,-1]
  # calculate hill values for chunk
  hill<- tsallis(matrix, hill=TRUE, scales=c(0,1,2))
  # bind plot ID
  hill<- cbind(plots, hill)
  # show the total hill dataframe, which will be binded together
  hill
}
# calculate time required
end<-Sys.time()
round(end-begin)
colnames(x)<- c("PlotObservationID","ENS0","ENS1","ENS2")
# stop cluster
stopCluster(my.cluster)
```


## 2.3 Analysis
```{r}
# check the values --> some difference --> related to summing different layers 
fullPlotData$PlotObservationID[!fullPlotData$PlotObservationID %in% x$PlotObservationID]
all.equal(x[,2], fullPlotData$numberOfVascularPlantSpecies)
x[x$ENS0 != fullPlotData$numberOfVascularPlantSpecies,]

# what causes it to be different
fullPlotData[fullPlotData$numberOfVascularPlantSpecies != x$ENS0, c(1,2,4,9)]
eva[eva$PlotObservationID=="205618",]
fullPlotData[fullPlotData$PlotObservationID=="205618",]
# some species were present multiple times

# these are all plots
test <- eva[eva$PlotObservationID %in% x$PlotObservationID[x$ENS0 != fullPlotData$numberOfVascularPlantSpecies],]
length(unique(test$PlotObservationID))
y <- plant |> group_by(PlotObservationID)|> summarise(n=n())
sum(y$n>1)

# so this is due to classifying previously as 'plant' and not 'plant 1'
setdiff(y$PlotObservationID[y$n>1], test$PlotObservationID)

# plot
apply(x[, 2:4], MARGIN=2, FUN = hist)

```



## 2.4 Save
```{r}
save=T
if(save){
  fullPlotData<- left_join(fullPlotData, x)
  head(fullPlotData)
  write_csv(fullPlotData, "../EVA data/fullPlotData_new.csv")
}
```
