==> 5.1_Cover_New_Classes.Rmd <==
---
title: 'Abundance analysis based on new classification'
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---

This script analyses whether neophytes are more dominant and frequent than native plants or not:

# 1 DATA
## 1.1 Preparation
We first clean the environment.
```{r, echo=T}
#rm(list=ls())
```

Subsequently we load the required packages.
```{r, warnings=F, message=F}
library(tidyverse)
library(lmerTest)
library(ggpubr)
library(vegdata)
library(doParallel)
library(emmeans)
library(DHARMa)
library(glmmTMB)
```

## 1.2 Data loaded
This script analyses whether neophytes are more dominant and frequent than native plants or not.  
First, we load data of EVA
```{r, warnings=F, message=F}
# Eva data for which plots are available and after correcting for incorrect or weird plot observations
eva <- read_csv("../EVA data/fullPlotEva_new.csv")
```

Load filtered Data containing the richness and indicator values for each plot. In case we want to save computing time: down sample it.
```{r, warnings=F, message=F}
# header like data, filtered and appended with EIVE and DIV
normal = T
if(normal){
  fullPlotData <- read_csv("../EVA data/fullPlotData_new.csv")
  fullPlotData$Date <- as.Date(fullPlotData$Date, format = c("%Y-%m-%d"))
  fullPlotData <- fullPlotData[fullPlotData$Date >= "1980-01-01",]
} else {
  fullPlotData_EUNIS <- read_csv("../EVA data/fullPlotData_new.csv")
  fullPlotData$Date <- as.Date(fullPlotData$Date, format = c("%Y-%m-%d"))
  fullPlotData <- fullPlotData[fullPlotData$Date >= "1980-01-01",]
  habitats <- unique(fullPlotData$habitat)
  # adapt here the type you want to assess, e.g. 'grasslands'=1
  fullPlotData <- fullPlotData[fullPlotData$habitat==habitats[3] & !is.na(fullPlotData$habitat),]
  eva <- eva[eva$PlotObservationID %in% fullPlotData$PlotObservationID,]
}


fast <- F
if(fast) {
  fullPlotData <- fullPlotData[runif(length(fullPlotData$PlotObservationID)) > 0.50,]
  eva <- eva[eva$PlotObservationID %in% fullPlotData$PlotObservationID,]
}
```

Load the neophyte definitions:
```{r, message=F}
# Data on which species are neophytes
native_intra_analysis=F
if(native_intra_analysis){
  species_country_status<- read_csv("../EVA data/country_species_new.csv", show_col_types = FALSE)
} else{
  species_country_status<- read_csv("../EVA data/country_species_new.csv", show_col_types = FALSE)
  species_country_status$Neophyte[species_country_status$Neophyte=="native_intra"] <- "native"
}

# Assign these names to the eva list
extra_EU <- unique(species_country_status$species[species_country_status$Neophyte=="extra"])
intra_EU <- unique(species_country_status$species[species_country_status$Neophyte=="intra"])
native_intra <- unique(species_country_status$species[species_country_status$Neophyte=="native_intra"])
native <-  unique(species_country_status$species[species_country_status$Neophyte=="native"])
```


Assign classification to species
```{r}
# Only observation ID and Region
fullPlot2<- fullPlotData[,c("PlotObservationID","Region")]
colnames(eva)
# Right join to keep only species present in fullplot (otherwise a lot of NAs)
eva<- right_join(eva, fullPlot2, by = c("PlotObservationID"="PlotObservationID"))
# Join eva and classification
colnames(eva)

eva_country_neophyte<- left_join(eva, species_country_status[, -c(2:4,6:7)], by= c("Region"= "Region", "name"= "name"))
# Look at how much from every type are present
table(eva_country_neophyte$Neophyte)
eva <- eva_country_neophyte

colnames(eva_country_neophyte)[9] <- "Cover %"
```


# 2 COVER
## 2.1 Synthesis
```{r}
fast=F
if(fast){
  # For speed optimization we extract only the necessary information from EVA
  reducedEva <- eva_country_neophyte |> select(PlotObservationID, Region,name, Neophyte,
                                               Layer,`Cover %`)
  
  # Add covers per layer
  reducedEva <- reducedEva |> group_by(PlotObservationID, Region,name, Neophyte, Layer) |>
    summarise(cover= sum(`Cover %`))
  
  # Fisher for multiple layers
  reducedEva <- reducedEva|> group_by(PlotObservationID, Region, name, Neophyte) |> 
    summarise(cover = (1-prod((1-ifelse(cover/100>1, 1, cover/100 ))))*100)
  
  # list unique species
  uniqueSpecies <- unique(reducedEva[, c(3,4)])
  
  # add total cover to fullPlotData
  fullPlotData <- reducedEva %>% group_by(PlotObservationID) %>% 
    summarise(totalCover = sum(cover)) %>% 
    left_join(fullPlotData, by = "PlotObservationID")
  
  # Copy the full dataset without aberrant totalCover
  fullPlotData <- fullPlotData %>% filter(totalCover <= 900)
  reducedEva <- reducedEva[reducedEva$PlotObservationID %in%
                             fullPlotData$PlotObservationID,]

  # Make empty dataframe
  speciesDominance <- data.frame(names = character(), coverMedian = numeric(), 
                                 coverMean = numeric(), var = numeric(), 
                                 numberOfPlots = integer(), neophyte = logical())
  
  # Prepare running in parallel
  parallel::detectCores()
  n.cores <- parallel::detectCores() - 2
  my.cluster <- parallel::makeCluster(n.cores, type = "PSOCK")
  print(my.cluster)
  doParallel::registerDoParallel(cl = my.cluster)
  foreach::getDoParRegistered()
  foreach::getDoParWorkers()
  
  i=1
  # Take time beginning
  begin<- Sys.time()
  # Do calculation of average and median cover+ variance in dataset
 speciesDominance <-foreach(i= 1:nrow(uniqueSpecies), .combine='rbind',
                            .packages=c("dplyr","mgcv")) %dopar% {
    # Get dataset for each unique species
      tmp <- reducedEva[reducedEva$name == uniqueSpecies$name[i]& 
                          reducedEva$Neophyte == uniqueSpecies$Neophyte[i],]
    # Make ID a factor
    tmp$PlotObservationID <- as.factor(tmp$PlotObservationID)
    # Group by plot and calculate the mean species cover for that species in all plots where
    # it is present (this step is not necessary anymore)
    tmp <- tmp |> group_by(PlotObservationID) |> summarise(cover = mean(`cover`))
    
    # We are only interested in the dominance of species occurring in at least 30 plots 
    if(length(tmp$PlotObservationID) >= 30) {
    # Get the summary (median and mean at place 3 and 4) for when the species is present, 
    # also incorporating the variance and number of Plots
      sum <- summary(tmp$cover)
      x <- c(names = uniqueSpecies$name[i], coverMedian = sum[[3]], coverMean = sum[[4]], 
             var = var(tmp$cover),numberOfPlots =length(tmp$PlotObservationID), 
             neophyte = uniqueSpecies$Neophyte[i])
  }
}
# Ending and calculate time it took to perform the calculations
end<- Sys.time()
round(end-begin)

# Make the matrix a dataframe and change the row names
speciesDominance<- as.data.frame(speciesDominance)
row.names(speciesDominance)<- c(1:length(speciesDominance$names))

# Make the Median, Mean, Variance and Number numeric
speciesDominance[,c(2:5)]<- apply(speciesDominance[, c(2:5)], MARGIN= 2, FUN= as.numeric)

# Stop the cluster
parallel::stopCluster(cl = my.cluster)
write_csv(speciesDominance, "../Results/speciesDominance_1980.csv")
} else{
  if(native_intra_analysis){
  speciesDominance <- read.csv("../Results/speciesDominance_1980_native_intra.csv")
  }else{
  speciesDominance<- read.csv("../Results/speciesDominance_1980.csv")
  }
}

```


```{r}
eva_names <- eva |> group_by(name, Neophyte) |> summarise(n=n())
colnames(eva_names) <- c("names","neophyte","n")
sum(eva_names$n >=30)

# difference is due to summing of cover within one plot
setdiff(eva_names[eva_names$n>=30, 1:2], speciesDominance[, c(1,6)])

sum(eva$name=="Juglans ailantifolia")

length(unique(eva$PlotObservationID[eva$name %in% "Juglans ailantifolia"]))
```


We also analyse which species are significantly negative / positive when compared to the distribution of the native species. We do this for all plants.
```{r}
# Get the quantile values of all native plant species.Cutoff values at 0.01 and 0.05 chances (both upper and lower limit)
quantiles <- quantile(speciesDominance$coverMean[speciesDominance$neophyte=="native"], probs = c(0, 0.01, 0.05, 0.95, 0.99, 1))

# Assign each species to a class
speciesDominance$significantCoverMean <- as.factor(cut(speciesDominance$coverMean, breaks = quantiles, labels = c("p=0.01 negative", "p=0.05 negative", "", "p=0.05 positive", "p=0.01 positive"), right = TRUE))
```



## 2.2 Analysis
We perform a small analysis on the data by looking at the mean values for native, intra and extra and analyse whether they are significantly different from each other
```{r}
phylo <- read_csv("../Extra data/Species names/phylo.csv", show_col_types = F)
speciesDominance[, c(8:9)] <- phylo[match(speciesDominance$names, phylo$name), 6:7]

names <- speciesDominance$names

check <-vegdata::parse.taxa(names)

genus <- check[is.na(check$epi1),]

speciesDominance <- speciesDominance[!speciesDominance$names %in% genus$original,]
#genus_country <- species_country_status[species_country_status$name %in% genus$original,]
#write_csv(genus_country, "genus_country.csv")
```



```{r}
# run for intra-European analysis

intra <- 	speciesDominance$names[speciesDominance$neophyte=="intra"]

speciesDominance$neophyte[speciesDominance$names %in% intra & speciesDominance$neophyte=="native"] <- "native_intra"

native_intra_analysis <- T
```




```{r}
# make it similar every time
set.seed(1)

speciesDominance$rel <- speciesDominance$coverMean/100

# how many have we of each origin
table(speciesDominance$neophyte)

speciesDominance

ggplot(speciesDominance, aes(x = rel, y= neophyte, fill = neophyte, colour = neophyte)) +
  ggridges::geom_density_ridges( alpha = 1) +
  ylab("Origin class") +
  xlab("Relative cover") +
  labs(fill = "Neophyte", colour = "Neophyte")

# glmmTMB
speciesDominance$rel <- speciesDominance$coverMean/100
model <- glmmTMB((rel) ~ -1+ neophyte + (1|family/genus/names), data= speciesDominance, family= beta_family() )
summary(model)
summary(emmeans::emmeans(model, pairwise ~ neophyte))

# check assumptions
simulation <- DHARMa::simulateResiduals(model)
plot(simulation)
# does not work


# we will 'bootstrap' or results using 1000 replications
result <- data.frame(contrast = character(), estimate = numeric(), SE= numeric(), df=numeric(), t.ratio=numeric(), p.value= numeric())

# Counter for successful runs
successful_runs <- 0

while (successful_runs < 1000) {
  # sample natives
  natives <- speciesDominance[speciesDominance$neophyte=="native",]
  natives <- natives[sample(nrow(natives),200, replace=F),]
  # sample intras
  intras <- speciesDominance[speciesDominance$neophyte=="intra",]
  intras <- intras[sample(nrow(intras),200, replace=T),]
  # sample extras
  extras <- speciesDominance[speciesDominance$neophyte=="extra",]
  extras <- extras[sample(nrow(extras),200, replace=T),]
  all <- rbind(natives, speciesDominance[!speciesDominance$neophyte=="native",])
  all <- rbind(natives, intras, extras)
  
  
  if(native_intra_analysis){
    native_intra <- speciesDominance[speciesDominance$neophyte=="native_intra",]
    native_intra <- native_intra[sample(nrow(native_intra),200, replace=T),]
    all <- rbind(natives, intras, extras, native_intra)
  }
  
  # model again
  model <-  glmmTMB((rel) ~ -1+ neophyte + (1|family/genus), data= all, family= beta_family() )
  summary(model)
  
  if (model$fit$convergence != 0) next
  
  # check pairwise differences
  emmeans_results <- emmeans::emmeans(model, pairwise ~ neophyte)
  # save in data frame
  sum <- summary(emmeans_results)$contrasts
  # add to large table
  result <- rbind(result, sum)
  
  successful_runs <- successful_runs+1
}
  
# look at aggregated results (mean and sd)
aggregated_result <- aggregate(. ~ contrast, data = result, FUN = function(x) c(mean = mean(x), sd = sd(x)))
  
# assess number of significant
result$significant <- result$p.value < 0.05
table(result$significant)
result |> group_by(contrast, significant) |> summarise(n=n())
```



```{r}
# Take a look at the most dominant species
topdominant<- speciesDominance[order(speciesDominance$coverMean, decreasing=T),]
topdominant<- topdominant[1:100,]
topdominant_intra<- topdominant[topdominant$neophyte=="intra",]
topdominant_extra <- topdominant[topdominant$neophyte=="extra",]

topdominant_alien <- topdominant[topdominant$neophyte != "native",]
library(readxl)
write_excel_csv2(topdominant_alien, "../Results/speciesDominance_top100_aliens.csv")
```


## 2.3 Plot
```{r}
# Violin plot
if(!native_intra_analysis){

speciesDominance <- speciesDominance %>% mutate(Neophyte = factor(neophyte, 
                          levels = c("native", "intra","extra"),
                          labels = c("native in the country", 
                                     "intra-European neophyte", 
                                     "extra-European neophyte")))  
  
  
p<-ggplot(speciesDominance, mapping = aes(x= (neophyte), y = log10(coverMean), group= Neophyte, colour = Neophyte, fill= Neophyte)) + 
  geom_violin(alpha=0.5, scale="width")+
  geom_boxplot(width= 0.25, alpha=0.8, fill="white")+
  theme_pubr()+
  stat_summary(fun= "mean",
               geom = "point", aes(group= Neophyte), size=3)+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"))+
  scale_fill_manual(values = c("#1E88E5", "#FFC107", "#004D40"))+
  guides(color="none")+
  theme(legend.position = "none")+
  scale_y_continuous(breaks = c(-1, 0, 1, 2), labels = c(0.1,1,10,100), limits=c(-1, 2.2))+
  scale_x_discrete(labels=  c("extra-European \n aliens",  "intra-European \n aliens", "native species")) +
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=12), axis.title.x = element_blank(), 
        axis.text.y = element_text(size=10))+
  ylab("Mean cover when present (%)")+
  annotate("text", x = 1, y = max(log10(speciesDominance$coverMean)), 
             label = "a", size = 4,vjust = -0.5, hjust = 0.5, alpha=1) +
  annotate("text", x = 2, y = max(log10(speciesDominance$coverMean)), 
             label = "b", size = 4, vjust = -0.5, hjust = 0.5, alpha=1)+
  annotate("text", x = 3, y = max(log10(speciesDominance$coverMean)), 
             label = "c", size = 4, vjust = -0.5, hjust = 0.5, alpha=1)+
  annotate("text", x=1, y=max(log10(speciesDominance$coverMean)), 
           label= paste("n=",length(speciesDominance$coverMean[speciesDominance$neophyte=="extra"])), 
           vjust = -2, hjust = 0.5, size=4)+
  annotate("text", x=2, max(log10(speciesDominance$coverMean)), 
           label= paste("n=",length(speciesDominance$coverMean[speciesDominance$neophyte=="intra"])), 
           vjust = -2, hjust = 0.5, size=4)+
  annotate("text", x=3, max(log10(speciesDominance$coverMean)), 
           label= paste("n=",length(speciesDominance$coverMean[speciesDominance$neophyte=="native"])), 
           vjust = -2, hjust = 0.5, size=4)+
  labs(subtitle = "Significance level: ***")+
  xlab(NULL)+
  theme(plot.subtitle = element_text( hjust = 0.01))

Dominance <- p
p
ggsave("../Images/Dominance.jpeg", p,width = 5, height = 5)
}


if(native_intra_analysis){
p<-ggplot(speciesDominance, mapping = aes(x= neophyte, y = log10(coverMean), group= neophyte,  fill= neophyte, colour=neophyte)) + 
  # create violin plot with x axis no title and y 
  geom_violin(alpha= 0.3, scale= "width")+
  geom_boxplot(width= 0.25, alpha=0.8, fill="white") +
  theme_pubr()+
  ylab("Mean cover when present (%)") +
  scale_colour_manual(values=c( "#004D40","#FFC107", "#1E88E5", "darkgreen"), 
                    breaks=c("extra","intra", "native",  "native_intra")) +
  scale_fill_manual(values=c( "#004D40","#FFC107", "#1E88E5", "darkgreen"), 
                    breaks=c("extra","intra", "native",  "native_intra")) +
  guides(color="none")+
  theme(legend.position = "none")+
  stat_summary(fun= "mean",
               geom = "point", aes(group= neophyte), size=3)+
  scale_y_continuous(breaks = c(-1, 0, 1, 2), labels = c(0.1,1,10,100), limits=c(-1, 2.2))+
  scale_x_discrete(labels=  c("extra-European \n neophytes",  "intra-European \n neophytes","native species", 
                             "native species \nalien elsewhere")) +
  theme(axis.text.x = element_text(size=10), axis.title.y=element_text(size=12), axis.title.x = element_blank(), 
        axis.text.y = element_text(size=10)) +
  annotate("text", x = 1, y = max(log10(speciesDominance$coverMean)), label = "a", size = 4, 
           vjust = -0.7, hjust = 0.5, alpha=0.8) +
  annotate("text", x = 2, y = max(log10(speciesDominance$coverMean)), label = "b", size = 4, 
           vjust = -0.7, hjust = 0.5, alpha=0.8)+
  annotate("text", x = 3, y = max(log10(speciesDominance$coverMean)), label = "c", size = 4, 
           vjust = -0.7, hjust = 0.5, alpha=0.8)+
  annotate("text", x = 4, y = max(log10(speciesDominance$coverMean)), label = "c", size = 4, 
           vjust = -0.7, hjust = 0.5, alpha=0.8)+
  annotate("text", x=1, y=2.0, label= paste("n=",sum(speciesDominance$neophyte=="extra")), vjust = -2, hjust = 0.5, size=3.5)+
  annotate("text", x=2, y=2.0, label= paste("n=",sum(speciesDominance$neophyte=="intra")), vjust = -2, hjust = 0.5, size=3.5)+
  annotate("text", x=3, y=2.0, label= paste("n=",sum(speciesDominance$neophyte=="native")), vjust = -2, hjust = 0.5, size=3.5)+
  annotate("text", x=4, y=2.0, label= paste("n=",sum(speciesDominance$neophyte=="native_intra")), vjust = -2, hjust = 0.5, size=3.5)
p
Dominance <- p
}
p
```

# 3 FREQUENCY
## 3.1 Analysis
The data is present in speciesDominance, which facilitates the calculations
```{r}
set.seed(1)

# plot to get idea about distribution
ggplot(speciesDominance, aes(x = numberOfPlots, y= neophyte, fill = neophyte, colour = neophyte)) +
  ggridges::geom_density_ridges( alpha = 1) +
  ylab("Origin class") +
  xlab("Relative cover") +
  labs(fill = "Neophyte", colour = "Neophyte")

# glmmTMB
model <- glmmTMB((numberOfPlots) ~ -1+ neophyte + (1|family/genus/names), data= speciesDominance, family= nbinom2() )
summary(model)
summary(emmeans::emmeans(model, pairwise ~ neophyte))

# check assumptions
simulation <- DHARMa::simulateResiduals(model)
plot(simulation)
# does not work


# we will 'bootstrap' or results using 1000 replications
result_freq <- data.frame(contrast = character(), estimate = numeric(), SE= numeric(), df=numeric(), t.ratio=numeric(), p.value= numeric())


# Counter for successful runs
successful_runs <- 0

# Run until we get 1000 successful models
while (successful_runs < 1000) {
  # sample natives
  natives <- speciesDominance[speciesDominance$neophyte=="native",]
  natives <- natives[sample(nrow(natives),200, replace=F),]
  # sample intras
  intras <- speciesDominance[speciesDominance$neophyte=="intra",]
  intras <- intras[sample(nrow(intras),200, replace=T),]
  # sample extras
  extras <- speciesDominance[speciesDominance$neophyte=="extra",]
  extras <- extras[sample(nrow(extras),200, replace=T),]
  all <- rbind(natives, speciesDominance[!speciesDominance$neophyte=="native",])
  all <- rbind(natives, intras, extras)
  
  
  if(native_intra_analysis){
    native_intra <- speciesDominance[speciesDominance$neophyte=="native_intra",]
    native_intra <- native_intra[sample(nrow(native_intra),200, replace=F),]
    all <- rbind(natives, intras, extras, native_intra)
  }
  
  
  # model again
  model <- glmmTMB((numberOfPlots) ~ -1+ neophyte + (1|family/genus), data= all, family= nbinom2() )
  
  summary(model)
  
  if (model$fit$convergence != 0) next
  
  # check pairwise differences
  emmeans_results <- emmeans::emmeans(model, pairwise ~ neophyte)
  # save in data frame
  sum <- summary(emmeans_results)$contrasts
  # add to large table
  result_freq <- rbind(result_freq, sum)
  
  successful_runs <- successful_runs+1
}

# look at aggregated results (mean and sd)
aggregated_result_freq <- aggregate(. ~ contrast, data = result_freq, FUN = function(x) c(mean = mean(x), sd = sd(x)))

# assess number of significant
result_freq$significant <- result_freq$p.value < 0.05
table(result_freq$significant)

result_freq |> group_by(contrast) |> summarise(n = sum(significant), n2 = sum(!significant))
```

```{r}
# Take a look at the most dominant species
topfrequent<- speciesDominance[order(speciesDominance$numberOfPlots, decreasing=T),]
topfrequent<- topfrequent[1:100,]
topdominant_intra<- topdominant[topdominant$neophyte=="intra",]
topdominant_extra <- topdominant[topdominant$neophyte=="extra",]
#write_csv(topfrequent, "../Results/speciesDominance_top100_frequent.csv")
```



## 3.2 Plot
```{r}
# Violin plot
# plot frequency
if(!native_intra_analysis){
  
p<-ggplot(speciesDominance, mapping = aes(x= (neophyte), y = log10(numberOfPlots), group= Neophyte, colour = Neophyte, fill= Neophyte)) + 
  geom_violin(alpha=0.5, scale="width")+
  geom_boxplot(width= 0.25, alpha=0.8, fill="white")+
  theme_pubr()+
  stat_summary(fun= "mean",
               geom = "point", aes(group= Neophyte), size=3)+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"))+
  scale_fill_manual(values = c("#1E88E5", "#FFC107", "#004D40"))+
  guides(color="none")+
  theme(legend.position = "none")+
  scale_x_discrete(labels=  c("extra-European \n aliens",  "intra-European \n aliens", "native species")) +
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=12), axis.title.x = element_blank(), 
        axis.text.y = element_text(size=10))+
  scale_y_continuous(limits=c(1.5, 5.3))+
  ylab(expression("log"[10]* " number of occurences")) + xlab(NULL)+
  annotate("text", x = 1, y = max(log10(speciesDominance$numberOfPlots)), 
             label = "b", size = 4,vjust = -0.5, hjust = 0.5, alpha=1) +
  annotate("text", x = 2, y = max(log10(speciesDominance$numberOfPlots)), 
             label = "c", size = 4, vjust = -0.5, hjust = 0.5, alpha=1)+
  annotate("text", x = 3, y = max(log10(speciesDominance$numberOfPlots)), 
             label = "a", size = 4, vjust = -0.5, hjust = 0.5, alpha=1)+
  annotate("text", x=1, y=max(log10(speciesDominance$numberOfPlots)), 
           label= paste("n=",length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="extra"])), 
           vjust = -2, hjust = 0.5, size=4)+
  annotate("text", x=2, max(log10(speciesDominance$numberOfPlots)), 
           label= paste("n=",length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="intra"])), 
           vjust = -2, hjust = 0.5, size=4)+
  annotate("text", x=3, max(log10(speciesDominance$numberOfPlots)), 
           label= paste("n=",length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="native"])), 
           vjust = -2, hjust = 0.5, size=4)+
  labs(subtitle = "Significance level: ***")+
  theme(plot.subtitle = element_text( hjust = 0.01))
Frequency <- p
p
ggsave("../Images/Frequency.jpeg", p,width = 5, height = 5)
} else {
p<- ggplot(speciesDominance, mapping = aes(x= neophyte, y = log10(numberOfPlots), colour = neophyte, group=neophyte, fill= neophyte)) + 
  geom_violin(alpha= 0.3, scale= "width")+
  geom_boxplot(width= 0.25, alpha=0.8, fill="white") +
  theme_pubr()+
  ylab(expression("log"[10]* " number of occurences")) + xlab(NULL)+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40", "darkgreen"), 
                    name="Legend",
                    breaks=c("native", "intra","extra", "native_intra"),
                    labels=c("native species", "intra European aliens", "extra European aliens", "native species alien elsewhere")) +
  scale_fill_manual(values=c( "#004D40","#FFC107", "#1E88E5", "darkgreen"), 
                    breaks=c("extra","intra", "native",  "native_intra")) +
  guides(color="none")+
  stat_summary(fun= "mean",
               geom = "point", aes(group= neophyte), size=3)+
  guides(color="none")+
  theme(legend.position = "none")+
  scale_x_discrete(labels= c("extra-European \n aliens",  "intra-European \n aliens", "native species",
                             "native species \nalien elsewhere")) +
  scale_y_continuous(limits=c(1.5, 5.5))+
  theme(axis.text.x = element_text(size=10), axis.title.y=element_text(size=12) )  +
  theme(plot.subtitle=element_text(size=12)) + 
  annotate("text", x = 1, y = max(log10(speciesDominance$numberOfPlots)), label = "b", size = 4, 
           vjust = -0.7, hjust = 0.5, alpha=0.8) +
  annotate("text", x = 2, y = max(log10(speciesDominance$numberOfPlots)), label = "c", size = 4, 
           vjust = -0.7, hjust = 0.5, alpha=0.8)+
  annotate("text", x = 3, y = max(log10(speciesDominance$numberOfPlots)), label = "b", size = 4, 
           vjust = -0.7, hjust = 0.5, alpha=0.8)+
  annotate("text", x = 4, y = max(log10(speciesDominance$numberOfPlots)), label = "a",size = 4, 
           vjust = -0.7, hjust = 0.5, alpha=0.8)+
  annotate("text", x=1, y=5.2, label= paste("n=",sum(speciesDominance$neophyte=="extra")), vjust = -2, hjust = 0.5, size=3.5)+
  annotate("text", x=2, y=5.2, label= paste("n=",sum(speciesDominance$neophyte=="intra")), vjust = -2, hjust = 0.5, size=3.5)+
  annotate("text", x=3, y=5.2, label= paste("n=",sum(speciesDominance$neophyte=="native")), vjust = -2, hjust = 0.5, size=3.5)+
  annotate("text", x=4, y=5.2, label= paste("n=",sum(speciesDominance$neophyte=="native_intra")), vjust = -2, hjust = 0.5, size=3.5)

Frequency <- p
p  

}


plot <- ggarrange(Dominance,Frequency, nrow=1, ncol=2, labels= c("a","b"),font.label = list(size = 12))
plot

#ggsave("Dominance_frequency.svg", plot= plot, width = 10, height = 5)

```





# 4 Analysis
## 4.1 Correlation
```{r}
cor.test(speciesDominance$coverMean, speciesDominance$numberOfPlots, method=c("spearman"))
cor.test(topdominant$coverMean, topdominant$numberOfPlots, method=c("spearman"))
cor.test(topfrequent$coverMean, topfrequent$numberOfPlots, method=c("spearman"))
```


```{r}
# Define relative positions for annotations
speciesDominance <- speciesDominance %>% mutate(Neophyte = factor(neophyte, 
                          levels = c("native", "intra","extra"),
                          labels = c("Native", 
                                     "Intra-European", 
                                     "Extra-European")))  


y_n_Dominance <- max(log10(speciesDominance$coverMean)) + 0.2 * diff(range(log10(speciesDominance$coverMean)))
y_text_Dominance <- max(log10(speciesDominance$coverMean)) + 0.1 * diff(range(log10(speciesDominance$coverMean)))

y_n_Frequency <- max(log10(speciesDominance$numberOfPlots)) + 0.2 * diff(range(log10(speciesDominance$numberOfPlots)))
y_text_Frequency <- max(log10(speciesDominance$numberOfPlots)) + 0.1 * diff(range(log10(speciesDominance$numberOfPlots)))

if(!native_intra_analysis){
# Violin plot for Dominance
Dominance <- ggplot(speciesDominance, aes(x=neophyte, y=log10(coverMean), group=Neophyte, colour=Neophyte, fill=Neophyte)) + 
  geom_violin(alpha=0.5, scale="width") +
  geom_boxplot(width=0.25, alpha=0.8, fill="white") +
  theme_pubr() +
  stat_summary(fun="mean", geom="point", aes(group=Neophyte), size=3) +
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40")) +
  scale_fill_manual(values=c("#1E88E5", "#FFC107", "#004D40")) +
  guides(color="none") +
  theme(legend.position="none") +
  scale_y_continuous(breaks=c(-1, 0, 1, 2), labels=c(0.1, 1, 10, 100), limits=c(-1, y_n_Dominance + 0)) +
  scale_x_discrete(labels=c("Extra-\nEuropean", "Intra-\nEuropean", "Native")) +
  theme(axis.text.x=element_text(size=12), axis.title.y=element_text(size=12), axis.title.x=element_blank(), 
        axis.text.y=element_text(size=10)) +
  
  theme(axis.title.y = element_text(margin = margin(r = 0)))+
  ylab("Dominance") +
  
  # Sample sizes (higher position)
  annotate("text", x=1, y=y_n_Dominance, label=paste("n=", length(speciesDominance$coverMean[speciesDominance$neophyte=="extra"])), size=4, hjust=0.5) +
  annotate("text", x=2, y=y_n_Dominance, label=paste("n=", length(speciesDominance$coverMean[speciesDominance$neophyte=="intra"])), size=4, hjust=0.5) +
  annotate("text", x=3, y=y_n_Dominance, label=paste("n=", length(speciesDominance$coverMean[speciesDominance$neophyte=="native"])), size=4, hjust=0.5) +
  
  # Significance letters (lower position)
  annotate("text", x=1, y=y_text_Dominance, label="a", size=4, hjust=0.5, alpha=1) +
  annotate("text", x=2, y=y_text_Dominance, label="b", size=4, hjust=0.5, alpha=1) +
  annotate("text", x=3, y=y_text_Dominance, label="c", size=4, hjust=0.5, alpha=1)

# Violin plot for Frequency
Frequency <- ggplot(speciesDominance, aes(x=neophyte, y=log10(numberOfPlots), group=Neophyte, colour=Neophyte, fill=Neophyte)) + 
  geom_violin(alpha=0.5, scale="width") +
  geom_boxplot(width=0.25, alpha=0.8, fill="white") +
  theme_pubr() +
  stat_summary(fun="mean", geom="point", aes(group=Neophyte), size=3) +
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40")) +
  scale_fill_manual(values=c("#1E88E5", "#FFC107", "#004D40")) +
  guides(color="none") +
  theme(legend.position="none") +
  scale_x_discrete(labels=c("Extra-\nEuropean", "Intra-\nEuropean", "Native")) +
  scale_y_continuous(breaks= seq(2, 5, by =1), limits=c(1.5, y_n_Frequency )) +
  ylab("Frequency")+
  xlab(element_blank())+
  theme(axis.title.y = element_text(margin = margin(r = 0)))+
  
  # Sample sizes (higher position)
  annotate("text", x=1, y=y_n_Frequency, label=paste("n=", length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="extra"])), size=4, hjust=0.5) +
  annotate("text", x=2, y=y_n_Frequency, label=paste("n=", length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="intra"])), size=4, hjust=0.5) +
  annotate("text", x=3, y=y_n_Frequency, label=paste("n=", length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="native"])), size=4, hjust=0.5) +
  
  # Significance letters (lower position)
  annotate("text", x=1, y=y_text_Frequency, label="b", size=4, hjust=0.5, alpha=1) +
  annotate("text", x=2, y=y_text_Frequency, label="c", size=4, hjust=0.5, alpha=1) +
  annotate("text", x=3, y=y_text_Frequency, label="a", size=4, hjust=0.5, alpha=1)
} else {
speciesDominance <- speciesDominance %>% mutate(Neophyte = factor(neophyte, 
                          levels = c("extra", "intra","native", "native_intra"),
                          labels = c("Extra-European",
                                     "Intra-European",
                                     "Native", 
                                     "native alien elsewhere")))    
  
  # Violin plot for Dominance
Dominance <- ggplot(speciesDominance, aes(x=neophyte, y=log10(coverMean), group=Neophyte, colour=Neophyte, fill=Neophyte)) + 
  geom_violin(alpha=0.5, scale="width") +
  geom_boxplot(width=0.25, alpha=0.8, fill="white") +
  theme_pubr() +
  stat_summary(fun="mean", geom="point", aes(group=Neophyte), size=3) +
  scale_colour_manual(values=c( "#004D40","#FFC107", "#1E88E5", "darkgreen")) +
  scale_fill_manual(values=c( "#004D40","#FFC107", "#1E88E5", "darkgreen")) +
  guides(color="none") +
  theme(legend.position="none") +
  scale_y_continuous(breaks=c(-1, 0, 1, 2), labels=c(0.1, 1, 10, 100), limits=c(-1, y_n_Dominance + 0)) +
  scale_x_discrete(labels=c("Extra-\nEuropean", "Intra-\nEuropean", "Native", "Native\nalien elsewhere")) +
  theme(axis.text.x=element_text(size=12), axis.title.y=element_text(size=12), axis.title.x=element_blank(), 
        axis.text.y=element_text(size=10)) +
  ylab("Dominance") +
  
  # Sample sizes (higher position)
  annotate("text", x=1, y=y_n_Dominance, label=paste("n=", length(speciesDominance$coverMean[speciesDominance$neophyte=="extra"])), size=4, hjust=0.5) +
  annotate("text", x=2, y=y_n_Dominance, label=paste("n=", length(speciesDominance$coverMean[speciesDominance$neophyte=="intra"])), size=4, hjust=0.5) +
  annotate("text", x=3, y=y_n_Dominance, label=paste("n=", length(speciesDominance$coverMean[speciesDominance$neophyte=="native"])), size=4, hjust=0.5)+ 
  annotate("text", x=4, y=y_n_Dominance, label=paste("n=", length(speciesDominance$coverMean[speciesDominance$neophyte=="native_intra"])), size=4, hjust=0.5) +
  # Significance letters (lower position)
  annotate("text", x=1, y=y_text_Dominance, label="a", size=4, hjust=0.5, alpha=1) +
  annotate("text", x=2, y=y_text_Dominance, label="b", size=4, hjust=0.5, alpha=1) +
  annotate("text", x=3, y=y_text_Dominance, label="c", size=4, hjust=0.5, alpha=1)+
  annotate("text", x=4, y=y_text_Dominance, label="c", size=4, hjust=0.5, alpha=1)

# Violin plot for Frequency
Frequency <- ggplot(speciesDominance, aes(x=neophyte, y=log10(numberOfPlots), group=Neophyte, colour=Neophyte, fill=Neophyte)) + 
  geom_violin(alpha=0.5, scale="width") +
  geom_boxplot(width=0.25, alpha=0.8, fill="white") +
  theme_pubr() +
  stat_summary(fun="mean", geom="point", aes(group=Neophyte), size=3) +
  scale_colour_manual(values=c( "#004D40","#FFC107", "#1E88E5", "darkgreen")) +
  scale_fill_manual(values=c( "#004D40","#FFC107", "#1E88E5", "darkgreen")) +
  guides(color="none") +
  theme(legend.position="none") +
  scale_x_discrete(labels=c("Extra-\nEuropean", "Intra-\nEuropean", "Native", "Native\nalien elsewhere")) +
  scale_y_continuous(breaks= seq(2, 5, by =1), limits=c(1.5, y_n_Frequency )) +
  ylab("Frequency")+
  xlab(element_blank())+
  
  # Sample sizes (higher position)
  annotate("text", x=1, y=y_n_Frequency, label=paste("n=", length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="extra"])), size=4, hjust=0.5) +
  annotate("text", x=2, y=y_n_Frequency, label=paste("n=", length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="intra"])), size=4, hjust=0.5) +
  annotate("text", x=3, y=y_n_Frequency, label=paste("n=", length(speciesDominance$numberOfPlots[speciesDominance$neophyte=="native"])), size=4, hjust=0.5) +  
  annotate("text", x=4, y=y_n_Frequency, label=paste("n=", length(speciesDominance$coverMean[speciesDominance$neophyte=="native_intra"])), size=4, hjust=0.5) +
  
  # Significance letters (lower position)
  annotate("text", x=1, y=y_text_Frequency, label="c", size=4, hjust=0.5, alpha=1) +
  annotate("text", x=2, y=y_text_Frequency, label="c", size=4, hjust=0.5, alpha=1) +
  annotate("text", x=3, y=y_text_Frequency, label="b", size=4, hjust=0.5, alpha=1)+
  annotate("text", x=4, y=y_text_Frequency, label="a", size=4, hjust=0.5, alpha=1)
}
Frequency
Dominance
# Arrange both plots in one figure
plot <- ggarrange(Dominance, Frequency, nrow=1, ncol=2, align="hv", labels=c("a","b"))
plot


ggsave("../Images/Dominance_frequency_native_intra.png", plot= plot, 
       width = 10, height = 5)

#EU <- EU + theme(plot.margin = margin(0, 0, 0, 0))
#Dominance <- Dominance + theme(plot.margin = margin(0, 0, 0, 0))
#Frequency <- Frequency + theme(plot.margin = margin(0, 0, 0, 0))

# plot together with files
plot <- ggarrange(EU,ggarrange(Dominance, Frequency, nrow=2, ncol=1, labels=c("b","c"),
                               font.label=list(size=12), align= "hv"), 
                  ncol=2, labels=c("a"), 
                  font.label=list(size=12), align="v", widths=  c(7,4), heights=c(6,6))
plot

ggsave("../Images/Dominance_frequency_map.png", plot= plot, 
       width = 12, height = 6)

ggsave("../Images/Dominance.png", plot= Dominance, width = 5, height = 3.5)
ggsave("../Images/Frequency.png", plot= Frequency, width = 5, height = 3.5)
eva[eva$name=="Rumex patientia aggr.",]

fullPlotData[fullPlotData$PlotObservationID %in% eva$PlotObservationID[eva$name=="Rumex patientia aggr."],]
(eva$name[eva$PlotObservationID %in% c(2195066)])
(eva$name[eva$PlotObservationID %in% c(2195067)])

```


## 4.2 Species
```{r}
rel_extra <- sum(!is.na(fullPlotData$cover_rel_extra))/ length(fullPlotData$PlotObservationID)
rel_intra <- sum(!is.na(fullPlotData$cover_rel_intra))/ length(fullPlotData$PlotObservationID) 

sort(plotsPerSpecies$numberOfPlots, decreasing=T)[1]/length(fullPlotData$PlotObservationID)
```
