==> 6_Base_Model.Rmd <==
---
title: "Base Model"
subtitle: "based on the work of Christian Ledergerber"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---
The purpose of this file is to explore which base model best explains the species richness.

# Start
We first clean the environment.
```{r, echo=T}
rm(list=ls())
```

Load packages
```{r, warnings=F, message=F}
library(readr)
library(AER)
library(mgcv)
library(modEvA)
library(lubridate)
library(visreg)
library(ggplot2)
library(parallel)
library(doParallel)
```

Load data
```{r, message=F}
# Data from header with the EIVE and DIV data
fullPlotData <- read_csv("fullPlotData.csv")
fullPlotData$Dataset <- as.factor(fullPlotData$Dataset)

fast <- F
if(fast) {
  fullPlotData <- fullPlotData[runif(length(fullPlotData$PlotObservationID)) > 0.99,]
}
```

# Explore correlations
Checking for correlation among predictor variables:
```{r}
# take normal DIV
correlation <- cor(fullPlotData[,c(51:55, 63:67)])
correlation[correlation < 0.7] <- 0
correlation
```

The following variables are all highly correlated: DistSeverity, DistSeverityHerblayer, SD_Grazing.Pressure, SD_Mowing.Frequency and SD_Soil.Disturbance. We therefore decided to only keep the overall disturbance severity and frequency in the model and remove the other correlated variables. DistFrequencyHerblayer and DistSeveretityHerblayer are only estimated for a subset of the habitats and hence should not be used for all habitats as we are doing. This is why we also removed this variable from the predictors.

Checking again, whether any of the predictors we plan to use are highly correlated shows that they are not:
```{r}
correlation <- cor(fullPlotData[,append(51:55, 77:78)])
correlation[correlation < 0.7] <- 0
correlation
```

# Base Model
## Straight line
Fitting a straight forward model to the data excluding the outliers:
```{r}
glm1 <- glm(numberOfVascularPlantSpecies ~ 
                    log(Area) + 
                    EIVEresM.sqrt + I(EIVEresM.sqrt^2) + 
                    EIVEresN.sqrt + I(EIVEresN.sqrt^2) + 
                    EIVEresR.sqrt + I(EIVEresR.sqrt^2) + 
                    EIVEresL.sqrt + I(EIVEresL.sqrt^2) + 
                    EIVEresT.sqrt + I(EIVEresT.sqrt^2) +
                     transformedDisturbanceSeverity + I(transformedDisturbanceSeverity^2) +
                    transformedDisturbanceFrequency + I(transformedDisturbanceFrequency^2)
                    , family=poisson, fullPlotData)

summary(glm1)
par(mfrow = c(2,2))
plot(glm1)
aicGlm1 <- AIC(glm1)
dsquaredGlm1 <- (glm1$null.deviance-glm1$deviance)/glm1$null.deviance
dsquaredGlm1

glm1.1 <- glm(numberOfVascularPlantSpecies ~ 
                    log(Area) + 
                    EIVEresM.sqrt + I(EIVEresM.sqrt^2) + 
                    EIVEresN.sqrt + I(EIVEresN.sqrt^2) + 
                    EIVEresR.sqrt + I(EIVEresR.sqrt^2) + 
                    EIVEresL.sqrt + I(EIVEresL.sqrt^2) + 
                    EIVEresT.sqrt + I(EIVEresT.sqrt^2) +
                    DistSeverity + I(DistSeverity^2) +
                    Soil.Disturbance + I(Soil.Disturbance^2) +
                    Grazing.Pressure + I(Grazing.Pressure^2)+
                    Mowing.Frequency + I(Mowing.Frequency^2)
                    , family=poisson, fullPlotData)

summary(glm1.1)
par(mfrow = c(2,2))
plot(glm1.1)
aicGlm1.1 <- AIC(glm1.1)
dsquaredGlm1.1 <- (glm1.1$null.deviance-glm1.1$deviance)/glm1.1$null.deviance
dsquaredGlm1.1
```
The residual plots look much better than before log transfroming the disturbance data and excluding the outliers. Hence, we test for overdispersion:

```{r}
dispersiontest(glm1)
```

The test for overdispersion is significant. Hence, we proceed with a quasipoisson model to correct for overdispersion.

```{r}
glm2 <- glm(numberOfVascularPlantSpecies ~ 
                    log(Area) + 
                    EIVEresM.sqrt + I(EIVEresM.sqrt^2) + 
                    EIVEresN.sqrt + I(EIVEresN.sqrt^2) + 
                    EIVEresR.sqrt + I(EIVEresR.sqrt^2) + 
                    EIVEresL.sqrt + I(EIVEresL.sqrt^2) + 
                    EIVEresT.sqrt + I(EIVEresT.sqrt^2) +
                    transformedDisturbanceSeverity + I(transformedDisturbanceSeverity^2) +
                    transformedDisturbanceFrequency + I(transformedDisturbanceFrequency^2)
                    , family=quasipoisson, fullPlotData)

summary(glm2)
par(mfrow = c(2,2))
aicGlm2 <- AIC(glm2)
plot(glm2)
dsquaredGlm2 <- (glm2$null.deviance-glm2$deviance)/glm2$null.deviance
dsquaredGlm2

glm1.2 <- glm(numberOfVascularPlantSpecies ~ 
                    log(Area) + 
                    EIVEresM.sqrt + I(EIVEresM.sqrt^2) + 
                    EIVEresN.sqrt + I(EIVEresN.sqrt^2) + 
                    EIVEresR.sqrt + I(EIVEresR.sqrt^2) + 
                    EIVEresL.sqrt + I(EIVEresL.sqrt^2) + 
                    EIVEresT.sqrt + I(EIVEresT.sqrt^2) +
                    DistSeverity + I(DistSeverity^2) +
                    Soil.Disturbance + I(Soil.Disturbance^2) +
                    Grazing.Pressure + I(Grazing.Pressure^2)+
                    Mowing.Frequency + I(Mowing.Frequency^2)
                    , family=quasipoisson, fullPlotData)
summary(glm1.2)
par(mfrow = c(2,2))
plot(glm1.2)
aicGlm1.2 <- AIC(glm1.2)
dsquaredGlm1.2 <- (glm1.2$null.deviance-glm1.2$deviance)/glm1.2$null.deviance
dsquaredGlm1.2

```

We next try fitting a GAM instead of a glm and treating Dataset as a random factor.
```{r}
gam0 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(transformedDisturbanceSeverity, bs = 'tp') +
              s(transformedDisturbanceFrequency, bs = 'tp'), 
            family = poisson(link = log), 
            data = fullPlotData, 
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)

AIC(glm1, gam0)
aicGam0 <- AIC(gam0)

dsquaredGam0 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam0), family = "poisson")
dsquaredGam0


gam0.1 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(DistFrequency, bs = 'tp'),
            family = poisson(link = log), 
            data = fullPlotData, 
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)


AIC(glm1, gam0, gam0.1)
gam0.1 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp'), 
            family = poisson(link = log), 
            data = fullPlotData, 
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)
aicGam0 <- AIC(gam0)

dsquaredGam0.1 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam0.1), family = "poisson")
dsquaredGam0.1
```

The AIC for the gam is `r aicGam0` while the AIC for the poisson link glm was `r aicGlm1`. This is a huge improvement. The models are comparable in complexity (poisson link, log size etc.) apart from replacing the use of quadratic indicator terms with smoothing terms. Hence, the GAM clearly outperforms the GLM.

We next add the Dataset as random factor and correct for spatial autocorrelation by adding a 2D smoothing term for Latitude / Longitude.
```{r}
begin<- Sys.time()
gam2 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(transformedDisturbanceSeverity.sqrt, bs = 'tp') +
              s(transformedDisturbanceFrequency.sqrt, bs = 'tp') +
              s(Dataset, bs = 're'), 
            family = poisson(link = log), 
            data = fullPlotData, 
            method = 'fREML',
            discrete=T,
            nthreads=4)
end<- Sys.time()
round(end-begin)

summary(gam2)
dsquaredGam2 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam2), family = "poisson")
dsquaredGam2

gam3 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(transformedDisturbanceSeverity.sqrt, bs = 'tp') +
              s(transformedDisturbanceFrequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're'), 
            family = poisson(link = log), 
            data = fullPlotData, 
            method = 'fREML', 
            discrete=TRUE,
            nthreads=4)

summary(gam3)

dsquaredGam3 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam3), family = "poisson")
dsquaredGam3

gam3.1 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're'), 
            family = poisson(link = log), 
            data = fullPlotData, 
            method = 'fREML', 
            discrete=TRUE,
            nthreads=4)

summary(gam3.1)

dsquaredGam3.1 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam3.1), family = "poisson")
dsquaredGam3.1
```

Finally, we compare the different weighting schemes for computing the average indicator values per plot:
```{r}
# Niche width weighing
gam4 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM.nw, bs = 'tp') +
              s(EIVEresN.nw, bs = 'tp') +
              s(EIVEresR.nw, bs = 'tp') +
              s(EIVEresL.nw, bs = 'tp') +
              s(EIVEresT.nw, bs = 'tp') +
              s(transformedDisturbanceSeverity, bs = 'tp') +
              s(transformedDisturbanceFrequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're'), 
            family = poisson(link = log), 
            data = fullPlotData, 
            method = 'fREML', 
            discrete=TRUE,
            nthreads=4)

summary(gam4)

dsquaredGam4 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam4), family = "poisson")
dsquaredGam4

# Cover over niche width
gam5 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM.cnw, bs = 'tp') +
              s(EIVEresN.cnw, bs = 'tp') +
              s(EIVEresR.cnw, bs = 'tp') +
              s(EIVEresL.cnw, bs = 'tp') +
              s(EIVEresT.cnw, bs = 'tp') +
              s(transformedDisturbanceSeverity.cnw, bs = 'tp') +
              s(transformedDisturbanceFrequency.cnw, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're'), 
            family = poisson(link = log), 
            data = fullPlotData,
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)

summary(gam5)

dsquaredGam5 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam5), family = "poisson")
dsquaredGam5

# No weighing
gam6 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(transformedDisturbanceSeverity, bs = 'tp') +
              s(transformedDisturbanceFrequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're'), 
            family = poisson(link = log), 
            data = fullPlotData,
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)

summary(gam6)

visreg(gam6, xvar = "Area", gg = TRUE) 
visreg(gam6, xvar = "EIVEresM", ylab = "Richness", xlab = "M", gg = TRUE, partial=FALSE, rug = FALSE) 
visreg(gam6, xvar = "EIVEresN", ylab = "Richness", xlab = "N", gg = TRUE, partial=FALSE, rug = FALSE) 
visreg(gam6, xvar = "EIVEresR", ylab = "Richness", xlab = "R", gg = TRUE, partial=FALSE, rug = FALSE) 
visreg(gam6, xvar = "EIVEresL", ylab = "Richness", xlab = "L", gg = TRUE, partial=FALSE, rug = FALSE) 
visreg(gam6, xvar = "EIVEresT", ylab = "Richness", xlab = "T", gg = TRUE, partial=FALSE, rug = FALSE) 
visreg(gam6, xvar = "transformedDisturbanceSeverity", ylab = "Richness", xlab = "Transformed Distrubance Severity", gg = TRUE, partial=FALSE, rug = FALSE) 
visreg(gam6, xvar = "transformedDisturbanceFrequency", ylab = "Richness", xlab = "Transformed Distrubance Frequency", gg = TRUE, partial=FALSE, rug = FALSE) 

dsquaredGam6 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam6), family = "poisson")
dsquaredGam6

par(mfrow = c(3,3))
plot(gam6)

# Try different disturbance
gam7 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(logDistSeverity, bs = 'tp') +
              s(logDistFrequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're'),
            family = poisson(link = log),
            data = fullPlotData,
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)

summary(gam7)

dsquaredGam7 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam7), family = "poisson")
dsquaredGam7

# Add Area as a polygone
gam8 <- bam(numberOfVascularPlantSpecies ~ 
              poly(log(Area),2) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(transformedDisturbanceSeverity, bs = 'tp') +
              s(transformedDisturbanceFrequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're'),
            family = poisson(link = log),
            data = fullPlotData,
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)

summary(gam8)

dsquaredGam8 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam8), family = "poisson")
dsquaredGam8
# Add Area as a smoother
gam9 <- bam(numberOfVascularPlantSpecies ~ 
              s(log(Area), bs = 'tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(transformedDisturbanceSeverity, bs = 'tp') +
              s(transformedDisturbanceFrequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're'),
            family = poisson(link = log),
            data = fullPlotData,
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)

summary(gam9)

dsquaredGam9 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam9), family = "poisson")
dsquaredGam9

par(mfrow = c(4,3))
plot(gam9)

AIC(gam0, gam2, gam3, gam4, gam5, gam6, gam7, gam8, gam9)
```

Conclusion: A model considering the log(size) of the plot, using the indicator as smoothers, taking different sampling cultures into account through Dataset and correcting for spatial autocorrelation using a two dimensional smoothing term on latitude / longitude seems to perform best when looking at AIC and Dsquared values.

Next we plot the residuals of the model in colour coded fashion on the europe map to visually assess the autocorrelation.
```{r}
plotTabel <- data.frame(predictedValues = predict(gam6), residuals = residuals(gam6), Longitude = fullPlotData$Longitude, Latitude = fullPlotData$Latitude)

ggplot(plotTabel, aes(Longitude, Latitude, col = residuals)) + 
  scale_colour_gradient2(name = "Residual species richness"  ) +
  geom_point(alpha = 1, stroke = 0.05, size = 0.005) + 
  coord_fixed(ratio = 1.4) + theme_dark()
```

Among the different weighting strategies used for computing the average indicator values per plot the strategy with no weighting outperforms all other strategies both in terms of d^2 as well as AIC.

Finally, explore, whether the different Datasets are significantly different using a fixed factor for Dataset:
```{r}
gam10 <- bam(numberOfVascularPlantSpecies ~ 
              log(Area) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(logDistSeverity, bs = 'tp') +
              s(logDistFrequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              Dataset,
            family = poisson(link = log),
            data = fullPlotData,
            method = 'fREML',
            discrete=TRUE,
            nthreads=4)

options(max.print=1000000)
summary(gam10)

dsquaredGam10 <- Dsquared(obs = fullPlotData$numberOfVascularPlantSpecies, pred = fitted(gam10), family = "poisson")
dsquaredGam10
```



Conclusion: not correcting for the skew in the data had caused a bad fit with completely unrealistic predictions and huge residuals. This has now been corrected. At the same time the model now no longer estimates a positive second order term for DistFrequency^2 but a negative one.
