==> 6_Base_Model_new_all_SR.Rmd <==
---
title: "Base model new"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---

The purpose of this file is to explore which base model best explains the species richness.

# 1 START
We first clean the environment.
```{r, echo=T}
rm(list=ls())
```

Load packages
```{r, warnings=F, message=F}
library(readr)
library(AER)
library(mgcv)
library(modEvA)
library(lubridate)
library(visreg)
library(ggplot2)
library(parallel)
library(doParallel)
library(IDPmisc)
library(ggeffects)
library(MASS)
```

## 1.1 Load
```{r, message=F}
# Data from header with the EIVE and DIV data
fullPlotData <- read_csv("fullPlotData_EUNIS_1980.csv")
fullPlotData$Dataset <- as.factor(fullPlotData$Dataset)

fullPlotData$Date <- as.Date(fullPlotData$Date, format = c("%Y-%m-%d"))
fullPlotData <- fullPlotData[fullPlotData$Date > "1980-01-01",]

fullPlotData<- fullPlotData[!is.na(fullPlotData$Date),]
fast <- F
if(fast) {
  fullPlotData <- fullPlotData[runif(length(fullPlotData$PlotObservationID)) > 0.99,]
}
```


## 1.2 Correlations
Checking for correlation among predictor variables:
```{r}
# take correlation omitting the INF values in the log disturbances
correlation <- cor(NaRV.omit(fullPlotData[,c(9:53)]))
# set all values to 0 which are not highly correlated
correlation[correlation < 0.7 & correlation> -0.7] <- 0
correlation
# see correlation--> correlation --> hard to see some things within dataset
# Interesting (but logical) --> EIVEresL correlated with disturbance frequency --> higher disturbance is species living more in sunlight
# Correlations between same indicator different calculations and weighting methods --> likely
```


The following variables are all highly correlated: DistFrequency and EIVEresL. We therefore decided to only keep the overall disturbance severity and frequency in the model and remove the other correlated variables. DistFrequencyHerblayer and DistSeveretityHerblayer are only estimated for a subset of the habitats and hence should not be used for all habitats as we are doing. This is why we also removed this variable from the predictors.

Check assumptions quasipoisson
```{r}
mean(fullPlotData$ENS0)
var(fullPlotData$ENS0)
```



# 2 MODEL
## 2.1 Distribution
We chose to include light instead of dist frequency
```{r}
model<- (ENS0 ~ log(Area) + 
                    EIVEresM.sqrt + I(EIVEresM.sqrt^2) + 
                    EIVEresN.sqrt + I(EIVEresN.sqrt^2) + 
                    EIVEresR.sqrt + I(EIVEresR.sqrt^2) + 
                    EIVEresL.sqrt + I(EIVEresL.sqrt^2) + 
                    EIVEresT.sqrt + I(EIVEresT.sqrt^2) +
                    DistSeverity + I(DistSeverity^2) +
                    Soil.Disturbance + I(Soil.Disturbance^2) +
                    Grazing.Pressure + I(Grazing.Pressure^2)+
                    Mowing.Frequency + I(Mowing.Frequency^2))
                   
# This is the simplest example possible, with normal possoin distribution
glm1 <- glm(model, family=poisson, fullPlotData)
dispersiontest(glm1)
# To account for overdispersion, we also test the quasipoisson distribution
glm2 <- glm(model, family= quasipoisson, fullPlotData)

summary(glm1)
summary(glm2)

# poisson
plot(glm1)
aicGlm1 <- AIC(glm1)
dsquaredGlm1 <- (glm1$null.deviance-glm1$deviance)/glm1$null.deviance
dsquaredGlm1

# semipoisson
plot(glm2)
dsquaredGlm2 <- (glm2$null.deviance-glm2$deviance)/glm2$null.deviance
dsquaredGlm2
```


## 2.2 Method
```{r}
model<- (ENS0 ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp'))
gam0 <- bam(model, family = poisson(link=log), data = fullPlotData,  method = 'fREML',discrete=TRUE, nthreads=4)
gam1 <- bam(model, family = quasipoisson, data = fullPlotData,  method = 'fREML',discrete=TRUE, nthreads=4)
gam1 <- bam(model, family = nb(), data = fullPlotData,  method = 'fREML',discrete=TRUE, nthreads=4)

summary(gam0)
summary(gam1)

# GAM has a considerably lower AIC
AIC(gam0, gam1)

dsquaredGam0 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam0), family = "poisson")
dsquaredGam0

dsquaredGam1 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam1), family = "poisson")
dsquaredGam1
```



# 3 FORMULA
## 3.1 Database
We add database (random effect) and longitude and latitude as smoothers
```{r}
model<- ENS0 ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model2<- ENS0 ~ 
              log(Area) + 
              s(EIVEresM.sqrt, bs = 'tp') +
              s(EIVEresN.sqrt, bs = 'tp') +
              s(EIVEresR.sqrt, bs = 'tp') +
              s(EIVEresL.sqrt, bs = 'tp') +
              s(EIVEresT.sqrt, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              Dataset
gam3 <- bam(model,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
#gam4 <- bam(model2,family = poisson(link = log), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)

summary(gam3)
#summary(gam4)

# GAM3 has a considerably higher Dsquared
dsquaredGam3 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam3), family = "poisson")
dsquaredGam3

# Incorporating the database considerably increases the Dsquared
# Making it a fixed factor does not increase much of the explained variation
# I do not run it here as it gives a very long output and produces NaNs
```

## 3.2 Weight EIVEs
```{r}
model1<- ENS0 ~ 
              log(Area) + 
              s(EIVEresM.nw, bs = 'tp') +
              s(EIVEresN.nw, bs = 'tp') +
              s(EIVEresR.nw, bs = 'tp') +
              s(EIVEresL.nw, bs = 'tp') +
              s(EIVEresT.nw, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')

model2<- ENS0 ~ 
              log(Area) + 
              s(EIVEresM.cnw, bs = 'tp') +
              s(EIVEresN.cnw, bs = 'tp') +
              s(EIVEresR.cnw, bs = 'tp') +
              s(EIVEresL.cnw, bs = 'tp') +
              s(EIVEresT.cnw, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model3<- ENS0 ~ 
              log(Area) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')

gam4 <- bam(model1,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam5 <- bam(model2,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam6 <- bam(model3,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)

# fREML is lowest for GAM with no weight
summary(gam3)
summary(gam4)
summary(gam5)
summary(gam6)

# AIC is lowest for GAM6 (in case of poisson, does not calculate for quasipoisson)
AIC(gam3,gam4,gam5,gam6)

# DSQUARED is lowest for GAM6
dsquaredGam4 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam4), family = "poisson")
dsquaredGam4
dsquaredGam5 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam5), family = "poisson")
dsquaredGam5
dsquaredGam6 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam6), family = "poisson")
dsquaredGam6
```


## 3.3 Weight DIVs
```{r}
model1<- ENS0 ~ 
              log(Area) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity, bs = 'tp') +
              s(Soil.Disturbance, bs = 'tp')+
              s(Grazing.Pressure, bs = 'tp')+
              s(Mowing.Frequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model2<- ENS0 ~ 
              log(Area) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model3<- ENS0 ~ 
              log(Area) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(logDistSeverity, bs = 'tp') +
              s(logSoil.Disturbance, bs = 'tp')+
              s(logGrazing.Pressure, bs = 'tp')+
              s(logMowingFrequency, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
gam7 <- bam(model1,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam8 <- bam(model2,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
#gam9 <- bam(model3,family = poisson(link = log), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)

# GAM 8 has a lower AIC
summary(gam7)
summary(gam8)

# GAM8 is the overall best model
AIC(gam7,gam8)

# Check gam 6 and 8 against each other
plot(gam6)
plot(gam8)

dsquaredGam4 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam7), family = "poisson")
dsquaredGam4
dsquaredGam5 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam8), family = "poisson")
dsquaredGam5
```


## 3.4 Area
```{r}
model<-  ENS0 ~ 
              poly(log(Area),2) + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model2<-  ENS0 ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model3<-  ENS0 ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')

gam9 <- bam(model,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam10 <- bam(model2,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam11 <- bam(model3,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)

# GAM 10 is the best now
summary(gam9)
summary(gam10)
summary(gam11)

AIC(gam11, gam9, gam10)


dsquaredGam4 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam9), family = "poisson")
dsquaredGam4
dsquaredGam5 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam10), family = "poisson")
dsquaredGam5
dsquaredGam6 <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(gam11), family = "poisson")
dsquaredGam6
```

## 3.5 HFP and Climate
```{r}
model2<-  ENS0 ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model3<-  ENS0 ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(hfp, bs="tp")+
              s(chelsaP, bs="tp")+
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model4<-  ENS0 ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(hfp, bs="tp")+
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model5<-  ENS0 ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(chelsaP, bs="tp")+
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
model6<-  ENS0 ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(hfp, bs="tp")+
              s(chelsaP, bs="tp")+
              s(elev, bs='tp')+
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')
gam12 <- bam(model2,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam13 <- bam(model3,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam14 <- bam(model4,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam15 <- bam(model5,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam16 <- bam(model6,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
summary(gam12)
summary(gam13)
summary(gam14)
summary(gam15)
summary(gam16)
AIC(gam1,gam3,gam4,gam5,gam6,gam7,gam8,gam9,gam10,gam11,gam12, gam13,gam14,gam15, gam16)

```


## 3.6 Final check
```{r}
gam10 <- bam(model6,family = poisson(link = log), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
gam11 <- bam(model6,family = nb(), data = fullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4)
AIC(gam10, gam11)
summary(gam10)
plot(gam10)
summary(gam11)
plot(gam11)
```


# 4 PLOT
```{r}
visreg(gam16, xvar = "Area", ylab = "Richness" ,gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "EIVEresM", ylab = "Richness", xlab = "M", gg = TRUE, partial=FALSE, rug = F,  scale="response") 
visreg(gam16, xvar = "EIVEresN", ylab = "Richness", xlab = "N", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "EIVEresR", ylab = "Richness", xlab = "R", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "EIVEresL", ylab = "Richness", xlab = "L", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "EIVEresT", ylab = "Richness", xlab = "T", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "DistSeverity.sqrt", ylab = "Richness", xlab = "Distrubance Severity (SQRT)", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "Soil.Disturbance.sqrt", ylab = "Richness", xlab = "Soil Disturbance (SQRT)", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "Grazing.Pressure.sqrt", ylab = "Richness", xlab = "Grazing Pressure (SQRT)", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "Mowing.Frequency.sqrt", ylab = "Richness", xlab = "Mowing Frequency (SQRT)", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "hfp", ylab = "Richness", xlab = "hfp", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "chelsaP", ylab = "Richness", xlab = "P mean", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
visreg(gam16, xvar = "elev", ylab = "Richness", xlab = "elev", gg = TRUE, partial=FALSE, rug = FALSE,  scale="response") 
```


```{r}
library(DHARMa)

MDL <- mgcv::bam(model6, family = nb(), data = fullPlotData,  
                 method = 'fREML',  discrete=T, nthreads=4)

# create simulation data
simulationOutput <- simulateResiduals(fittedModel = MDL, plot = F)
# look at simulation 
# QQ plot --> line normal (is it correct distribution and is there no overdispersion)
# residual plot --> if correct line at 0.5
plot(simulationOutput)

# test dispersion
testDispersion(simulationOutput)
plotResiduals(simulationOutput)
```
