==> 7_Impact_Classes.Rmd <==
---
title: "Impact analysis per class"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---

Based on the best model as determined in the baseModelExploration.qmd we assess the impact of individual neophytes in this script. We compute impact of all plants in EVA based on the response ratio between observed and expected species richness and divide this over multiple cover classes.

# 1 START
We first clean the environment.
```{r, echo=F, warning=F, message=FALSE}
rm(list=ls())
gc()
```

We load the required packages
```{r, echo=F, message=F, warning=F}
library(readr)
library(dplyr)
library(mgcv)
library(modEvA)
library(ggplot2)
library(tidyr)
library(foreach)
library(doParallel)
library(rstatix)
library(DescTools)
library(asbio)
```

## 1.1 Eva + header
We load the eva data with the species per plot
```{r}
eva <- read_csv("fullPlotEva_ESy.csv",show_col_types = FALSE)
```


Load filtered Data containing the richness and indicator values for each plot. In case we want to save computing time: down sample it:
```{r}
fullPlotData <- read_csv("fullPlotData_ESy.csv",show_col_types = FALSE)

fast <- T
if(fast) {
  fullPlotData <- fullPlotData[runif(length(fullPlotData$PlotObservationID)) > 0.90,]
  eva <- eva[eva$PlotObservationID %in% fullPlotData$PlotObservationID,]
}

#eva<- read.csv("downsampledEva.csv", header=T)
#colnames(eva)[8]<- "Cover %"
#fullPlotData<- read.csv("downsampledPlotData.csv")
```


## 1.2 Neophyte
Load the neophyte definitions:
```{r, message=F}
# Data on which species are neophytes
native_intra_analysis=F
if(native_intra_analysis){
  species_country_status<- read_csv("country_species_ESy.csv", show_col_types = FALSE)
} else{
  species_country_status<- read_csv("country_species_ESy.csv", show_col_types = FALSE)
  species_country_status$Neophyte[species_country_status$Neophyte=="native_intra"] <- "native"
# or if we want to do it with intra seperately
}

# Assign these names to the eva list
extra_EU <- unique(species_country_status$species[species_country_status$Neophyte=="extra"])
intra_EU <- unique(species_country_status$species[species_country_status$Neophyte=="intra"])
native_intra <- unique(species_country_status$species[species_country_status$Neophyte=="native_intra"])

```


Assign classification to species
```{r}
# Only observation ID and Region
fullPlot2<- fullPlotData[,c("PlotObservationID","Region")]
# Right join to keep only species present in fullplot (otherwise a lot of NAs)
eva<- right_join(eva, fullPlot2, by = c("PlotObservationID"="PlotObservationID"))
# Join eva and classification
eva_country_neophyte<- left_join(eva, species_country_status[, -c(2:4,6:8)], by= c("Region"= "Region", "name"= "name"))
# Look at how much from every type are present
table(eva_country_neophyte$Neophyte)
eva <- eva_country_neophyte
```



# 2 ANALYSIS
The idea is to compute the gam for every plot, which gives us the expected number of species based on the EIVE and DIV (and other indicators). We then compare this fitted (expected) values with the observed (real) species richness through the natural logarithm of the response ratio, which is defined as observed/ expected (commonly used in for example review papers on alien species).

Since Christians code is rather long, I try to simplify it by cutting it into more chunks.

## 2.1 Prepare
First, we make a list of the number of plots per species in our smaller dataset (if we do the analysis for all plots we can just make the file created earlier).
```{r, message=F}

if(fast) {
  reducedEva <- eva_country_neophyte |> select(PlotObservationID, Region,name, Neophyte,Layer,`Cover %`)

  reducedEva <- reducedEva |> group_by(PlotObservationID, Region,name, Neophyte, Layer) |> summarise(cover= sum(`Cover %`))
  
  reducedEva <- reducedEva|> group_by(PlotObservationID, Region, name, Neophyte) |> 
    summarise(cover = (1-prod((1-ifelse(cover/100>1, 1, cover/100 ))))*100)
  
  uniqueSpecies <- unique(reducedEva[, c(3,4)])
  
  # Empty dataset

  # Prepare parallel
  parallel::detectCores()
  n.cores <- parallel::detectCores() - 2
  my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
  print(my.cluster)
  doParallel::registerDoParallel(cl = my.cluster)
  foreach::getDoParRegistered()
  foreach::getDoParWorkers()
  
  
  plotsPerSpecies <- data.frame(names = character(), numberOfPlots = integer(), Neophyte= character())
  # Count in how many plots a native species occurs
  begin<- Sys.time()
plotsPerSpecies<-foreach(i= 1:nrow(uniqueSpecies), .combine='rbind', .packages=c("dplyr","mgcv")) %dopar% {
    tmp <- reducedEva$PlotObservationID[reducedEva$name == uniqueSpecies$name[i]& 
                                          reducedEva$Neophyte == uniqueSpecies$Neophyte[i]]
    names=uniqueSpecies$name[i]
    numberOfPlots= length(unique(tmp))
    plotsPerSpecies[i,]<- c(names, numberOfPlots,Neophyte=uniqueSpecies$Neophyte[i] )
}

  plotsPerSpecies<- as.data.frame(plotsPerSpecies)
  plotsPerSpecies<-setNames(plotsPerSpecies, c("names", "numberOfPlots","Neophyte"))
  plotsPerSpecies$numberOfPlots <- as.numeric(plotsPerSpecies$numberOfPlots)
  row.names(plotsPerSpecies)<- c(1: nrow(uniqueSpecies))
  
  end<- Sys.time()
  round(end-begin)
  
  parallel::stopCluster(cl = my.cluster)
  #write.csv(plotsPerSpecies,"plotsPerSpeciesInEurope_native_intra.csv", row.names = F )
} else {
  if(!native_intra_analysis){
  plotsPerSpecies <- read_csv("speciesDominance.csv", show_col_types = FALSE)
  plotsPerSpecies <- plotsPerSpecies[,c(1,5:6)]
  colnames(plotsPerSpecies) <- c("names","numberOfPlots","neophyte")
  } else {
  plotsPerSpecies <- read_csv("speciesDominance_native_intra.csv", show_col_types = FALSE)
  plotsPerSpecies <- plotsPerSpecies[,c(1,5:6)]
  colnames(plotsPerSpecies) <- c("names","numberOfPlots","neophyte")
  }
}
```


Get the names of the species that are present in more than the required minimum number of plots
```{r, message=F}
minimumNumberOfPlots<- 30
taxaForCoverClassAnalysis <- plotsPerSpecies[plotsPerSpecies$numberOfPlots >= minimumNumberOfPlots,]
```

Run this (eval=T) if you want to use the second method (in which we want at least x observations in the lowest and highest cover class)
```{r, message=F}
eval<-F
if (eval){
  # Prepare parallel
  parallel::detectCores()
  n.cores <- parallel::detectCores() - 2
  my.cluster <- parallel::makeCluster(
    n.cores, 
    type = "PSOCK"
  )
  print(my.cluster)
  doParallel::registerDoParallel(cl = my.cluster)
  foreach::getDoParRegistered()
  foreach::getDoParWorkers()
  
  # Get the names of the species which have at least 20 plots (since we want 10 smaller and 10 bigger)
  candidatesForCoverClassAnalysis <- plotsPerSpecies[plotsPerSpecies$numberOfPlots >= 20,]
  # create empty vector
  taxaForCoverAnalysis <- data.frame(names= character(), Neophyte= character())
  
  # run for all candidates
  begin<- Sys.time()
  taxaForCoverAnalysis<- foreach(i= 1:length(candidatesForCoverClassAnalysis$names),.combine='rbind',
                                 .packages=c("dplyr","mgcv")) %dopar% {
    # if the sum of all species with a cover of 50% for a species and less than 1% each are more than 10 
    # && makes sure to only look at the second if it is already T
    taxon<- candidatesForCoverClassAnalysis$names[i]
    if(sum(reducedEva$name == taxon & reducedEva$cover >= 70)>5 && 
       sum(reducedEva$name == taxon & reducedEva$cover < 1) > 5) {
      # write the species name to the list
      x<- c(taxon, candidatesForCoverClassAnalysis$neophyte[i] )
      x
    }
  }
  
  end<- Sys.time()
  round(end-begin)
  
  parallel::stopCluster(cl = my.cluster)
  
  taxaForCoverAnalysis<- as.data.frame(taxaForCoverAnalysis)
  taxaForCoverAnalysis<-setNames(taxaForCoverAnalysis, c("names", "Neophyte"))
  row.names(taxaForCoverAnalysis)<- c(1: nrow(taxaForCoverAnalysis))
}
```




## 2.2 Base model
We compute our base model with the best solution (see [rmarkdown link](6_Base_Model.Rmd)), which includes smoothers for all EIVE without weighting, smoothers for the DIV, a smoother for log(area), a smoother for spatial autocorrelation and the dataset as a random effect. We use a quasipoisson distribution with a log link function. 

Note that this base model can be adjusted in future works based on what we decide (eg human footprint map values).
```{r}
# Make Dataset a factor
fullPlotData$Dataset <- as.factor(fullPlotData$Dataset)
# GAM
model<-  ENS0 ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(chelsaP, bs='tp')+
              s(hfp, bs='tp')+
              s(elev, bs='tp')+
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')

baseModel <- bam(model,family = poisson, data = fullPlotData,  
                 method = 'fREML',  discrete=TRUE, nthreads=4)
dsquaredGam <- Dsquared(obs = fullPlotData$ENS0, pred = fitted(baseModel), family = "poisson")
dsquaredGam
```


## 2.3 Residuals
We use three methods, but only discuss the response ratio (RR)
```{r, message=F}
# Look at the relative residuals in the dataset
fullPlotData$relativRes <- (fullPlotData$numberOfVascularPlantSpecies - fitted(baseModel))/fitted(baseModel)
fullPlotData$logRelativRes <- (log(fullPlotData$numberOfVascularPlantSpecies) - log(fitted(baseModel)))/log(fitted(baseModel))
fullPlotData$RR<- (log(fullPlotData$ENS1/fitted(baseModel)))
summary(fullPlotData$relativRes)
summary(fullPlotData$logRelativRes)
summary(fullPlotData$RR)
```


## 2.4 Analysis
Function calculating impact and summarizing this per cover class. In the original work, we divided the cover classes in way larger classes, but following our new species selection based on a minimum number of occurrences this is not necessary
```{r}
# currently we are still assessing the data based on every layer, maybe it is useful to omit the alien species dataframe
# eva<- filter(eva, Layer == 0 | (Layer >=4 ) )

# Get Eva with plot, species and cover
reducedEva <- reducedEva_official[,c("PlotObservationID", "species", "Cover %", "Neophyte")]

# Get a vector with the name of the plant to analyse
plantName<- reducedEva$species[1]
status<- reducedEva$Neophyte[1]

# Function to calculate the impact for each species (stored in plantName)
computePlantImpact2 <- function(plantName, status) {
  
  #Create a temporary copy of fullPlotData
  tmpFullPlotData <- fullPlotData
  
  # Get the index of all observations of the species
  indexOfPlant <- reducedEva$species == plantName & reducedEva$Neophyte== status
  
  # The majority of the time is spent in computing the cover classes. 
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We group these observations by plot observation ID
  # We make a new variable calles coverClass for which we compute the mean cover per plot observation ID for the species under assessment.
  # We label them using breaks
  coverClassForPresent <- reducedEva[indexOfPlant,] |>
    group_by(PlotObservationID) |>
    summarise(coverClass = cut( `Cover %`, breaks = c(-1, 0, 1, 10, 30, 50, 70, 101), labels = 
                                 c("0%", "0%-1%", "1%-10%", "10%-30%", "30%-50%","50%-70%,","70%-100%"), right = TRUE) )
  
  # Because the cover class was only computed for plots containing the species we need to assign 0% class to the remaining plots.
  tmpFullPlotData <- left_join(tmpFullPlotData, coverClassForPresent, by = "PlotObservationID")
  tmpFullPlotData$coverClass[is.na(tmpFullPlotData$coverClass)] <- "0%"
  
  # Get the mean RR per cover class
  # We subtract the value for when the species is absent (as a baseline)
  result <- tmpFullPlotData |> group_by(coverClass) |> summarise(meanRel = mean(relativRes), meanLogRel = mean(logRelativRes), 
                                                                 meanRR= mean(RR))
  result$meanRel <- result$meanRel - result$meanRel[1]
  result$meanLogRel <- result$meanLogRel - result$meanLogRel[1]
  result$meanRR<- result$meanRR-result$meanRR[1]

  # Make empty dataframe to store the results. We want to have the species name, mean residuals per cover, total number of plots. The log defines the method applied
  res <- data.frame(taxa = character(), log = character(),  
                    absent = numeric(), present = numeric(), low = numeric(), medlow = numeric(), med= numeric(),medhigh=numeric(), 
                    dominant = numeric(), numberOfPlots = integer(), status= character())
  
  
  # Add the findings to the dataframe
  res <- add_row(res, taxa = plantName, log = "1",
                 absent = result$meanRel[1], present = result$meanRel[2], low = result$meanRel[3], medlow = result$meanRel[4], med = 
                 result$meanRel[5], medhigh = result$meanRel[6], dominant = result$meanRel[7],
                 numberOfPlots = sum(!(tmpFullPlotData$coverClass == "0%")), status= status)
  res <- add_row(res, taxa = plantName, log ="2",
                 absent = result$meanLogRel[1], present = result$meanLogRel[2], low = result$meanLogRel[3], medlow = result$meanLogRel[4], 
                 med = result$meanLogRel[5], medhigh = result$meanLogRel[6], dominant = result$meanLogRel[7],
                 numberOfPlots = sum(!(tmpFullPlotData$coverClass == "0%")), status= status)
  res <- add_row(res, taxa = plantName, log ="3",
                 absent = result$meanRR[1], present = result$meanRR[2], low = result$meanRR[3], medlow = result$meanRR[4], med = 
                 result$meanRR[5], medhigh = result$meanRR[6], dominant = result$meanRR[7],
                 numberOfPlots = sum(!(tmpFullPlotData$coverClass == "0%")), status= status)
  res
}
```




## 2.5 Run all
```{r, message=F}
if(fast) {
# Get the impact
# Make new dataframe
# If you want to perform the analysis following the second method (with at least 5 observations in both the lowest and the highest class), run the following line
#taxaForCoverClassAnalysis<- taxaForCoverAnalysis

# All species which are deemed good to analyse based on criteria
start <- 1
end<- length(taxaForCoverClassAnalysis$names)

# Prepare parallel
parallel::detectCores()
n.cores <- parallel::detectCores() - 4
my.cluster <- parallel::makeCluster(
  n.cores, 
  type = "PSOCK"
)
print(my.cluster)
doParallel::registerDoParallel(cl = my.cluster)
foreach::getDoParRegistered()
foreach::getDoParWorkers()



begin<- Sys.time()
# Run for all species
impact <- foreach(i=start:end, .combine='rbind', .packages=c("dplyr","mgcv")) %dopar% {
 res<-  computePlantImpact2(taxaForCoverClassAnalysis$names[i], taxaForCoverClassAnalysis$Neophyte[i])
}  


end<- Sys.time()
round(end-begin)

parallel::stopCluster(cl = my.cluster)

# Copy in new data frame
coverClassImpactForCandidates <- impact

#fileName <- paste("coverClassImpactForCandidates_new_all.csv")
#write_csv(coverClassImpactForCandidates, fileName)
}
```

There are some problems with running foreach on the whole database, hence we run seperate for loop if we want to compute the whole dataset (takes about 1.5 hours)
```{r}
if(!fast){
# Get the impact
# Make new dataframe
coverClassImpactForCandidates <- data.frame(taxa = character(), log = character(),  
                    absent = numeric(), present = numeric(), low = numeric(), medlow = numeric(), med= numeric(),medhigh=numeric(), 
                    dominant = numeric(), numberOfPlots = integer(), status= character())

# All species which are deemed good to analyse based on criteria
start <- 1
minimumNumberOfPlots<- 50
taxaForCoverClassAnalysis <- plotsPerSpecies[plotsPerSpecies$numberOfPlots >= minimumNumberOfPlots,]
end<- length(taxaForCoverClassAnalysis$names)

#if(!is.na(params$start) && params$start != "NA") { start <- params$start}
#if(!is.na(params$end) && params$end != "NA") { end <- params$end}

begin<- Sys.time()
# Run for all species
for (i in start:end) {
    # Take previously made function and store temporarily in res
    res <- computePlantImpact2(taxaForCoverClassAnalysis$names[i], taxaForCoverClassAnalysis$neophyte[i])
    # Enlarge the final dataframe
    coverClassImpactForCandidates <- rbind(coverClassImpactForCandidates, res)
}

coverClassImpactForCandidates

end<- Sys.time()
round(end-begin)

# Check for loop
#taxaForCoverClassAnalysis[!((taxaForCoverClassAnalysis[(taxaForCoverClassAnalysis %in% eva$species[eva$Neophyte== "native"])]) %in% (unique(coverClassImpactForCandidates$taxa[coverClassImpactForCandidates$status=="native"])))]

#fileName <- paste("coverClassImpactForCandidates_cover_all_layer_cleaned.csv")
#write_csv(coverClassImpactForCandidates, fileName)
}
```


## 2.6 Synthesis
```{r, warning=F, message=F}
coverClassImpactForCandidates <- read.csv("coverClassImpactForCandidates_cover_all_layer_cleaned.csv")
# Select all columns from the dataset except number of plots and pivot this data so that we have two new columns with the cover class name as a factor and their values
dataToPlot <- coverClassImpactForCandidates |> 
  select(-numberOfPlots) |> 
  pivot_longer(cols =c(`absent`,`present`,`low`,`medlow`,`med`,`medhigh`,`dominant`),names_to = "cover_class", values_to ="diversity_impact") |>
  mutate(cover_class = factor(cover_class, levels = c("absent", "present", "low","medlow", "med", "medhigh","dominant")))

dataToPlot<- dataToPlot[dataToPlot$log=="3",]
dataToPlot<- dataToPlot[!is.na(dataToPlot$diversity_impact),]

# Summarize the data per classification, cover class and method and calculate the mean, standard deviation and confidence interval
if(fast){
summarizedData <- dataToPlot |>  group_by(status, cover_class)|> 
  summarise(n=n(), mean= mean(diversity_impact, na.rm=T),sd = sd(diversity_impact, na.rm=T), ci= qt(p=1-0.05/2, df=n-1)*sd/sqrt(n)) 
} else {
  summarizedData <- dataToPlot |>  group_by(status, cover_class)|> 
  summarise(n=n(), mean= mean(diversity_impact, na.rm=T),sd = sd(diversity_impact, na.rm=T), ci= qt(p=1-0.05/2, df=n-1)*sd/sqrt(n), median=ci.median(diversity_impact, conf = 0.95)$ci[1], ci.lower= ci.median(diversity_impact,conf = 0.95)$ci[2], ci.higher= ci.median(diversity_impact, conf = 0.95)$ci[3])
}
# I also wanted to add the median (as it might be interesting to see how this relates to the mean value --> are there some 'outlier' plants that significantly change our observations) but it is not always possible to run this due to a lack of observations 
```


## 2.7 Significance
Compare the different classes using ranked ANOVA test and Tukey post-hoc test:
```{r}
# ranked ANOVA test
# Interaction included here
result<-(aov(rank(diversity_impact) ~ status * cover_class, dataToPlot))
summary(result)
# Tukey post hoc test
TukeyHSD(result, which= "cover_class") 
TukeyHSD(result, which= "status")
plot(TukeyHSD(result, conf.level=.95), las = 2)

# Better to use Welch ANOVA
# Welch Anova
result<- oneway.test(diversity_impact ~ status, dataToPlot, var.equal = F)
result
# Games howell post hoc test (Tukey for uneven variances)
games_howell_test(data=dataToPlot, diversity_impact ~ status, conf.level = 0.95, detailed=F)
games_howell_test(data=dataToPlot, diversity_impact ~ cover_class, conf.level = 0.95)
```



## 2.8 Save
```{r}

#fileName <- paste("coverClassImpactForCandidates_new_all_100%.csv")
#write_csv(coverClassImpactForCandidates, fileName)

```



# 3 PLOT
## 3.1 SummarizedData
```{r}
# Remove the absent class
summarizedData<- summarizedData[!summarizedData$cover_class=="absent",]

if(!native_intra_analysis){
p<- ggplot(summarizedData, aes(x=cover_class, y=mean, group=status, color=status)) + 
  geom_line(linewidth=1, position=position_dodge(0.2)) +
  ggtitle("Mean impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_point(position=position_dodge(0.2))+ylab("ln(OBS/EXP)") + 
  geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci), width=.2, position=position_dodge(0.2), linewidth=1)+
  xlab("Cover Class") +
  scale_x_discrete(labels = c('[0%-1%]','[1% - 10%]', '[10%-30%]','[30%-50%]', '[50%-70%]','[70%-100%]'))+
  scale_y_continuous(limits = c(-0.7, 0.30))+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"), 
                    name="Legend",
                    breaks=c("native", "intra","extra"),
                    labels=c("native species", "intra European neophyte", "extra European neophyte"))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))
  
p

#ggsave("RR_Cover_Class.jpeg", p, width = 25, height = 15, units = "cm")

# It should be noted that we can convert RR to percentage (see Mollot 2017 who also did it), but we cannot convert the confidence interval so it is not possible to make a graph out of this


# RR but as barplot
p<- ggplot(summarizedData, aes(x=cover_class, y=mean,   fill=status)) + 
  geom_bar(stat="identity",position = position_dodge(0.7), width = 0.7) +
  ggtitle("Mean impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_point(position=position_dodge(0.7))+
  geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci), width=.2, position=position_dodge(0.7)) +
  ylab("ln(OBS/EXP)") + 
  xlab("Cover Class") +
    scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"), 
                    name="Legend",
                    breaks=c("native", "intra","extra"),
                    labels=c("native species", "intra European neophyte", "extra European neophyte"))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))
p

#ggsave("RR_Cover_Class_Barplot.jpeg", p,width = 25, height = 15, units = "cm")

# I also wanted to visualise the median (to omit some species that strongly bias the outcome)
if(!fast){
p<- ggplot(summarizedData, aes(x=cover_class, y=median, group=status, color=status)) + 
  geom_line(linewidth=1, position=position_dodge(0.2)) +
  ggtitle("Median impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_point(position=position_dodge(0.2))+ylab("ln(OBS/EXP)") + 
  geom_errorbar(aes(ymin=ci.lower, ymax=ci.higher), width=.2, position=position_dodge(0.2), linewidth=1)+
  xlab("Cover Class") +
  scale_x_discrete(labels = c('[0%-1%]','[1% - 10%]', '[10%-30%]','[30%-50%]', '[50%-70%]','[70%-100%]'))+
  scale_y_continuous(limits = c(-0.7, 0.30))+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"), 
                    name="Legend",
                    breaks=c("native", "intra","extra"),
                    labels=c("native species", "intra European neophyte", "extra European neophyte"))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))
p
#ggsave("RR_Cover_Class_median.jpeg", p, width = 25, height = 15, units = "cm")
}
} else {
p<- ggplot(summarizedData, aes(x=cover_class, y=mean, group=status, color=status)) + 
  geom_line(linewidth=1, position=position_dodge(0.2)) +
  ggtitle("Mean impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_point(position=position_dodge(0.2))+ylab("ln(OBS/EXP)") + 
  geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci), width=.2, position=position_dodge(0.2), linewidth=1)+
  xlab("Cover Class") +
  scale_x_discrete(labels = c('[0%-1%]','[1% - 10%]', '[10%-30%]','[30%-50%]', '[50%-70%]','[70%-100%]'))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))+
  scale_y_continuous(limits = c(-0.9, 0.30))+
  scale_colour_manual(values=c("#1E88E5","#B71C1C", "#FFC107", "#388E3C"), 
                    name="Legend",
                    breaks=c("native","native_intra", "intra","extra"),
                    labels=c("native species","native species alien elsewhere", "intra European neophyte", "extra European neophyte"))
p

#ggsave("RR_Cover_Class.jpeg", p, width = 25, height = 15, units = "cm")

# It should be noted that we can convert RR to percentage (see Mollot 2017 who also did it), but we cannot convert the confidence interval so it is not possible to make a graph out of this


# RR but as barplot
p<- ggplot(summarizedData, aes(x=cover_class, y=mean,   fill=status)) + 
  geom_bar(stat="identity",position = position_dodge(0.7), width = 0.7) +
  ggtitle("Mean impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_point(position=position_dodge(0.7))+
  geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci), width=.2, position=position_dodge(0.7)) +
  ylab("ln(OBS/EXP)") + 
  xlab("Cover Class") +
  scale_colour_manual(values=c("#1E88E5","#B71C1C", "#FFC107", "#388E3C"), 
                    name="Legend",
                    breaks=c("native","native_intra", "intra","extra"),
                    labels=c("native species","native species alien elsewhere", "intra European neophyte", "extra European neophyte"))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))
p

#ggsave("RR_Cover_Class_Barplot.jpeg", p,width = 25, height = 15, units = "cm")

# I also wanted to visualise the median (to omit some species that strongly bias the outcome)
if(!fast){
p<- ggplot(summarizedData, aes(x=cover_class, y=median, group=status, color=status)) + 
  geom_line(linewidth=1, position=position_dodge(0.2)) +
  ggtitle("Median impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_point(position=position_dodge(0.2))+ylab("ln(OBS/EXP)") + 
  geom_errorbar(aes(ymin=ci.lower, ymax=ci.higher), width=.2, position=position_dodge(0.2), linewidth=1)+
  xlab("Cover Class") +
  scale_x_discrete(labels = c('[0%-1%]','[1% - 10%]', '[10%-30%]','[30%-50%]', '[50%-70%]','[70%-100%]'))+
  scale_y_continuous(limits = c(-0.35, 0.30))+
  scale_colour_manual(values=c("#1E88E5","#B71C1C", "#FFC107", "#388E3C"), 
                    name="Legend",
                    breaks=c("native","native_intra", "intra","extra"),
                    labels=c("native species","native species alien elsewhere", "intra European neophyte", "extra European neophyte"))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))
p
#ggsave("RR_Cover_Class_median.jpeg", p, width = 25, height = 15, units = "cm")
}
}
```


## 3.2 Boxplot
```{r}
dataToPlot<- dataToPlot[!dataToPlot$cover_class=="absent",]
if(!native_intra_analysis){
p<- ggplot(dataToPlot[!is.na(dataToPlot$diversity_impact),], aes(x=cover_class, y=diversity_impact, color=status)) + 
  ggtitle("Mean impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_boxplot(position = position_dodge(0.7), width = 0.7)+
  ylab("ln(OBS/EXP)") + 
  xlab("Cover Class") +
  scale_x_discrete(labels = c('[0%-1%]','[1% - 10%]', '[10%-30%]','[30%-50%]', '[50%-70%]','[70%-100%]'))+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"), 
                    name="Legend",
                    breaks=c("native", "intra","extra"),
                    labels=c("native species", "intra European neophyte", "extra European neophyte"))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))
p
#ggsave("RR_Cover_Class_Boxplot.jpeg", p, width = 25, height = 15, units = "cm")
} else {
p<- ggplot(dataToPlot[!is.na(dataToPlot$diversity_impact),], aes(x=cover_class, y=diversity_impact, color=status)) + 
  ggtitle("Mean impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_boxplot(position = position_dodge(0.7), width = 0.7)+
  ylab("ln(OBS/EXP)") + 
  xlab("Cover Class") +
  scale_x_discrete(labels = c('[0%-1%]','[1% - 10%]', '[10%-30%]','[30%-50%]', '[50%-70%]','[70%-100%]'))+
  scale_colour_manual(values=c("#1E88E5","#B71C1C", "#FFC107", "#388E3C"), 
                    name="Legend",
                    breaks=c("native","native_intra", "intra","extra"),
                    labels=c("native species","native species alien elsewhere", "intra European neophyte", "extra European neophyte"))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))
p
#ggsave("RR_Cover_Class_Boxplot.jpeg", p, width = 25, height = 15, units = "cm")
  
}

```


# 4 FURTHER
## 4.1 Dominant
Which species are most dominant
```{r}
dominant <- coverClassImpactForCandidates[order(coverClassImpactForCandidates$dominant),]
# Take top 100
dominant<- dominant[dominant$log=="3",]
dominant <- dominant[1:100,]
print(dominant[1:20,])
# Count species in the top
sum(dominant$status=="intra")
sum(dominant$status=="extra")
sum(dominant$status=="native")
dominant$taxa[dominant$status=="intra"]
```


## 4.2 Check average
```{r}
# Mean impact for all species per status
mean(dataToPlot$diversity_impact[dataToPlot$log=="3"& dataToPlot$status=="intra"], na.rm=T)
mean(dataToPlot$diversity_impact[dataToPlot$log=="3"& dataToPlot$status=="native"], na.rm=T)
mean(dataToPlot$diversity_impact[dataToPlot$log=="3"& dataToPlot$status=="extra"], na.rm=T)
mean(dataToPlot$diversity_impact[dataToPlot$log=="3"& dataToPlot$status=="native_intra"], na.rm=T)

# Mean impact across all classes
mean(dataToPlot$diversity_impact[dataToPlot$log=="3"], na.rm=T)

# Possible reasons why we have a higher impact than assumed purely based on mathematical reasons 
      # we only assess species present in more than 50 plots

# Mean number of species per plot
mean(fullPlotData$numberOfVascularPlantSpecies)
```

## 4.3 Invasives
```{r}
gisd <- read.csv("../EIVE Data/GISD.csv", sep = ";")
coverClassImpactForCandidates$invasive <- coverClassImpactForCandidates$taxa %in% gisd$Species
coverClassImpactForCandidates$status[coverClassImpactForCandidates$invasive==T & coverClassImpactForCandidates$status== "native"] <- "native_invasive"
coverClassImpactForCandidates$status[coverClassImpactForCandidates$invasive==T & coverClassImpactForCandidates$status== "intra"] <- "intra_invasive"
coverClassImpactForCandidates$status[coverClassImpactForCandidates$invasive==T & coverClassImpactForCandidates$status== "extra"] <- "extra_invasive"


dataToPlot <- coverClassImpactForCandidates |> 
  select(-numberOfPlots) |> 
  pivot_longer(cols =c(`absent`,`present`,`low`,`medlow`,`med`,`medhigh`,`dominant`),names_to = "cover_class", values_to ="diversity_impact") |>
  mutate(cover_class = factor(cover_class, levels = c("absent", "present", "low","medlow", "med", "medhigh","dominant")))

dataToPlot<- dataToPlot[dataToPlot$log=="3",]
dataToPlot<- dataToPlot[!is.na(dataToPlot$diversity_impact),]

# Summarize the data per classification, cover class and method and calculate the mean, standard deviation and confidence interval
if(fast){
summarizedData <- dataToPlot |>  group_by(status, cover_class)|> 
  summarise(n=n(), mean= mean(diversity_impact, na.rm=T),sd = sd(diversity_impact, na.rm=T), ci= qt(p=1-0.05/2, df=n-1)*sd/sqrt(n)) 
} else {
  summarizedData <- dataToPlot |>  group_by(status, cover_class)|> 
  summarise(n=n(), mean= mean(diversity_impact, na.rm=T),sd = sd(diversity_impact, na.rm=T), ci= qt(p=1-0.05/2, df=n-1)*sd/sqrt(n), median=ci.median(diversity_impact, conf = 0.95)$ci[1], ci.lower= ci.median(diversity_impact,conf = 0.95)$ci[2], ci.higher= ci.median(diversity_impact, conf = 0.95)$ci[3])
}
```

```{r}
summarizedData <- summarizedData[!summarizedData$cover_class=="absent",]
p<- ggplot(summarizedData, aes(x=cover_class, y=mean, group=status, color=status)) + 
  geom_line(linewidth=1, position=position_dodge(0.2)) +
  ggtitle("Mean impact of Plant Species on Species Richness")+
  theme(plot.title = element_text(hjust= 0.5, size=22))+
  geom_point(position=position_dodge(0.2))+ylab("ln(OBS/EXP)") + 
  geom_errorbar(aes(ymin=mean-ci, ymax=mean+ci), width=.2, position=position_dodge(0.2), linewidth=1)+
  xlab("Cover Class") +
  scale_x_discrete(labels = c('[0%-1%]','[1% - 10%]', '[10%-30%]','[30%-50%]', '[50%-70%]','[70%-100%]'))+
  scale_y_continuous(limits = c(-0.7, 0.30))+
  scale_colour_manual(values = c("#1E88E5", "#FFC107", "#004D40", "#FF5733", "#3C33FF", "#00FF33"), 
                    name="Legend",
                    breaks=c("native", "intra","extra", "native_invasive", "intra_invasive","extra_invasive"),
                    labels=c("native species", "intra European neophyte", "extra European neophyte",
                             "native_invasive", "intra_invasive","extra_invasive"))+
  theme_classic()+ geom_hline(yintercept = 0, linetype="longdash")+ 
  theme(axis.text.x = element_text(size=12), axis.title.y=element_text(size=14) ,axis.title.x=element_text(size=14))+
  theme(legend.text= element_text(size=12), legend.title =element_text(size=14))
  
p
```


## 4.4 Rel Species
```{r}
test <- dataToPlot |> group_by(cover_class, status)|> summarise(n=n())
test<- test |>group_by( status) |> mutate(rel= round(n/max(n), 3))

p<- ggplot(test, aes(x=cover_class, y= rel, color=status, group=status))+
  scale_x_discrete(labels = c('[0%]','[0%-1%]','[1% - 10%]', '[10%-30%]','[30%-50%]', '[50%-70%]','[70%-100%]'))+
         geom_bar(stat = "identity", position = position_dodge(0.7), width = 0.7)
#ggsave("Cover class relative distribution.jpeg", p, width = 25, height = 15, units = "cm")
```


## 4.5 Rel Observations
```{r}
# Get Eva with plot, species and cover
reducedEva <- eva[,c("PlotObservationID", "species", "Cover %", "Neophyte")]

# Function to calculate the number of observations per class for each species (stored in plantName)
computePlantImpact3 <- function(plantName, status) {
  #Create a temporary copy of fullPlotData
  tmpFullPlotData <- fullPlotData
  
  # Get the index of all observations of the species
  indexOfPlant <- reducedEva$species == plantName & reducedEva$Neophyte== status
  
  # The majority of the time is spent in computing the cover classes. 
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We group these observations by plot observation ID
  # We make a new variable calles coverClass for which we compute the mean cover per plot observation ID for the species under assessment.
  # We label them using breaks
  coverClassForPresent <- reducedEva[indexOfPlant,] |>
    group_by(PlotObservationID) |>
    summarise(coverClass = cut( `Cover %`, breaks = c(-1, 0, 1, 10, 30, 50, 70, 101), labels = 
                                 c("0%", "0%-1%", "1%-10%", "10%-30%", "30%-50%","50%-70%,","70%-100%"), right = TRUE) )
  
  # Because the cover class was only computed for plots containing the species we need to assign 0% class to the remaining plots.
  tmpFullPlotData <- left_join(tmpFullPlotData, coverClassForPresent, by = "PlotObservationID")
  tmpFullPlotData$coverClass[is.na(tmpFullPlotData$coverClass)] <- "0%"
  
  # summarise and count number of observations per class
  test <- tmpFullPlotData |> group_by(coverClass) |> summarise(n=n())
  # give also plant species and name
  test<- cbind(c(plantName), c(status), test)
  # for ease change colnames
  colnames(test)<- c("species","status","coverClass", "n")
  test
}

# All species which are deemed good to analyse based on criteria previously
start <- 1
end<- length(taxaForCoverClassAnalysis$names)

# empty dataframe
impact<- data.frame(species=character(), status=character(), coverClass=character(), n=numeric())
# foreach does not work for all species so we run with normal for-loop
for(i in start: end){
  res<-  computePlantImpact3(taxaForCoverClassAnalysis$names[i], taxaForCoverClassAnalysis$neophyte[i])
  impact<- rbind(impact, res)
}

# we save the results
#write_csv(impact, "Method2_number_of_observations_per_class.csv")
```


```{r}
# remove all absent observations (too much, make plot less visible)
test<- impact[!impact$coverClass=="0%",]

# plot the number of observations for each species with n
# if you want to run the total number run test below and change n to sum
p <- ggplot(test, aes(coverClass, n, fill = status)) +
  geom_bar(stat = "identity", position = position_dodge(0.7), width = 0.7) + 
  geom_point(position = position_dodge(0.7)) +
  theme(plot.title = element_text(hjust = 0.5, size = 22)) +
  ylab("Number of observations") + 
  xlab("Cover Class") +
  scale_fill_manual(
    values = c("#1E88E5", "#FFC107", "#004D40"), 
    name = "Legend",
    breaks = c("native", "intra", "extra"),
    labels = c("native species", "intra European neophyte", "extra European neophyte")
  ) +
  theme_classic() + 
  geom_hline(yintercept = 0, linetype = "longdash") + 
  theme(
    axis.text.x = element_text(size = 12),
    axis.title.y = element_text(size = 14),
    axis.title.x = element_text(size = 14),
    legend.text = element_text(size = 12),
    legend.title = element_text(size = 14)
  ) +
  facet_wrap(~ status, scales = "free_y", ncol = 1)
p

test <- test|> group_by(status, coverClass) |> summarise(sum= sum(n))


#ggsave("Number of observations per species.jpeg", p, width = 25, height = 15, units = "cm")

```
