==> 8_Impact_Base_Model.Rmd <==
---
title: "Impact analysis with base model"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---


```{r, echo=FALSE}
rm(list=ls())
```


Load packages
```{r, warnings=F, message=F}
library(readr)
library(dplyr)
library(mgcv)
library(modEvA)
library(ggplot2)
library(doParallel)
library(foreach)
```


# Start
Here we set some parameters that are useful in the script
```{r}
correctArea<- F
```


Load eva and plot data
```{r, warnings=F, message=F}
eva <- read_csv("fullPlotEva.csv")
fullPlotData <- read_csv("fullPlotData.csv")
```


Downsample by a factor of 100 if fast is selected
```{r}
fast <- T
if(fast) {
  fullPlotData <- fullPlotData[runif(length(fullPlotData$PlotObservationID)) > 0.90,]
  eva <- eva[eva$PlotObservationID %in% fullPlotData$PlotObservationID,]
}
```


Load neophyte definitions and check which of the data exists in EVA
```{r, warnings=F, message=F}
neophyteDefinitions <- read_csv("../Neophyte Assignments/UniqueTaxaEurope-2023-04-23.csv")
neophyteNames <- neophyteDefinitions$species[neophyteDefinitions$statusEurope == "neo"]
neophyteNames <- neophyteNames[neophyteNames %in% eva$species]
```


Reduce EVA to a smaller dataset to accelerate computations.  
Count the number of plots in which the species in the dataset exist.
```{r}
if(fast) {
  reducedEva <- eva |> select(PlotObservationID, species, `Cover %`)
  
  uniqueSpecies <- unique(reducedEva$species)
  
  plotsPerSpecies <- data.frame(names = character(), numberOfPlots = integer())
  
  # Count number of plots
  for(i in 1:length(uniqueSpecies)) {
    tmp <- reducedEva$PlotObservationID[reducedEva$species == uniqueSpecies[i]]
    
    plotsPerSpecies <- 
      add_row(plotsPerSpecies, names = uniqueSpecies[i], numberOfPlots = length(unique(tmp)))
  }
} else {
  plotsPerSpecies <- read_csv("plotsPerSpeciesInEurope.csv")
}
```


# Data preparation
We want all species that are present in at least 50 plots, we order this set from highest to lowest and subsequently make a function to compute the total cover of the species across all sites.
```{r}
# determine which plant species should be computed
minimumNumberOfPlots<- 50
allPlants <- plotsPerSpecies[plotsPerSpecies$numberOfPlots >= minimumNumberOfPlots,]
allPlants <- allPlants[order(-allPlants$numberOfPlots),]

computeSpeciesCover <- function(name, species, cover) {
  result <- sum(cover[species==name]) # we use the sum here, since otherwise the total cover / species cover cannot reach one.
  if(is.na(result)) {
    0
  } else {
    result
  }
}
```


Test the function
```{r, warnings=F, message=F}
# Tests of the above function
computeSpeciesCover("nonExistentPlant", c("plant1", "plant2"), c(1,2)) # If the plant is not present 0 is returned.
computeSpeciesCover("plant1", c("plant1", "plant2"), c(1,2)) # If the is present the right value is returned
computeSpeciesCover("plant2", c("plant1", "plant2"), c(1,2))
computeSpeciesCover("plant2", c("plant1", "plant2", "plant2"), c(1,2,3)) # If the plant is present multiple times the sum is returned

reducedEva <- eva[,c("PlotObservationID", "species", "Cover %")]
```
# Function
```{r}
# The name of the first species
plantName = allPlants$names[1]

# Function to compute the impact for each species, so we only need the plant name as input
computePlantImpact <- function(plantName) {
  
  # Copy dataset temporarily to be able to make changes
  tmpFullPlotData <- fullPlotData
  
  # Get the indices of the plant in the eva dataset
  indexOfPlant <- eva$species == plantName
  
  # The majority of the time is spent in computing the cover classes. 
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We make the dataset smaller (only those plots where the species is present), group the data per plot and compute the total cover (all species) and cover of the species we are highlighting now
  plotsWherePlantOccurs <- reducedEva[reducedEva$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant],] |>
    group_by(PlotObservationID) |>
    summarise(totalCover = sum(`Cover %`), 
              speciesCover = computeSpeciesCover(plantName, species, `Cover %`))
  
  
  # Because the cover class was only computed for plots containing the species we need to assign 0% class to the remaining plots.
  # Join the computed cover plots with the full data
  tmpFullPlotData <- left_join(tmpFullPlotData, plotsWherePlantOccurs, by = "PlotObservationID")
  # Identify all plots where the species investigated occurs
  tmpFullPlotData$plantOccurs = tmpFullPlotData$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant]
  
  # If species is not present:
  # Make 1 total cover and 0 species cover
  tmpFullPlotData$totalCover[is.na(tmpFullPlotData$totalCover)] <- 1.0
  tmpFullPlotData$speciesCover[is.na(tmpFullPlotData$speciesCover)] <- 0.0
  
  # Exclude all plots which have 0 total cover
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$totalCover > 0,] 
  
  # Do SR - 1 for all plots in which the species is present
  tmpFullPlotData$numberOfVascularPlantSpecies <- tmpFullPlotData$numberOfVascularPlantSpecies - tmpFullPlotData$plantOccurs
  
  # Correct area: if true we calculate the area that is reserved for the other plant species. 
  if(correctArea) {
    tmpFullPlotData$Area <- tmpFullPlotData$Area*(1.0 - tmpFullPlotData$speciesCover/tmpFullPlotData$totalCover)
  }
  
  # There is a numerical problem in the bam function if we provide empty plots...
  # This should normally not be possible (as we did data screening) but I will check again
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$numberOfVascularPlantSpecies > 0 & tmpFullPlotData$Area > 0,]
  
  # Make dataset a factor to add it as a random variable
  tmpFullPlotData$Dataset <- as.factor(tmpFullPlotData$Dataset)
  
  # Count number of plots in which the species is present
  sum(tmpFullPlotData$plantOccurs)
  
  # Base model
  gam11 <- bam(numberOfVascularPlantSpecies ~ 
                 log(Area) + 
                 s(EIVEresM, bs = 'tp') +
                 s(EIVEresN, bs = 'tp') +
                 s(EIVEresR, bs = 'tp') +
                 s(EIVEresL, bs = 'tp') +
                 s(EIVEresT, bs = 'tp') +
                 s(transformedDisturbanceSeverity, bs = 'tp') +
                 s(transformedDisturbanceFrequency, bs = 'tp') +
                 s(Latitude, Longitude, bs = 'tp') +
                 s(Dataset, bs = 're') +
                 plantOccurs, 
               family = poisson(link = log), 
               data = tmpFullPlotData, 
               method = 'fREML',
               discrete=TRUE,
               nthreads=1)
  
  # Get summary from model
  gam11.sum <- summary(gam11)
  
  # Make a dataframe to add the residuals to
  res <- data.frame(taxa = character(), Estimate = numeric(), StdErr = numeric(), `zValue` = numeric(), pr = numeric(), numberOfPlots = integer())
  
  res <- add_row(res, taxa = plantName, Estimate = gam11.sum$p.table[3,1], StdErr = gam11.sum$p.table[3,2], zValue = gam11.sum$p.table[3,3], pr = gam11.sum$p.table[3,4], numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID))
  res
}
```


# Impact
Run whole function for smaller dataset
```{r}
allPlantsImpact <- data.frame(taxa = character(), Estimate = numeric(), StdErr = numeric(), `zValue` = numeric(), pr = numeric(), numberOfPlots = integer())

#testSpecies<- uniqueSpecies[1:100]

#for(j in 1: round(length(uniqueSpecies)/100)){
 # start<- 100*(j-1)+1
 # end<- 100*j
#  for(i in start:end) {
 # res <- computePlantImpact(allPlants$names[i])
  #allPlantsImpact <- rbind(allPlantsImpact, res)
 # }
  #name<-paste("allPlantImpact",j, sep = "")
  #assign(name,allPlantsImpact)
#}

#allPlantImpact94$neophyte<- allPlantImpact94$taxa %in% neophyteNames

#ggplot(allPlantImpact94, aes(x=neophyte, y= Estimate))+geom_violin()+ stat_summary(fun= "mean",  geom = "point", aes(group= neophyte), size=3)

#stat.test<-t.test((Estimate) ~ neophyte, allPlantImpact94)
```


Use foreach to speed up computations
```{r}
n.cores<- parallel::detectCores()-2 
my.cluster <- parallel::makeCluster(n.cores, type = "PSOCK" )
print(my.cluster)
doParallel::registerDoParallel(cl = my.cluster)
foreach::getDoParRegistered()
foreach::getDoParWorkers()

allPlantsImpact <- data.frame(taxa = character(), Estimate = numeric(), StdErr = numeric(), `zValue` = numeric(), pr = numeric(), numberOfPlots = integer())
length(allPlants$names)

begin<-Sys.time()
x<- foreach(i= 1:length(allPlants$names), .combine = "rbind", .packages = c("dplyr","mgcv")) %dopar% {
  res <- computePlantImpact(allPlants$names[i])
}
end<-Sys.time()
round(end-begin, 2)

x$neophyte<- x$taxa %in% neophyteNames

length(unique(x$taxa[x$neophyte]))

ggplot(x, aes(x=neophyte, y= Estimate))+geom_violin()+
  stat_summary(fun= "mean",
               geom = "point", aes(group= neophyte), size=3)

mean(x$Estimate[x$neophyte])
mean(x$Estimate[!x$neophyte])
t.test(Estimate~neophyte, x)
parallel::stopCluster(cl = my.cluster)
```




# 2 ANALYSIS
## 2.1 Function 1
```{r, meassage=F}
# The name of the first species (to test the function)
plantName = allPlants$names[1]
plantStatus= allPlants$Neophyte[1]


# Function to compute the impact for each species, so we only need the plant name as input
computePlantImpact <- function(plantName, plantStatus) {
  
  # Copy dataset temporarily to be able to make changes
  tmpFullPlotData <- fullPlotData
  
  # Select only explanatory variables
  tmpFullPlotData <- tmpFullPlotData %>% 
    select(numberOfVascularPlantSpecies,PlotObservationID, Area, 
           EIVEresM, EIVEresN, EIVEresR, EIVEresL, EIVEresT, 
           DistSeverity.sqrt, DistFrequency.sqrt, Soil.Disturbance.sqrt, 
           Grazing.Pressure.sqrt, Mowing.Frequency.sqrt, 
           Latitude, Longitude, Dataset, totalCover, chelsaP, elev, hfp)
  
  # Copy the full dataset without aberrant totalCover
  tmpFullPlotData <- tmpFullPlotData %>% filter(totalCover <= 900)
  reducedEva <- eva %>% filter(PlotObservationID %in% tmpFullPlotData$PlotObservationID)

  # Get the indices of the plant in the eva dataset
  indexOfPlant <- reducedEva$species == plantName & reducedEva$Neophyte== plantStatus
  
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We make the dataset smaller (only those plots where the species is present), group the data per plot and compute the total cover (all     species) and cover of the species we are highlighting now
  plotsWherePlantOccurs <- reducedEva[reducedEva$PlotObservationID %in%
                                      reducedEva$PlotObservationID[indexOfPlant],] |>
                           group_by(PlotObservationID) |>
                           summarise(speciesCover = computeSpeciesCover(plantName, species, `Cover %`))
  
  
  # Join the computed cover plots with the full data
  tmpFullPlotData <- left_join(tmpFullPlotData, plotsWherePlantOccurs, 
                             by = "PlotObservationID")
  
  # Identify all plots where the species investigated occurs
  tmpFullPlotData$plantOccurs = tmpFullPlotData$PlotObservationID %in%
    reducedEva$PlotObservationID[indexOfPlant]
  

  # If species is not present:
  # Make 100 total cover and 0 species cover
  tmpFullPlotData$totalCover[is.na(tmpFullPlotData$totalCover)] <- 100.0
  tmpFullPlotData$speciesCover[is.na(tmpFullPlotData$speciesCover)] <- 0.0
  
  # Exclude all plots which have 0 total cover
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$totalCover >=1,] 
  
  # Do SR - 1 for all plots in which the species is present
  tmpFullPlotData$numberOfVascularPlantSpecies <- tmpFullPlotData$numberOfVascularPlantSpecies - tmpFullPlotData$plantOccurs
  
  # Correct area: if true we calculate the area that is reserved for the other plant species. 
  if(correctArea) {
    tmpFullPlotData$Area <- tmpFullPlotData$Area*(1.0 - tmpFullPlotData$speciesCover/tmpFullPlotData$totalCover)
  }

  # There is a numerical problem in the bam function if we provide empty plots...
  tmpFullPlotData <- tmpFullPlotData |> filter(numberOfVascularPlantSpecies > 0, Area > 0)
  
  # Make dataset a factor to add it as a random variable
  tmpFullPlotData$Dataset <- as.factor(tmpFullPlotData$Dataset)
  
  # Add relative species cover
  tmpFullPlotData <- tmpFullPlotData %>% 
    mutate(relSpeciesCover = 100 * speciesCover / totalCover)
  
  
  # Species occurrence as factor (not logical)
  tmpFullPlotData$plantOccursF <- factor(tmpFullPlotData$plantOccurs)
  
  # Base model
  model<-  numberOfVascularPlantSpecies ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')+
              plantOccurs

  base <- bam(model,family = poisson, 
              data = tmpFullPlotData,  method = 'fREML',  discrete=TRUE, 
              nthreads=4)


  library(DHARMa)
  library(mgcViz)

  test<- simulateResiduals(base)
  plot(test) 
  testDispersion(base)
  
  # Get summary from model
  baseModel <- summary(base)
  baseModel
  
  # Make a dataframe to add the residuals to
  res <- data.frame(taxa = character(), Estimate = numeric(), 
                    StdErr = numeric(), `zValue` = numeric(), 
                    pr = numeric(), numberOfPlots = integer(), 
                    Neophyte= character(),
                    RelDiff= numeric())
  
  res <- add_row(res, 
                 taxa = plantName, Estimate = baseModel$p.table[2,1], 
                 StdErr = baseModel$p.table[2,2], 
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4], 
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$plantOccursF == TRUE]) -
                    mean(fitted(base)[tmpFullPlotData$plantOccursF == FALSE])) /
                    mean(fitted(base)[tmpFullPlotData$plantOccursF == FALSE]))
  
  res
}
```


## 2.1 Function 1.1
```{r, meassage=F}
# The name of the first species (to test the function)
plantName = allPlants$names[1]
plantStatus= allPlants$Neophyte[1]


# Function to compute the impact for each species, so we only need the plant name as input
computePlantImpact1 <- function(plantName, plantStatus) {
  
  # Copy dataset temporarily to be able to make changes
  tmpFullPlotData <- fullPlotData
  
  # Select only explanatory variables
  tmpFullPlotData <- tmpFullPlotData %>% 
    select(numberOfVascularPlantSpecies,PlotObservationID, Area, 
           EIVEresM, EIVEresN, EIVEresR, EIVEresL, EIVEresT, 
           DistSeverity.sqrt, DistFrequency.sqrt, Soil.Disturbance.sqrt, 
           Grazing.Pressure.sqrt, Mowing.Frequency.sqrt, 
           Latitude, Longitude, Dataset, totalCover, chelsaP, elev, hfp, PC1,PC2,PC3,PC4)
  
  # Copy the full dataset without aberrant totalCover
  tmpFullPlotData <- tmpFullPlotData %>% filter(totalCover <= 900)
  reducedEva <- eva %>% filter(PlotObservationID %in% tmpFullPlotData$PlotObservationID)

  # Get the indices of the plant in the eva dataset
  indexOfPlant <- reducedEva$species == plantName & reducedEva$Neophyte== plantStatus
  
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We make the dataset smaller (only those plots where the species is present), group the data per plot and compute the total cover (all     species) and cover of the species we are highlighting now
  plotsWherePlantOccurs <- reducedEva[reducedEva$PlotObservationID %in%
                                      reducedEva$PlotObservationID[indexOfPlant],] |>
    group_by(PlotObservationID) |>
    summarise(speciesCover = computeSpeciesCover(plantName, species, `Cover %`))
  
  
  # Join the computed cover plots with the full data
  tmpFullPlotData <- left_join(tmpFullPlotData, plotsWherePlantOccurs, 
                             by = "PlotObservationID")
  
  # Identify all plots where the species investigated occurs
  tmpFullPlotData$plantOccurs = tmpFullPlotData$PlotObservationID %in%
    reducedEva$PlotObservationID[indexOfPlant]
  
  
  # reduce the dataset to obtain only those plots in which the species is present
  species_plots <- tmpFullPlotData[tmpFullPlotData$plantOccurs, c(21:24)  ]
  # make matrix and reduce duplicated columns
  species_plots<- as.matrix(species_plots)
  duplicated<- duplicated(species_plots)
  species_plots<- species_plots[!duplicated, ]
  
  # take ConvexHull around the PCA coordinates in which the species is present
  convex_hull <- convhulln((species_plots))

  # Extract all points from the dataset that are additionally within this ConvexHull and do not contain the species
  is_within_hull <- inhulln(convex_hull, as.matrix(tmpFullPlotData[, c(21:24)]))
  subset_data <- tmpFullPlotData[is_within_hull, ]
  
  # Take size of subset data
  size<- length(subset_data$plantOccurs)
  # change name tmpFullPlotData
  tmpFullPlotData <- subset_data
  
  # If species is not present:
  # Make 1 total cover and 0 species cover
  tmpFullPlotData$totalCover[is.na(tmpFullPlotData$totalCover)] <- 100.0
  tmpFullPlotData$speciesCover[is.na(tmpFullPlotData$speciesCover)] <- 0.0
  
  # Exclude all plots which have 0 total cover
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$totalCover >=1,] 
  
  # Do SR - 1 for all plots in which the species is present
  tmpFullPlotData$numberOfVascularPlantSpecies <- tmpFullPlotData$numberOfVascularPlantSpecies - tmpFullPlotData$plantOccurs
  
  # Correct area: if true we calculate the area that is reserved for the other plant species. 
  if(correctArea) {
    tmpFullPlotData$Area <- tmpFullPlotData$Area*(1.0 - tmpFullPlotData$speciesCover/tmpFullPlotData$totalCover)
  }

  # There is a numerical problem in the bam function if we provide empty plots...
  tmpFullPlotData <- tmpFullPlotData |> filter(numberOfVascularPlantSpecies > 0, Area > 0)
  
  # Make dataset a factor to add it as a random variable
  tmpFullPlotData$Dataset <- as.factor(tmpFullPlotData$Dataset)
  
  # Add relative species cover
  tmpFullPlotData <- tmpFullPlotData %>% 
    mutate(relSpeciesCover = 100 * speciesCover / totalCover)
  
  
  # Species occurrence as factor (not logical)
  tmpFullPlotData$plantOccursF <- factor(tmpFullPlotData$plantOccurs)
  
  # Base model
  model<-  numberOfVascularPlantSpecies ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs = 're')+
              plantOccurs

  base <- bam(model,family = poisson, 
              data = tmpFullPlotData,  method = 'fREML',  discrete=TRUE, 
              nthreads=4)
  
  # Get summary from model
  baseModel <- summary(base)
  baseModel
  
  # Make a dataframe to add the residuals to
  res <- data.frame(taxa = character(), Estimate = numeric(), 
                    StdErr = numeric(), `zValue` = numeric(), 
                    pr = numeric(), numberOfPlots = integer(), 
                    Neophyte= character(),
                    RelDiff= numeric(),
                    size= numeric())
  
  res <- add_row(res, 
                 taxa = plantName, Estimate = baseModel$p.table[2,1], 
                 StdErr = baseModel$p.table[2,2], 
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4], 
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$plantOccursF == TRUE]) -
                    mean(fitted(base)[tmpFullPlotData$plantOccursF == FALSE])) /
                    mean(fitted(base)[tmpFullPlotData$plantOccursF == FALSE]),
                 size= size)
  
  res
}
```




## 2.2 Function 2
```{r, meassage=F}
# The name of the first species (to test the function)
plantName = allPlants$names[1]
plantStatus= allPlants$Neophyte[1]
i=1

# Function to compute the impact for each species, so we only need the plant name as input
computePlantImpact2 <- function(plantName, plantStatus) {
  
  
  # Copy the full dataset without aberrant totalCover
  tmpFullPlotData <- fullPlotData %>% filter(totalCover < 850)
  reducedEva <- eva %>% filter(PlotObservationID %in% tmpFullPlotData$PlotObservationID)
  
  # Select only explanatory variables
  tmpFullPlotData <- tmpFullPlotData %>% 
    select(PlotObservationID, numberOfVascularPlantSpecies, Area, 
           EIVEresM, EIVEresN, EIVEresR, EIVEresL, EIVEresT, 
           DistSeverity.sqrt, DistFrequency.sqrt, Soil.Disturbance.sqrt, 
           Grazing.Pressure.sqrt, Mowing.Frequency.sqrt, 
           Latitude, Longitude, Dataset, totalCover)
  
  # Get the indices of the plant in the eva dataset
  indexOfPlant <- eva$species == plantName & eva$Neophyte== plantStatus
  
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We make the dataset smaller (only those plots where the species is present), group the data per plot and 
  # compute the cover of the species we are highlighting now
  plotsWherePlantOccurs <- reducedEva[reducedEva$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant],] |>
    group_by(PlotObservationID) |>
    summarise(speciesCover = computeSpeciesCover(plantName, species, `Cover %`))
  
  # Join the computed cover plots with the full data
  tmpFullPlotData <- left_join(tmpFullPlotData, plotsWherePlantOccurs, by = "PlotObservationID")
  # Identify all plots where the species investigated occurs
  tmpFullPlotData$plantOccurs = tmpFullPlotData$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant]
  
  # If species is not present:
  tmpFullPlotData$speciesCover[is.na(tmpFullPlotData$speciesCover)] <- 0.0
  
  # Exclude all plots which have 0 total cover
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$totalCover >= 1,] 
  
  # Do SR - 1 for all plots in which the species is present
  tmpFullPlotData$numberOfVascularPlantSpecies <- tmpFullPlotData$numberOfVascularPlantSpecies - tmpFullPlotData$plantOccurs
  
  # Correct area: if true we calculate the area that is reserved for the other plant species. 
  if(correctArea) {
    tmpFullPlotData$Area <- tmpFullPlotData$Area*(1.0 - tmpFullPlotData$speciesCover/tmpFullPlotData$totalCover)
  }
  
  # There is a numerical problem in the bam function if we provide empty plots...
  tmpFullPlotData <- tmpFullPlotData %>% filter(numberOfVascularPlantSpecies > 0, Area > 0)
  
  # Make dataset a factor to add it as a random variable
  tmpFullPlotData$Dataset <- as.factor(tmpFullPlotData$Dataset)

  # Add relative species cover
  tmpFullPlotData <- tmpFullPlotData %>% mutate(relSpeciesCover = 100 * speciesCover / totalCover)

  # Species occurrence as factor (not logical)
  tmpFullPlotData$plantOccursF <- factor(tmpFullPlotData$plantOccurs)
  
  # Base model
  model<-  (numberOfVascularPlantSpecies) ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs="re")+
              (relSpeciesCover)+ I(relSpeciesCover^2)
  
  
  base <- bam(model,family = poisson, data = tmpFullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4, link= "identity")
  baseModel<- summary(base)
  baseModel
  
    
  visreg::visreg(base, xvar="relSpeciesCover", type="conditional", scale='response')
  
  # Make a dataframe to add the residuals to
  res <- data.frame(taxa = character(), Estimate = numeric(),Estimate2 = numeric(), 
                    StdErr = numeric(),StdErr2 = numeric(), 
                    `zValue` = numeric(),`zValue2` = numeric(), 
                    pr = numeric(),pr2 = numeric(), 
                    Intercept= numeric(), 
                    numberOfPlots = integer(), 
                    Neophyte= character(),
                    minx= numeric(), maxx= numeric())
  
  res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],Estimate2 = baseModel$p.table[3,1],
                 StdErr = baseModel$p.table[2,2],StdErr2 = baseModel$p.table[3,2], 
                 zValue = baseModel$p.table[2,3], zValue2 = baseModel$p.table[3,3], 
                 pr = baseModel$p.table[2,4],pr2 = baseModel$p.table[3,4], 
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 minx= min(tmpFullPlotData$relSpeciesCover),
                 maxx=max(tmpFullPlotData$relSpeciesCover))
  res
}
```


## 2.3 Function 3
```{r, meassage=F}
# The name of the first species (to test the function)
plantName = allPlants$names[1]
plantStatus= allPlants$Neophyte[1]
i=1

# Function to compute the impact for each species, so we only need the plant name as input
computePlantImpact3 <- function(plantName, plantStatus) {
  
  
  # Copy the full dataset without aberrant totalCover
  tmpFullPlotData <- fullPlotData %>% filter(totalCover < 850)
  reducedEva <- eva %>% filter(PlotObservationID %in% tmpFullPlotData$PlotObservationID)
  
  # Select only explanatory variables
  tmpFullPlotData <- tmpFullPlotData %>% 
    select(PlotObservationID, numberOfVascularPlantSpecies, Area, 
           EIVEresM, EIVEresN, EIVEresR, EIVEresL, EIVEresT, 
           DistSeverity.sqrt, DistFrequency.sqrt, Soil.Disturbance.sqrt, 
           Grazing.Pressure.sqrt, Mowing.Frequency.sqrt, 
           Latitude, Longitude, Dataset, totalCover)
  
  # Get the indices of the plant in the eva dataset
  indexOfPlant <- eva$species == plantName & eva$Neophyte== plantStatus
  
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We make the dataset smaller (only those plots where the species is present), group the data per plot and 
  # compute the cover of the species we are highlighting now
  plotsWherePlantOccurs <- reducedEva[reducedEva$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant],] |>
    group_by(PlotObservationID) |>
    summarise(speciesCover = computeSpeciesCover(plantName, species, `Cover %`))
  
  # Join the computed cover plots with the full data
  tmpFullPlotData <- left_join(tmpFullPlotData, plotsWherePlantOccurs, by = "PlotObservationID")
  # Identify all plots where the species investigated occurs
  tmpFullPlotData$plantOccurs = tmpFullPlotData$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant]
  
  # If species is not present:
  tmpFullPlotData$speciesCover[is.na(tmpFullPlotData$speciesCover)] <- 0.0
  
  # Exclude all plots which have 0 total cover
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$totalCover >= 1,] 
  
  # Do SR - 1 for all plots in which the species is present
  tmpFullPlotData$numberOfVascularPlantSpecies <- tmpFullPlotData$numberOfVascularPlantSpecies - tmpFullPlotData$plantOccurs
  
  # Correct area: if true we calculate the area that is reserved for the other plant species. 
  if(correctArea) {
    tmpFullPlotData$Area <- tmpFullPlotData$Area*(1.0 - tmpFullPlotData$speciesCover/tmpFullPlotData$totalCover)
  }
  
  # There is a numerical problem in the bam function if we provide empty plots...
  tmpFullPlotData <- tmpFullPlotData %>% filter(numberOfVascularPlantSpecies > 0, Area > 0)
  
  # Make dataset a factor to add it as a random variable
  tmpFullPlotData$Dataset <- as.factor(tmpFullPlotData$Dataset)

  # Add relative species cover
  tmpFullPlotData <- tmpFullPlotData %>% mutate(relSpeciesCover = 100 * speciesCover / totalCover)

  # Species occurrence as factor (not logical)
  tmpFullPlotData$plantOccursF <- factor(tmpFullPlotData$plantOccurs)
  tmpFullPlotData
}

# Make a dataframe to add the residuals to
  res <- data.frame(taxa = character(), Estimate = numeric(),Estimate2 = numeric(), StdErr = numeric(),StdErr2 = numeric(), `zValue` = numeric(),`zValue2` = numeric(), pr = numeric(),pr2 = numeric(), Intercept= numeric(), numberOfPlots = integer(), Neophyte= character(), minx=numeric(), maxx=numeric())
```


## 2.4 Function 4
```{r, meassage=F}
# The name of the first species (to test the function)
plantName = allPlants$names[1]
plantStatus= allPlants$Neophyte[1]
i=1

# Function to compute the impact for each species, so we only need the plant name as input
computePlantImpact4 <- function(plantName, plantStatus) {
  
  
  # Copy the full dataset without aberrant totalCover
  tmpFullPlotData <- fullPlotData %>% filter(totalCover <=900)
  reducedEva <- eva %>% filter(PlotObservationID %in% tmpFullPlotData$PlotObservationID)
  
  # Select only explanatory variables
  tmpFullPlotData <- tmpFullPlotData %>% 
    select(PlotObservationID, numberOfVascularPlantSpecies, Area, 
           EIVEresM, EIVEresN, EIVEresR, EIVEresL, EIVEresT, 
           DistSeverity.sqrt, DistFrequency.sqrt, Soil.Disturbance.sqrt, 
           Grazing.Pressure.sqrt, Mowing.Frequency.sqrt, 
           Latitude, Longitude, Dataset, elev, chelsaP, hfp, totalCover)
  
  # Get the indices of the plant in the eva dataset
  indexOfPlant <- reducedEva$species == plantName & reducedEva$Neophyte== plantStatus
  
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We make the dataset smaller (only those plots where the species is present), group the data per plot and 
  # compute the cover of the species we are highlighting now
  plotsWherePlantOccurs <- reducedEva[reducedEva$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant],] |>
    group_by(PlotObservationID) |>
    summarise(speciesCover = computeSpeciesCover(plantName, species, `Cover %`)) 
  
  # Join the computed cover plots with the full data
  tmpFullPlotData <- left_join(tmpFullPlotData, plotsWherePlantOccurs, by = "PlotObservationID")
  # Identify all plots where the species investigated occurs
  tmpFullPlotData$plantOccurs = tmpFullPlotData$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant]
  
  # If species is not present:
  tmpFullPlotData$speciesCover[is.na(tmpFullPlotData$speciesCover)] <- 0.0
  
  # Exclude all plots which have 0 total cover
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$totalCover >= 1,] 
  
  # Do SR - 1 for all plots in which the species is present
  tmpFullPlotData$numberOfVascularPlantSpecies <- tmpFullPlotData$numberOfVascularPlantSpecies - tmpFullPlotData$plantOccurs
  
  # Correct area: if true we calculate the area that is reserved for the other plant species. 
  if(correctArea) {
    tmpFullPlotData$Area <- tmpFullPlotData$Area*(1.0 - tmpFullPlotData$speciesCover/tmpFullPlotData$totalCover)
  }
  
  # There is a numerical problem in the bam function if we provide empty plots...
  tmpFullPlotData <- tmpFullPlotData %>% filter(numberOfVascularPlantSpecies > 0, Area > 0)
  
  # Make dataset a factor to add it as a random variable
  tmpFullPlotData$Dataset <- as.factor(tmpFullPlotData$Dataset)

  # Add relative species cover
  tmpFullPlotData <- tmpFullPlotData %>% 
    mutate(relSpeciesCover = 100 * speciesCover / totalCover) |>
    mutate(coverClass = cut(relSpeciesCover,breaks = c(-1, 0, 1, 10, 30, 50, 70, 101), 
                            labels = c("0%", "0%-1%", "1%-10%", "10%-30%", "30%-50%","50%-70%","70%-100%")))
  
  tmpFullPlotData$coverClass<- as.character( tmpFullPlotData$coverClass)
  tmpFullPlotData$coverClass[(tmpFullPlotData$speciesCover==0.0)] <- c("0%")
  
  # Species occurrence as factor (not logical)
  tmpFullPlotData$plantOccursF <- factor(tmpFullPlotData$plantOccurs)
  tmpFullPlotData$coverClass <- factor(tmpFullPlotData$coverClass)
  
  # Base model
  model<-  (numberOfVascularPlantSpecies) ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(chelsaP, bs='tp')+
              s(hfp, bs='tp')+
              s(elev, bs='tp')+
              s(Dataset, bs="re")+
              coverClass
  
  
  base <- bam(model,family = poisson, data = tmpFullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4, link= "identity")
  baseModel<- summary(base)
  baseModel
  
  # Make a dataframe to add the residuals to
  res <- data.frame(taxa = character(), Estimate = numeric(),
                    StdErr = numeric(),
                    `zValue` = numeric(),
                    pr = numeric(),
                    Intercept= numeric(), 
                    numberOfPlots = integer(), 
                    Neophyte= character(),
                    RelDiff= numeric(),
                    class= character(), n=numeric())
  
  res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "0%-1%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "0%-1%", n= sum(tmpFullPlotData$coverClass == "0%-1%"))
   res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "1%-10%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "1%-10%", n= sum(tmpFullPlotData$coverClass == "1%-10%"))
   
   res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "10%-30%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "10%-30%", n= sum(tmpFullPlotData$coverClass == "10%-30%"))
  res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "30%-50%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "30%-50%", n= sum(tmpFullPlotData$coverClass == "30%-50%"))
   res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "50%-70%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class="50%-70%", n= sum(tmpFullPlotData$coverClass == "50%-70%"))
  
  res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff = 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "70%-100%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "70%-100%", n= sum(tmpFullPlotData$coverClass == "70%-100%"))
  
res
}
```

## 2.5 Function 5
```{r, meassage=F}
# The name of the first species (to test the function)
plantName = allPlants$names[1]
plantStatus= allPlants$Neophyte[1]

# Function to compute the impact for each species, so we only need the plant name as input
computePlantImpact5 <- function(plantName, plantStatus) {
  
  
  # Copy the full dataset without aberrant totalCover
  tmpFullPlotData <- fullPlotData %>% filter(totalCover <=900)
  reducedEva <- eva %>% filter(PlotObservationID %in% tmpFullPlotData$PlotObservationID)
  
  # Select only explanatory variables
  tmpFullPlotData <- tmpFullPlotData %>% 
    select(PlotObservationID, numberOfVascularPlantSpecies, Area, 
           EIVEresM, EIVEresN, EIVEresR, EIVEresL, EIVEresT, 
           DistSeverity.sqrt, DistFrequency.sqrt, Soil.Disturbance.sqrt, 
           Grazing.Pressure.sqrt, Mowing.Frequency.sqrt, 
           Latitude, Longitude, Dataset, elev, chelsaP, hfp, totalCover, PC1, PC2, PC3,PC4)
  
  # Get the indices of the plant in the eva dataset
  indexOfPlant <- reducedEva$species == plantName & reducedEva$Neophyte== plantStatus
  
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We make the dataset smaller (only those plots where the species is present), group the data per plot and 
  # compute the cover of the species we are highlighting now
  plotsWherePlantOccurs <- reducedEva[reducedEva$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant],] |>
    group_by(PlotObservationID) |>
    summarise(speciesCover = computeSpeciesCover(plantName, species, `Cover %`)) 
  
  # Join the computed cover plots with the full data
  tmpFullPlotData <- left_join(tmpFullPlotData, plotsWherePlantOccurs, by = "PlotObservationID")
  # Identify all plots where the species investigated occurs
  tmpFullPlotData$plantOccurs = tmpFullPlotData$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant]
  
   # reduce the dataset to obtain only those plots in which the species is present
  species_plots <- tmpFullPlotData[tmpFullPlotData$plantOccurs, c(21:24)  ]
  # make matrix and reduce duplicated columns
  species_plots<- as.matrix(species_plots)
  duplicated<- duplicated(species_plots)
  species_plots<- species_plots[!duplicated, ]
  
  # take ConvexHull around the PCA coordinates in which the species is present
  convex_hull <- convhulln((species_plots))

  # Extract all points from the dataset that are additionally within this ConvexHull and do not contain the species
  is_within_hull <- inhulln(convex_hull, as.matrix(tmpFullPlotData[, c(21:24)]))
  subset_data <- tmpFullPlotData[is_within_hull, ]
  
  # Take size of subset data
  size<- length(subset_data$plantOccurs)
  # change name tmpFullPlotData
  tmpFullPlotData <- subset_data
  
  
  # If species is not present:
  tmpFullPlotData$speciesCover[is.na(tmpFullPlotData$speciesCover)] <- 0.0
  tmpFullPlotData$totalCover[is.na(tmpFullPlotData$speciesCover)] <- 100.0

  # Exclude all plots which have 0 total cover
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$totalCover >= 1,] 
  
  # Do SR - 1 for all plots in which the species is present
  tmpFullPlotData$numberOfVascularPlantSpecies <- tmpFullPlotData$numberOfVascularPlantSpecies - tmpFullPlotData$plantOccurs
  
  # Correct area: if true we calculate the area that is reserved for the other plant species. 
  if(correctArea) {
    tmpFullPlotData$Area <- tmpFullPlotData$Area*(1.0 - tmpFullPlotData$speciesCover/tmpFullPlotData$totalCover)
  }
  
  # There is a numerical problem in the bam function if we provide empty plots...
  tmpFullPlotData <- tmpFullPlotData %>% filter(numberOfVascularPlantSpecies > 0, Area > 0)
  
  # Make dataset a factor to add it as a random variable
  tmpFullPlotData$Dataset <- as.factor(tmpFullPlotData$Dataset)

  # Add relative species cover
  tmpFullPlotData <- tmpFullPlotData %>% 
    mutate(relSpeciesCover = 100 * speciesCover / totalCover) |>
    mutate(coverClass = cut(relSpeciesCover,breaks = c(-1, 0, 1, 10, 30, 50, 70, 101), 
                            labels = c("0%", "0%-1%", "1%-10%", "10%-30%", "30%-50%","50%-70%","70%-100%")))
  
  tmpFullPlotData$coverClass<- as.character( tmpFullPlotData$coverClass)
  tmpFullPlotData$coverClass[(tmpFullPlotData$speciesCover==0.0)] <- c("0%")
  
  # Species occurrence as factor (not logical)
  tmpFullPlotData$plantOccursF <- factor(tmpFullPlotData$plantOccurs)
  tmpFullPlotData$coverClass <- factor(tmpFullPlotData$coverClass)
  
  # Base model
  model<-  (numberOfVascularPlantSpecies) ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(chelsaP, bs='tp')+
              s(hfp, bs='tp')+
              s(elev, bs='tp')+
              s(Dataset, bs="re")+
              coverClass
  
  
  base <- bam(model,family = poisson, data = tmpFullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4, link= "identity")
  baseModel<- summary(base)
  baseModel
  
  # Make a dataframe to add the residuals to
  res <- data.frame(taxa = character(), Estimate = numeric(),
                    StdErr = numeric(),
                    `zValue` = numeric(),
                    pr = numeric(),
                    Intercept= numeric(), 
                    numberOfPlots = integer(), 
                    Neophyte= character(),
                    RelDiff= numeric(),
                    class= character(), n=numeric())
  
  res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "0%-1%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "0%-1%", n= sum(tmpFullPlotData$coverClass == "0%-1%"))
   res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "1%-10%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "1%-10%", n= sum(tmpFullPlotData$coverClass == "1%-10%"))
   
   res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "10%-30%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "10%-30%", n= sum(tmpFullPlotData$coverClass == "10%-30%"))
  res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "30%-50%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "30%-50%", n= sum(tmpFullPlotData$coverClass == "30%-50%"))
   res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff= 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "50%-70%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class="50%-70%", n= sum(tmpFullPlotData$coverClass == "50%-70%"))
  
  res <- add_row(res, taxa = plantName, 
                 Estimate = baseModel$p.table[2,1],
                 StdErr = baseModel$p.table[2,2],
                 zValue = baseModel$p.table[2,3], 
                 pr = baseModel$p.table[2,4],
                 Intercept= baseModel$p.table[1,1],  
                 numberOfPlots = length(plotsWherePlantOccurs$PlotObservationID), 
                 Neophyte= plantStatus,
                 RelDiff = 100* 
                   (mean(fitted(base)[tmpFullPlotData$coverClass == "70%-100%"]) -
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"])) /
                    mean(fitted(base)[tmpFullPlotData$coverClass == "0%"]),
                 class= "70%-100%", n= sum(tmpFullPlotData$coverClass == "70%-100%"))
  
res
}
```


## 2.5 Impact
Use foreach to speed up computations
```{r}
if(fast){

# prepare parallel
n.cores<- parallel::detectCores()-2
my.cluster <- parallel::makeCluster(n.cores, type = "PSOCK" )
print(my.cluster)
doParallel::registerDoParallel(cl = my.cluster)
foreach::getDoParRegistered()
foreach::getDoParWorkers()

# run for all and time
begin<-Sys.time()
x<- foreach(i= 1:length(allPlants$names), 
            .combine = "rbind", 
            .packages = c("dplyr","mgcv","geometry","factoextra")) %dopar% {
  res <- computePlantImpact5(allPlants$names[i], allPlants$Neophyte[i])
}
end<-Sys.time()
round(end-begin, 2)

parallel::stopCluster(cl = my.cluster)
#write.csv(x, 'Impact2_new70%_ENS0_reldiff_PCA.csv')
}
```


```{r}
if(!fast){

allPlantsImpact <-  data.frame(taxa = character(), Estimate = numeric(), StdErr = numeric(), `zValue` = numeric(), pr = numeric(), numberOfPlots = integer(), Neophyte= character())

i=1
begin<- Sys.time()
for(i in 1:length(allPlants$names)) {
   res <- computePlantImpact2(allPlants$names[i], allPlants$Neophyte[i])
   allPlantsImpact <- rbind(allPlantsImpact, res)
}
end<- Sys.time()
round(end-begin)
}

x<- allPlantsImpact
#write.csv(x, 'Impact2_new100%_ENS0_squared.csv')

```


```{r}
if(method==3){
# prepare parallel
n.cores<- parallel::detectCores()-6
my.cluster <- parallel::makeCluster(n.cores, type = "PSOCK" )
print(my.cluster)
doParallel::registerDoParallel(cl = my.cluster)
foreach::getDoParRegistered()
foreach::getDoParWorkers()
i<-1
# run for all and time
begin<-Sys.time()

x<- foreach(i= 1:length(allPlants$names), .combine = "rbind", .packages = c("dplyr","mgcv")) %dopar% {
  tmpFullPlotData <- computePlantImpact3(allPlants$names[i], allPlants$Neophyte[i])

  # Base model
  model<-  (numberOfVascularPlantSpecies) ~ 
              s(log(Area),bs='tp') + 
              s(EIVEresM, bs = 'tp') +
              s(EIVEresN, bs = 'tp') +
              s(EIVEresR, bs = 'tp') +
              s(EIVEresL, bs = 'tp') +
              s(EIVEresT, bs = 'tp') +
              s(DistSeverity.sqrt, bs = 'tp') +
              s(Soil.Disturbance.sqrt, bs = 'tp')+
              s(Grazing.Pressure.sqrt, bs = 'tp')+
              s(Mowing.Frequency.sqrt, bs = 'tp') +
              s(Latitude, Longitude, bs = 'tp') +
              s(Dataset, bs="re")+
              (speciesCover)+ I(speciesCover^2)
  
  
  base <- bam(model,family = poisson, data = tmpFullPlotData,  method = 'fREML',  discrete=TRUE, nthreads=4, link= "identity")
  baseModel<- summary(base)

  # Make a dataframe to add the residuals to
  res <- data.frame(taxa = character(), Estimate = numeric(),Estimate2 = numeric(), StdErr = numeric(),StdErr2 = numeric(), `zValue` = numeric(),`zValue2` = numeric(), pr = numeric(),pr2 = numeric(), Intercept= numeric(), numberOfPlots = integer(), Neophyte= character(), minx=numeric(), maxx=numeric())
  
  res <- add_row(res, taxa = allPlants$names[i], Estimate = baseModel$p.table[2,1],Estimate2 = baseModel$p.table[3,1], StdErr = baseModel$p.table[2,2],StdErr2 = baseModel$p.table[3,2], zValue = baseModel$p.table[2,3], zValue2 = baseModel$p.table[3,3], pr = baseModel$p.table[2,4],pr2 = baseModel$p.table[3,4], Intercept= baseModel$p.table[1,1],  numberOfPlots = sum(tmpFullPlotData$plantOccurs), Neophyte= allPlants$Neophyte[i], minx= min(tmpFullPlotData$speciesCover), maxx=max(tmpFullPlotData$speciesCover))
}
end<- Sys.time()
round(end-begin)
stopCluster(my.cluster)
}
```



## 2.6 Significance
```{r}
# Probably (but currently only 10% run) no normality for especially the intra class and also very little data compared to the rest, so ranked anova performed
method=4
result<-(aov(rank(Estimate) ~ Neophyte, x))
summary(result)
TukeyHSD(result)

if(method==1){
result<-(aov(rank(RelDiff) ~ Neophyte, x))
summary(result)
TukeyHSD(result)
}

if(method==4){
result<-(aov(rank(RelDiff) ~ Neophyte + class, x))
summary(result)
TukeyHSD(result)
}
```



# 3 PLOT
## 3.1 Estimate
```{r}
if(!native_intra_analysis){
x <- x %>% mutate(Species = factor(Neophyte, 
                          levels = c("native", "intra","extra"),
                          labels = c("native in the country", 
                                     "intra-European neophyte", 
                                     "extra-European neophyte")))
p<-ggplot(x, aes(x=Neophyte, y= RelDiff, group = Species,
             color = Species, fill = Species))+
  geom_violin(alpha=0.5)+
  stat_summary(fun= "mean",
               geom = "point", aes(group= Neophyte), size=3)+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"))+
  scale_fill_manual(values = c("#1E88E5", "#FFC107", "#004D40")) +
  theme_minimal()+ 
  theme(legend.position = "bottom")+theme(legend.text= element_text(size=12), 
                                          legend.title =element_text(size=14))+ 
  annotate("text", x = 1, y = mean(x$RelDiff[x$Neophyte=="extra"], na.rm=T), 
           label = "a", size = 3.5, vjust = -1, hjust = 0.5) +
  annotate("text", x = 2, y = mean(x$RelDiff[x$Neophyte=="intra"], na.rm=T), 
           label = "a", size = 3.5, vjust = -1, hjust = 0.5)+
  annotate("text", x = 3, y = mean(x$RelDiff[x$Neophyte=="native"], na.rm=T), 
           label = "b", size = 3.5, vjust = -1, hjust = 0.5)+
  annotate("text", x=1, y=210, 
           label= paste("n=",length(x$Estimate[x$Neophyte=="extra"])), 
           vjust = -0.5, hjust = 0.5, size=3.5)+
  annotate("text", x=2, y=210, 
           label= paste("n=",length(x$Estimate[x$Neophyte=="intra"])), 
           vjust = -0.5, hjust = 0.5, size=3.5)+
  annotate("text", x=3, y=210, 
           label= paste("n=",length(x$Estimate[x$Neophyte=="native"])), 
           vjust = -0.5, hjust = 0.5, size=3.5)
p

#ggsave("Impact_base_model_new_PCA.jpeg", p, width = 25, height = 15, units = "cm")
} else {
x <- x %>% mutate(Species = factor(Neophyte, 
                          levels = c("native", "intra","extra", "native_intra"),
                          labels = c("native in the country", 
                                     "intra-European neophyte", 
                                     "extra-European neophyte",
                                     "native in country alien elsewhere")))
p<-ggplot(x, aes(x=Neophyte, y= Estimate, color=Neophyte))+geom_violin()+
  stat_summary(fun= "mean",
               geom = "point", aes(group= Neophyte), size=3)+
   scale_colour_manual(values=c("#1E88E5","#B71C1C", "#FFC107", "#388E3C"), 
                    name="Legend",
                    breaks=c("native","native_intra", "intra","extra"),
                    labels=c("native species","native species alien elsewhere", "intra European neophyte", "extra European neophyte"))+
  theme_classic()+ 
  theme(legend.position = "right")+theme(legend.text= element_text(size=12), legend.title =element_text(size=14))+
  annotate("text", x = 1, y = mean(x$Estimate[x$Neophyte=="extra"], na.rm=T), label = "a", size = 3.5, vjust = -1, hjust = 0.5) +
  annotate("text", x = 2, y = mean(x$Estimate[x$Neophyte=="intra"], na.rm=T), label = "ab", size = 3.5, vjust = -1, hjust = 0.5)+
  annotate("text", x = 3, y = mean(x$Estimate[x$Neophyte=="native"], na.rm=T), label = "c", size = 3.5, vjust = -1, hjust = 0.5)+
  annotate("text", x = 4, y = mean(x$Estimate[x$Neophyte=="native_intra"], na.rm=T), label = "b", size = 3.5, vjust = -1, hjust = 0.5)+
  annotate("text", x=1, y=0.8, label= paste("n=",length(x$Estimate[x$Neophyte=="extra"])), vjust = -0.5, hjust = 0.5, size=3.5)+
  annotate("text", x=2, y=0.8, label= paste("n=",length(x$Estimate[x$Neophyte=="intra"])), vjust = -0.5, hjust = 0.5, size=3.5)+
  annotate("text", x=3, y=0.8, label= paste("n=",length(x$Estimate[x$Neophyte=="native"])), vjust = -0.5, hjust = 0.5, size=3.5)+
  annotate("text", x=4, y=0.8, label= paste("n=",length(x$Estimate[x$Neophyte=="native_intra"])), vjust = -0.5, hjust = 0.5, size=3.5)
p

#ggsave("Impact_base_model_new10_ENS1.jpeg", p, width = 25, height = 15, units = "cm") 
}
```

## 3.2 Squared species
```{r}
# make x values from 0 to 100
xaxis <- seq(from=0, to=100, by=1)
res<-x

# Calculate all values for the estimations
matrix<-c()
for(i in 1: length(res$Estimate)){
  yvalues<- data.frame(xax=numeric(), mean = numeric(), status =character(), taxa= character(), n=numeric())
  xax<- c(floor(res$minx[i]): floor(res$maxx[i]))
  mean<- res$Intercept[i]+ xax*(res$Estimate[i])+ xax^2*(res$Estimate2[i])
  mean<- exp(mean)
  yvalues<- data.frame( xax=xax, mean=mean, status=res$Neophyte[i], taxa= res$taxa[i], n=res$numberOfPlots[i])
  matrix<- rbind(matrix, yvalues)
}


# Make dataframe
df<- as.data.frame(matrix)
# change order 
df$status <- factor(df$status, levels = c("native", "native_intra", "intra", "extra"))

alpha_values <- c("native" = 0.5, "native_intra" = 1, "intra" = 1, "extra" = 1)

# Plot with adjusted alpha values
ggplot(df, aes(x = xax, y = mean, color= status, group = taxa)) +
  geom_line(aes(alpha = status), linewidth = 0.5)+  # Set size of the lines
  scale_alpha_manual(values = alpha_values,name = "Legend",
                      breaks = c("native", "native_intra", "intra", "extra"),
                      labels = c("native species", "native species alien elsewhere", "intra European neophyte", "extra European neophyte")) +  # Set alpha values
  scale_colour_manual(values = c("#1E88E5", "#B71C1C", "#FFC107", "#388E3C"), 
                      name = "Legend",
                      breaks = c("native", "native_intra", "intra", "extra"),
                      labels = c("native species", "native species alien elsewhere", "intra European neophyte", "extra European neophyte"))
```



## 3.3 Squared mean
```{r}
df2<- df
status<- unique(df2$status)
for(i in status){
  sum<- sum(df2$n[df$status==i])
  df2$n[df$status==i]<- df2$n[df$status==i]/sum
}

df2<- df2 |> group_by(status, xax) |> summarise(mean= weighted.mean(mean, n))

# Plot
ggplot(df2, aes(x = xax, y = mean, color = status)) +
  geom_line(aes(color = status), linewidth=1) + 
  scale_fill_manual(values = c("#1E88E5", "#B71C1C", "#FFC107", "#388E3C"),
                    name = "Legend",
                    breaks = c("native", "native_intra", "intra", "extra"),
                    labels = c("native species", "native species alien elsewhere", "intra European neophyte", "extra European neophyte")) +
  scale_colour_manual(values = c("#1E88E5", "#B71C1C", "#FFC107", "#388E3C"), 
                      name = "Legend",
                      breaks = c("native", "native_intra", "intra", "extra"),
                      labels = c("native species", "native species alien elsewhere", "intra European neophyte", "extra European neophyte"))+theme_classic()+  scale_y_continuous(limits = c(0, 22))
```


## 3.4 Classes
Here we plot for when we use the all classes method (function 4)
```{r}
x <- x %>% mutate(Species = factor(Neophyte, 
                          levels = c("native", "intra","extra", "native_intra"),
                          labels = c("native in the country", 
                                     "intra-European neophyte", 
                                     "extra-European neophyte",
                                     "native in country alien elsewhere")))

obs_counts <- with(x, tapply(RelDiff, list(class, Species), length))

p<-ggplot(x, aes(x=class, y= RelDiff, color=Species, fill=Species))+geom_violin(alpha=0.5, position=position_dodge(1))+
  stat_summary(fun= "mean",
               geom = "point", aes(group= Neophyte), size=3, position = position_dodge(-1))+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"))+
  scale_fill_manual(values = c("#1E88E5", "#FFC107", "#004D40")) +
  theme_minimal()+ 
  theme(legend.position = "bottom")+theme(legend.text= element_text(size=10), 
                                          legend.title =element_text(size=12))+
  theme(axis.text.x = element_text(size=10), axis.title.y=element_text(size=12) ,axis.title.x=element_text(size=12))+
  ylab("Relative difference \n compared to absence species") + xlab("Cover class")+
  geom_boxplot(width=0.2, position = position_dodge(1), color="black", alpha=0.1)
p

all.equal(x,y)
#ggsave("Impact_base_model_classes.jpeg", p, width = 25, height = 15, units = "cm") 
```

I also wanted to include the relative amount of species in this calculation, since a lot more of extra species reach the highest cover class
```{r}
y<- x[!x$n==0,]
test <- y |> group_by(Neophyte, class) |> summarise(n= length(unique(taxa))) |> mutate(rel= n/max(n))

y<-y|> left_join(test, by=c("Neophyte"="Neophyte", "class"="class"))|> mutate(result= RelDiff*rel)

p<-ggplot(y, aes(x=class, y= result, color=Species, fill=Species))+geom_violin(alpha=0.5, position=position_dodge(1))+
  stat_summary(fun= "mean",
               geom = "point", aes(group= Neophyte), size=3, position = position_dodge(-1))+
  scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"))+
  scale_fill_manual(values = c("#1E88E5", "#FFC107", "#004D40")) +
  theme_minimal()+ 
  theme(legend.position = "bottom")+theme(legend.text= element_text(size=10), 
                                          legend.title =element_text(size=12))+
  theme(axis.text.x = element_text(size=10), axis.title.y=element_text(size=12) ,axis.title.x=element_text(size=12))+
  ylab("Relative difference \n compared to absence species") + xlab("Cover class")+
  geom_boxplot(width=0.2, position = position_dodge(1), color="black", alpha=0.1)
p
```


Some tests to get regression value
```{r}
library(lme4)
library(broom)
library(sjPlot)

MDL<- lmer(RelDiff ~ class+ (1|Neophyte), data=y)
summary(MDL)


plot_model(MDL,type="pred",terms= c("class","Neophyte"), pred.type="re")

# Extract coefficients using broom::tidy
coefficients <- tidy(MDL)

# Plot lines for each neophyte class
ggplot(coefficients, aes(x = class, y = estimate)) +
  geom_line() +
  labs(x = "Numeric Value of Cover Class", y = "Coefficient Estimate") +
  scale_color_manual(values = c("red", "blue", "green", "purple")) +  # Adjust color palette as needed
  theme_minimal()


plot(MDL$, MDL$fitted.values)

test <- test|> mutate(value= )

test <- y[y$Neophyte=="native",]
MDL<-lm(RelDiff ~ class, data=test)
plot( MDL$coefficients)

plot(MDL$class$, MDL$fitted.values)
```




## 3.5 All classes
```{r}
# This is data for which you have to run 7.1 a part to obtain the dataset 'taxaforCoveranalysis'
others<- x[x$taxa %in% taxaForCoverAnalysis$names,]

# new status delineation
status<- unique(others$status)

# mean and ci calculation
matrix<-c()
for(i in status){
  yvalues<- data.frame(n= numeric(), xax=numeric(), mean = numeric(), sd= numeric(), status =character())
  n<- length(others$Estimate[others$Neophyte==i])
  mean<-xaxis*mean(others$Estimate[others$Neophyte==i])+xaxis^2*mean(others$Estimate2[others$Neophyte==i])
  sd<-xaxis*sd(others$Estimate[others$Neophyte==i])+ xaxis^2*sd(others$Estimate2[others$Neophyte==i])
  ci<- qt(p=1-0.05/2, df=n-1)*sd/sqrt(n)
  yvalues<- data.frame(n=n, xax=xaxis, mean=mean, sd=sd, status=i)
  matrix<- rbind(matrix, yvalues)
}

# dataframe
df<- as.data.frame(matrix)

# plot
ggplot(df, aes(x = xax, y = mean, color = status)) +
  geom_line(aes(color = status), linewidth=0.1)+
  geom_ribbon(aes(ymin = mean - ci, ymax = mean + ci, fill = status), alpha = 0.1, color = NA) +
  scale_fill_manual(values = c("#1E88E5", "#B71C1C", "#FFC107", "#388E3C"),
                    name = "Legend",
                    breaks = c("native", "native_intra", "intra", "extra"),
                    labels = c("native species", "native species alien elsewhere", "intra European neophyte", "extra European neophyte")) +
  scale_colour_manual(values = c("#1E88E5", "#B71C1C", "#FFC107", "#388E3C"), 
                      name = "Legend",
                      breaks = c("native", "native_intra", "intra", "extra"),
                      labels = c("native species", "native species alien elsewhere", "intra European neophyte", "extra European neophyte"))
```




```{r}
 # Copy dataset temporarily to be able to make changes
  tmpFullPlotData <- fullPlotData
  
  # Select only explanatory variables
  tmpFullPlotData <- tmpFullPlotData %>% 
    select(numberOfVascularPlantSpecies,PlotObservationID, Area, 
           EIVEresM, EIVEresN, EIVEresR, EIVEresL, EIVEresT, 
           DistSeverity.sqrt, Soil.Disturbance.sqrt, 
           Grazing.Pressure.sqrt, Mowing.Frequency.sqrt, 
           Latitude, Longitude, Dataset, totalCover, chelsaP, hfp, PC1,PC2,PC3,PC4)
  
  # Copy the full dataset without aberrant totalCover
  tmpFullPlotData <- tmpFullPlotData %>% filter(totalCover <= 900)
  reducedEva <- eva %>% filter(PlotObservationID %in% tmpFullPlotData$PlotObservationID)

  # Get the indices of the plant in the eva dataset
  indexOfPlant <- reducedEva$name == plantName & reducedEva$Neophyte== plantStatus
  
  # We select only the rows containing the species first because this is much faster than doing it for all plots.
  # We make the dataset smaller (only those plots where the species is present), group the data per plot and compute the total cover (all     species) and cover of the species we are highlighting now
  plotsWherePlantOccurs <- reducedEva[reducedEva$PlotObservationID %in%
                                      reducedEva$PlotObservationID[indexOfPlant], ]|>
    group_by(PlotObservationID) |>
    summarise(speciesCover = computeSpeciesCover(plantName, name, `Cover %`))
 
  
  
  # Join the computed cover plots with the full data
  tmpFullPlotData <- left_join(tmpFullPlotData, plotsWherePlantOccurs, 
                             by = "PlotObservationID")
  
  # Identify all plots where the species investigated occurs
  tmpFullPlotData$plantOccurs = tmpFullPlotData$PlotObservationID %in% reducedEva$PlotObservationID[indexOfPlant]
  
  
  # reduce the dataset to obtain only those plots in which the species is present
  species_plots <- tmpFullPlotData[tmpFullPlotData$plantOccurs, c(19:22)  ]
  # make matrix and reduce duplicated columns
  species_plots<- as.matrix(species_plots)
  duplicated<- duplicated(species_plots)
  species_plots<- species_plots[!duplicated, ]
  
  # take ConvexHull around the PCA coordinates in which the species is present
  convex_hull <- convhulln((species_plots))
  # Extract all points from the dataset that are additionally within this ConvexHull and do not contain the species
  is_within_hull <- inhulln(convex_hull, as.matrix(tmpFullPlotData[, c(19:22)]))
  subset_data <- tmpFullPlotData[is_within_hull, ]
  

  # Take size of subset data
  size<- length(subset_data$plantOccurs)
  # change name tmpFullPlotData
  tmpFullPlotData <- subset_data
  
  # If species is not present:
  # Make 1 total cover and 0 species cover
  tmpFullPlotData$totalCover[is.na(tmpFullPlotData$totalCover)] <- 100.0
  tmpFullPlotData$speciesCover[is.na(tmpFullPlotData$speciesCover)] <- 0.0
  
  # Exclude all plots which have 0 total cover
  tmpFullPlotData <- tmpFullPlotData[tmpFullPlotData$totalCover >0,] 
  
  # Do SR - 1 for all plots in which the species is present
  tmpFullPlotData$numberOfVascularPlantSpecies <- tmpFullPlotData$numberOfVascularPlantSpecies - tmpFullPlotData$plantOccurs
  
  correctArea= F
  # Correct area: if true we calculate the area that is reserved for the other plant species. 
  if(correctArea) {
    tmpFullPlotData$Area <- tmpFullPlotData$Area*(1.0 - tmpFullPlotData$speciesCover/tmpFullPlotData$totalCover)
  }

  # There is a numerical problem in the bam function if we provide empty plots...
  tmpFullPlotData <- tmpFullPlotData |> filter(numberOfVascularPlantSpecies > 0, Area > 0)
  
  # Make dataset a factor to add it as a random variable
  tmpFullPlotData$Dataset <- as.factor(tmpFullPlotData$Dataset)
  
  # Add relative species cover
  tmpFullPlotData <- tmpFullPlotData %>% 
    mutate(relSpeciesCover = 100 * speciesCover / totalCover)
  
  
  # Species occurrence as factor (not logical)
  tmpFullPlotData$plantOccursF <- factor(tmpFullPlotData$plantOccurs)
  
```
