==> 9.2_Traits.Rmd <==
---
title: "Functional traits"
author: "Willem Goossens"
date: "`r Sys.Date()`"
output: html_document
---


# 1 Load
```{r}
#rm(list=ls())
```


```{r}
library(readr)
library(dplyr)
library(lmerTest)
library(lme4)
library(plantlist)
library(ggpubr)
library(emmeans)
library(vegdata)
library(ggeffects)
```

# 2 Data
## 2.1 Load
```{r}
# Load data
Diaz <- readxl::read_excel("../Extra data/Traits/Diaz.xlsx", guess_max = min(2000, n_max = NULL))


eva <- read_csv("../Eva data/fullPlotEva_new.csv", show_col_types = FALSE)
fullPlotData <- read_csv("../Eva data/fullPlotData_new.csv", show_col_types = FALSE)
fullPlotData <- fullPlotData[fullPlotData$PlotObservationID %in% eva$PlotObservationID,]

fast <- F
if(fast) {
  fullPlotData <- fullPlotData[runif(length(fullPlotData$PlotObservationID)) > 0.99,]
  eva <- eva[eva$PlotObservationID %in% fullPlotData$PlotObservationID,]
}
```


```{r}
# vector with species names
Diaz_names<- Diaz$`Species name standardized against TPL`
Diaz_names<- as.data.frame(Diaz_names)
# vector with eva names
eva_names<- as.data.frame(unique(eva$`Matched concept`))
colnames(eva_names) <- "Matched concept"
eva_names$name <- eva$name[match( eva_names$`Matched concept`, eva$`Matched concept`)]
eva_names$species <- eva$species[match( eva_names$`Matched concept`, eva$`Matched concept`)]

# check how much species in our dataset possess a trait value in Diaz
sum(unique(eva_names$name) %in% Diaz_names$Diaz_names)
sum(eva_names$name %in% Diaz_names$Diaz_names)
```



# 3 Analyse
## 3.1 Join names
```{r}
eva_names <- eva |> group_by(name, irena, `Matched concept`, species, `Turboveg2 concept`) |> summarise(n=n())

# name
Diaz$name <- eva_names$name[match(Diaz$`Species name standardized against TPL`, eva_names$name)]
sum(!is.na(Diaz$name))
# remove aggr.
Diaz$name[is.na(Diaz$name)] <- eva_names$name[match(Diaz$`Species name standardized against TPL`[is.na(Diaz$name)], gsub(" aggr\\.", "", eva_names$name))]
sum(!is.na(Diaz$name))
# vegdata
Diaz$name[is.na(Diaz$name)] <- eva_names$name[match(Diaz$`Species name standardized against TPL`[is.na(Diaz$name)], vegdata::taxname.abbr((eva_names$name)))]
sum(!is.na(Diaz$name))
Diaz$name[is.na(Diaz$name)] <- eva_names$name[match(Diaz$`Species name standardized against TPL`[is.na(Diaz$name)], vegdata::taxname.simplify((eva_names$name)))]
sum(!is.na(Diaz$name))

# check species
Diaz$name[is.na(Diaz$name)] <- eva_names$name[match(Diaz$`Species name standardized against TPL`[is.na(Diaz$name)], eva_names$species)]
sum(!is.na(Diaz$name))
Diaz$name[is.na(Diaz$name)] <- eva_names$name[match(Diaz$`Species name standardized against TPL`[is.na(Diaz$name)], eva_names$irena)]
sum(!is.na(Diaz$name))
Diaz$name[is.na(Diaz$name)] <- eva_names$name[match(Diaz$`Species name standardized against TPL`[is.na(Diaz$name)], eva_names$`Matched concept`)]
sum(!is.na(Diaz$name))
Diaz$name[is.na(Diaz$name)] <- eva_names$name[match(Diaz$`Species name standardized against TPL`[is.na(Diaz$name)], eva_names$`Turboveg2 concept`)]
sum(!is.na(Diaz$name))

Diaz_paste <- Diaz[, c(33,16,20,26,24,18,22,2,32, 10,11,6,5)]
colnames(Diaz_paste) <- c("name","LA","LMA","SSD","SM","Nmass","H","Diaz_name","number","woody","growth","Family","Genus")

Diaz_paste <- Diaz_paste[!is.na(Diaz_paste$name),]
dup <- Diaz_paste[duplicated(Diaz_paste$name) | duplicated(Diaz_paste$name, fromLast=T),]

Diaz2 <- Diaz |> group_by(name) |> summarise(LA = mean(`Leaf area (mm2)`, na.rm=T), LMA = mean(`LMA (g/m2)`, na.rm=T), 
                                             SSD= mean(`SSD combined (mg/mm3)`, na.rm=T), SM = mean(`Diaspore mass (mg)`, na.rm=T), 
                                             Nmass = mean(`Nmass (mg/g)`, na.rm=T), H = mean(`Plant height (m)`, na.rm=T),
                                             Diaz_name =  names(sort(table((`Species name standardized against TPL`)), decreasing=T)[1]), 
                                             number = max(`Number of traits with values`))

Diaz2 <- Diaz2[!is.na(Diaz2$name),]
hist(as.integer( Diaz2$number))
sum(Diaz2$number == 6)
sum(Diaz2$number == 5)
sum(Diaz2$number == 4)
sum(Diaz2$number == 3)
sum(Diaz2$number == 2)
sum(Diaz2$number == 1)

sum(Diaz2$number)/(length(Diaz2$name)*6)
sum(!is.na(Diaz2[,2:7]))/(nrow(Diaz2)*6)

length(unique(eva$name))

extra <- Diaz[!is.na(Diaz$Woodiness),]
Diaz2$woody <- extra$Woodiness[match(Diaz2$name, extra$name)]

extra <- Diaz[!is.na(Diaz$`Growth Form`),]
Diaz2$growth <- extra$`Growth Form`[match(Diaz2$name, extra$name)]
Diaz2$Family <- Diaz$Family[match(Diaz2$name, Diaz$name)]
Diaz2$Genus <- Diaz$Genus[match(Diaz2$name, Diaz$name)]

eva_names <- left_join(eva_names, Diaz2, by= c("name"="name"))
eva_names2 <- eva_names
sum(!is.na(eva_names$Diaz_name))/length(eva_names$Diaz_name)

#write_csv(eva_names, "../Extra data/Traits/Eva_diaz.csv")
```

This was a test but does not make sense, here we get less traits
```{r}
eva_names <- eva |> group_by(name, irena, `Matched concept`, species, `Turboveg2 concept`) |> summarise(n=n())

eva_names$Diaz <- Diaz$`Species name standardized against TPL`[match( gsub(" aggr\\.", "", eva_names$name), Diaz$`Species name standardized against TPL`)]
sum(!is.na(eva_names$Diaz))
eva_names$Diaz[is.na(eva_names$Diaz)] <- Diaz$`Species name standardized against TPL`[match(vegdata::taxname.abbr((eva_names$name[is.na(eva_names$Diaz)])), 
                                                                                            Diaz$`Species name standardized against TPL`)]
sum(!is.na(eva_names$Diaz))
eva_names$Diaz[is.na(eva_names$Diaz)] <- Diaz$`Species name standardized against TPL`[match(vegdata::taxname.simplify((eva_names$name[is.na(eva_names$Diaz)])), 
                                                                                            Diaz$`Species name standardized against TPL`)]
sum(!is.na(eva_names$Diaz))
eva_names$Diaz[is.na(eva_names$Diaz)] <- Diaz$`Species name standardized against TPL`[match(eva_names$species[is.na(eva_names$Diaz)], 
                                                                                            Diaz$`Species name standardized against TPL`)]
sum(!is.na(eva_names$Diaz))
eva_names$Diaz[is.na(eva_names$Diaz)] <- Diaz$`Species name standardized against TPL`[match(eva_names$`Matched concept`[is.na(eva_names$Diaz)], 
                                                                                            Diaz$`Species name standardized against TPL`)]
sum(!is.na(eva_names$Diaz))
eva_names$Diaz[is.na(eva_names$Diaz)] <- Diaz$`Species name standardized against TPL`[match(eva_names$irena[is.na(eva_names$Diaz)], 
                                                                                            Diaz$`Species name standardized against TPL`)]
sum(!is.na(eva_names$Diaz))
eva_names$Diaz[is.na(eva_names$Diaz)] <- Diaz$`Species name standardized against TPL`[match(eva_names$`Turboveg2 concept`[is.na(eva_names$Diaz)], 
                                                                                            Diaz$`Species name standardized against TPL`)]
sum(!is.na(eva_names$Diaz))

colnames(eva_names2)
colnames(Diaz)
Diaz_paste <- Diaz[, c(16,20,26,24,18,22,2,32, 10,11,6,5)]
colnames(Diaz_paste) <- c("LA","LMA","SSD","SM","Nmass","H","Diaz_name","number","woody","growth","Family","Genus")
eva_names <- left_join(eva_names, Diaz_paste, by= c("Diaz"="Diaz_name"))
  
sum(!is.na(eva_names$Diaz))/length(eva_names$Diaz)
```

Numbers
```{r}
eva_names <- eva_names2
sum(is.na(eva_names[, c(7:12)]))/6/length(eva_names$name)

eva_names <- eva_names %>% group_by(name)  %>%  
  summarise(LA= mean(LA, na.rm=T), LMA= mean(LMA, na.rm=T), SSD= mean(SSD, na.rm=T), SM = mean(SM, na.rm=T), Nmass= mean(Nmass, na.rm=T), H= mean(H, na.rm=T),
            #Diaz_name =  names(sort(table((Diaz_name)), decreasing=T)[1]),
            #woody = names(sort(table((woody)), decreasing=T)[1]),
            #growth =  names(sort(table((growth)), decreasing=T)[1]),
            #Genus =  names(sort(table((Genus)), decreasing=T)[1]),
            #Family =  names(sort(table((Family)), decreasing=T)[1])
            )
```



## 3.2 Traits
```{r}
mean <- read_csv("../Results/trait_normalized_new.csv")

eva_names$LA[is.na(eva_names$LA)] <- mean$LA[match(eva_names$name[is.na(eva_names$LA)], mean$SpeciesName)]
eva_names$LMA[is.na(eva_names$LMA)] <- mean$LMA[match(eva_names$name[is.na(eva_names$LMA)], mean$SpeciesName)]
eva_names$SSD[is.na(eva_names$SSD)] <- mean$SSD[match(eva_names$name[is.na(eva_names$SSD)], mean$SpeciesName)]
eva_names$SM[is.na(eva_names$SM)] <- mean$SM[match(eva_names$name[is.na(eva_names$SM)], mean$SpeciesName)]
eva_names$H[is.na(eva_names$H)] <- mean$H[match(eva_names$name[is.na(eva_names$H)], mean$SpeciesName)]
eva_names$Nmass[is.na(eva_names$Nmass)] <- mean$Nmass[match(eva_names$name[is.na(eva_names$Nmass)], mean$SpeciesName)]

```

```{r}
# standardize
eva_names[, c(2:7)] <- log(eva_names[,c(2:7)])
eva_names[, c(2:7)] <- scale(eva_names[, c(2:7)])

# remove outliers (see Diaz et al 2016 for number 4)
for(j in 2:7){
  tr <- as.matrix(eva_names[,j])
  print(sum(is.na(tr)))
  tr <- ifelse(tr <= -4, NA, tr)
  tr <- ifelse(tr >= 4, NA, tr)
  print(sum(is.na(tr)))
  eva_names[,j] <- as.data.frame(tr)
}
```


```{r}
eva_names$Diaz_name <- Diaz2$Diaz_name[match(eva_names$name, Diaz2$name)]
eva_names$growth <- Diaz2$growth[match(eva_names$name, Diaz2$name)]
eva_names$woody <- Diaz2$woody[match(eva_names$name, Diaz2$name)]
eva_names$Family <- Diaz2$Family[match(eva_names$name, Diaz2$name)]
eva_names$Genus <- Diaz2$Genus[match(eva_names$name, Diaz2$name)]
```



```{r}
# give Diaz name
# eva
eva <- left_join(eva, eva_names[,c(1:12)], by= c("name"="name"))

# check number of tree species in our database
sum(eva_names$growth=="tree", na.rm=T)

sum(!is.na(eva$Diaz_name))/length((eva$Diaz_name))

length(unique(eva$name[!is.na(eva$Diaz_name)]))/length(unique(eva$name))
```



Evaluate number of observations
```{r}
# Join eva and mean trait database
# check number of missing values per trait
sum(is.na(eva$LA))
sum(is.na(eva$LMA))
sum(is.na(eva$H))
sum(is.na(eva$SM))
sum(is.na(eva$Nmass))
sum(is.na(eva$SSD))

sum(!is.na(eva_names$Diaz_name))/ length(eva_names$name)

# 60.1% of the species have their name in the Diaz database (this is not correct, check the 9.1 file --> we need to remove genus level observations)
length(unique(eva$name[!is.na(eva$Diaz_name)]))/length(unique(eva$name))
# 90.8% of the observations do have a value for at least one trait
sum(!is.na(eva$Diaz_name))/ length(eva$PlotObservationID)
```

Alien
```{r}
colnames(eva)
# Data on which species are neophytes
native_intra_analysis=F
if(native_intra_analysis){
  species_country_status<- read_csv("../EVA data/country_species_new.csv", show_col_types = FALSE)
} else{
  species_country_status<- read_csv("../EVA data/country_species_new.csv", show_col_types = FALSE)
  species_country_status$Neophyte[species_country_status$Neophyte=="native_intra"] <- "native"
# or if we want to do it with intra seperately
}
# Only observation ID and Region
fullPlot2<- fullPlotData[,c("PlotObservationID","Region")]
# Right join to keep only species present in fullplot (otherwise a lot of NAs)
eva_country_neophyte <- right_join(eva, fullPlot2, by = c("PlotObservationID"="PlotObservationID"))
# Join eva and classification
eva_country_neophyte<- left_join(eva_country_neophyte, species_country_status[, -c(2:4,6:7)], by= c("Region"= "Region", "name"= "name"))
# Look at how much from every type are present
table(eva_country_neophyte$Neophyte)

# there are still na's present --> due to being named 'plant'
nas<- eva_country_neophyte[is.na(eva_country_neophyte$Neophyte),]
eva_country_neophyte <- (eva_country_neophyte[!eva_country_neophyte$name=="Plant",])
test <- eva_country_neophyte |> group_by(name, Neophyte, Diaz_name) |> summarise(n= n())
test <- test |> group_by(Neophyte) |> summarise(n = sum(!(is.na(Diaz_name))))
```




## 3.3 Plot level
```{r}
colnames(eva)[9]<-"Cover %"
y<- eva |> group_by(PlotObservationID) |> summarise(cwm_SSD= weighted.mean(SSD, `Cover %`, na.rm=T), 
                                                     cwm_LMA= weighted.mean(LMA, `Cover %`, na.rm=T),
                                                     cwm_height = weighted.mean(H, `Cover %`, na.rm=T),
                                                     cwm_seed= weighted.mean(SM, `Cover %`, na.rm=T),
                                                     cwm_LA= weighted.mean(LA, `Cover %`, na.rm=T),
                                                     cwm_N =weighted.mean(Nmass, `Cover %`, na.rm=T))

# check correlation with plot level EIVE values
cor(y[,2:7], fullPlotData[,15:19],use="pairwise.complete.obs", method="pearson")
cor(y[,2:7], fullPlotData[,20:24],use="pairwise.complete.obs", method="pearson")

# we notice quite high correlations for EIVE N and leaf nitrogen content
# also high for EIVE L and SSD and Height
# Also dist frequency and height/SSD
```



# 4 Impact 
## 4.1 Explore
```{r}
# Load impact data
impact<- read.csv("I:/Impact_1980_final.csv")

# possibly only use non-tree species (can maybe influence calculation)
# Diaz <- Diaz |> filter(!`Growth Form`=="tree")


# remove genera
check <-vegdata::parse.taxa(unique(impact$taxa))
genus <- check[is.na(check$epi1),]
impact <- impact[!impact$taxa %in% genus$original,]
```


## 4.2 Join
```{r}
#save <- eva_names[, c(1,2:12)]
#write.csv(save, "../PNAS/traits.csv")
# Merge 7.1 and traits data
impact<- left_join(impact, eva_names[, c(1,2:12)], by=c("taxa"="name"))
test <- unique(impact[, c(1,8)])
test |> group_by(Neophyte) |> summarise(n=n())

# if relative add 13 in join
test <- impact |> group_by(taxa, Diaz_name) |> summarise(n=n())
sum(!is.na(test$Diaz_name))/length(test$Diaz_name)

# Calculate correlations
cor_extra<-cor(impact[impact$Neophyte=="extra", c(2,9, 14:19)],  use= "pairwise.complete.obs", method="spearman")
cor_intra <-cor(impact[impact$Neophyte=="intra", c(2,9,  14:19)],  use="pairwise.complete.obs", method="spearman")
cor_native <-cor(impact[impact$Neophyte=="native", c(2,9,  14:19)],  use="pairwise.complete.obs", method="spearman")
cor <- cor(impact[, c(2,9,  14:19)],  use="pairwise.complete.obs", method="spearman")
```


## 4.3 Phylogenetic
```{r}
phylo <- read.csv("../Extra data/Species names/phylo.csv")
impact[,23:24] <- phylo[match(impact$taxa, phylo$name), 7:6]

```


## 4.4 General impact
```{r}
impact <- cbind(impact, eva[match(impact$taxa, eva$name), c(11:15, 17:23)]) 
colnames(impact)

# Change names FT
colnames(impact)[30:36]<- c("Sev","Sev_herb","Freq","Freq_herb","Gras","Mow","Soil") 

# compute general impact
general <- impact |> 
  group_by(taxa, Neophyte, LA, Nmass, LMA, H, SM, SSD, Family, Genus, growth, woody, 
           EIVEresM, EIVEresT, EIVEresR, EIVEresL, EIVEresN, 
           Sev,Sev_herb,Freq,Freq_herb,Gras,Mow,Soil) |> 
  summarise(impact= sum((RelDiff*n)/numberOfPlots))
```

Change growth
```{r}
general$growth <- gsub("herbaceous non-graminoid","herb",general$growth)
general$growth <- gsub("herbaceous graminoid","graminoid",general$growth)
general$growth <- ifelse(general$H <5, gsub("shrub/tree","shrub",general$growth), gsub("shrub/tree","tree",general$growth))
general$growth <- ifelse(general$H <0.5, gsub("herb/shrub","herb",general$growth), gsub("herb/shrub","shrub",general$growth))
general$growth <- gsub("succulent","other",general$growth)
general$growth <- gsub("fern","other",general$growth)
general$growth <- gsub("climber","other",general$growth)

general$growth <- as.factor(general$growth)
general$growth <- relevel(general$growth, ref="herb")
```


# 5 Model
```{r}
# create new vector to only assess species that have all data
general2 <- general[!is.na(general$LA)& !is.na(general$LMA)& !is.na(general$SSD)& !is.na(general$SM)& !is.na(general$Nmass)& !is.na(general$H),]

# run this if you want to see which species we loose
lost_species <- general[!general$taxa %in% general2$taxa,]
```

## 5.1 PCA
```{r}
# test PCA
PCA <- prcomp(general2[, c(3:8)], scale=F)
eig.val <- factoextra::get_eigenvalue(PCA)
eig.val
biplot(PCA)
PCA
```

PCA rotation
```{r}
# if you want to rotate do it here
angle <- pi/8
rotation_matrix <- matrix(c(cos(angle), -sin(angle), sin(angle), cos(angle)), ncol=2)

# change scores
scores <- as.matrix(PCA$x[, 1:2]) %*% rotation_matrix
scores <- as.data.frame(scores)
scores$Neophyte <- general2$Neophyte
colnames(scores) <- c("PC1","PC2","Neophyte")

# change loadings
loadings <- as.matrix(PCA$rotation[,1:2])%*% rotation_matrix 
loadings<- as.data.frame(loadings)
colnames(loadings) <- c("PC1","PC2")
rownames(loadings)[2] <- c("Leaf N")

# mirror PC2 to resemble Diaz
scores$PC2 <- scores$PC2*-1
loadings$PC2 <- loadings$PC2*-1

rotated_variances <- apply(scores[1:2], 2, var)
total_variance <- sum(apply(as.matrix(PCA$x) , 2, var))
rotated_variances / total_variance

# assign axes to the data
general2$PC1 <- scores[,1]
general2$PC2 <- scores[,2]
```


```{r}
dev_1<- round(apply(scores[1], 2,var)/sum(PCA$sdev^2)*100, 1)
dev_1
dev_2 <- round(apply(scores[2], 2,var)/sum(PCA$sdev^2)*100, 1)
# Plot
PCA_plot <-ggplot(data = scores, aes(x = PC1, y = PC2, color = Neophyte)) +
  geom_point(aes(alpha= ifelse(Neophyte=="Native", 0.1, 0.5), shape = Neophyte), size = 2) +
  geom_segment(data = loadings, aes(x = 0, y = 0, xend = PC1*5, yend = PC2*5),
               arrow = arrow(length = unit(0.2, "cm")), color = "black") +
  geom_text(data = loadings, aes(x = PC1*5, y = PC2*5, label = rownames(loadings)),
            color = "black", vjust = -0.5, size=5) +
  theme_pubr() +
  xlab((expression(PC[size]*" ("* 37.7* "%)"))) +
  ylab((expression(PC[economics]* " ("* 29.1* "%)"))) +
  scale_colour_manual(values=c( "#004D40", "#FFC107","#1E88E5"))+
  scale_fill_manual(values = c( "#004D40", "#FFC107","#1E88E5"))+ 
  scale_alpha_identity() + 
  theme(legend.position="none")+ 
  labs(fill = "Origin", colour= "Origin", shape= "Origin")
PCA_plot

#ggsave("../Images/PCA.svg", PCA_plot, width=4, height=4)
```



## 5.2 GLMM
```{r}
general2 <- general2 %>% mutate(Neophyte = factor(Neophyte, 
                          levels = c("native","intra", "extra"))) 

# run model with PC1 (size) and PC2 (econ), also include the growth form
MDL<- lmer(impact ~ -1 +Neophyte +PC1* Neophyte + PC2* Neophyte   + woody+ (1|Family/Genus), general2)
anova(MDL)

# summarise and look at the differences
summary(MDL)

# compare
emmeans::emmeans(MDL, pairwise ~Neophyte)
emmeans::emmeans(MDL, pairwise ~Neophyte*PC1)

summary(emmeans::emtrends(MDL,  pairwise~ "Neophyte", var = "PC1"))
summary(emmeans::emtrends(MDL,  pairwise~ "Neophyte", var = "PC2"))

test(emtrends(MDL, ~ Neophyte, var="PC1"))
test(emtrends(MDL, ~ Neophyte, var="PC2"))

#emmeans(MDL, pairwise ~woody)
emmeans::emmeans(MDL, pairwise ~woody)

(anova(MDL))
library(MuMIn)
r.squaredGLMM(MDL)
```


```{r}
# glmm
model <- lmer(impact ~ Neophyte +PC1* Neophyte + PC2* Neophyte   + woody+ (1|Family/Genus), general2)
summary(model)

simulation <- DHARMa::simulateResiduals(model)
plot(simulation)

# check pairwise differences
emmeans_results <- emmeans::emmeans(model, pairwise ~ Neophyte)
summary(emmeans_results)

table(general2$Neophyte)

anova(model)

# we will 'bootstrap' or results using 1000 replications
result <- data.frame()

result_PC1 <- data.frame()
result_PC2 <- data.frame()
result_woody <- data.frame()


# Counter for successful runs
successful_runs <- 0

# Run until we get 1000 successful models
while (successful_runs < 1000) {
  # sample natives
  natives <- general2[general2$Neophyte=="native",]
  natives <- natives[sample(nrow(natives),200, replace=F),]
  # sample intras
  intras <- general2[general2$Neophyte=="intra",]
  intras <- intras[sample(nrow(intras),200, replace=T),]
  # sample extras
  extras <- general2[general2$Neophyte=="extra",]
  extras <- extras[sample(nrow(extras),200, replace=T),]
  all <- rbind(natives, intras, extras)
  
  # model again
  model<- lmer(impact ~ -1 +Neophyte +PC1* Neophyte + PC2* Neophyte   + woody+ (1|Family/Genus), all)
  summary(model)
  
  if(isSingular(model)) next
  
  sum <- test(emtrends(model, ~ Neophyte, var="PC1"))
  sum <- summary(sum)

  result_PC1 <- rbind(result_PC1, sum)
  
  sum <- test(emtrends(model, ~ Neophyte, var="PC2"))
  sum <- summary(sum)

  result_PC2 <- rbind(result_PC2, sum)
  
  sum <- emmeans(model, pairwise ~woody)  
  sum <- summary(sum)$contrasts

  result_woody <- rbind(result_woody, sum)
  
  # Check pairwise differences
  sum <- anova(model)
  sum$name <- rownames(sum)
  sum$singular <- isSingular(model)

  # Append to results dataframe
  result <- rbind(result, sum)
  
  successful_runs <- successful_runs+1
}

# look at aggregated results (mean and sd)
aggregated_result <- aggregate(. ~ name, data = result, FUN = function(x) c(mean = mean(x), sd = sd(x)))

# assess number of significant
result$significant <- result$`Pr(>F)` < 0.05
table(result$significant)

result_PC1$significant <- result_PC1$p.value < 0.05
table(result_PC1$significant)

result_PC2$significant <- result_PC2$p.value < 0.05
table(result_PC2$significant)

result_woody$significant <- result_woody$p.value < 0.05
table(result_PC2$significant)


test <- result |> group_by(name, significant) |> summarise(n=n(), diff= mean(`Sum Sq`), sd= sd(`Sum Sq`), se= sd/sqrt(n))
test

testPC1 <- result_PC1 |> group_by(Neophyte, significant) |> summarise(n=n(), diff= mean(`PC1.trend`), sd= sd(`PC1.trend`), se= sd/sqrt(n))
testPC1

testPC2 <- result_PC2 |> group_by(Neophyte, significant) |> summarise(n=n(), diff= mean(`PC2.trend`), sd= sd(`PC2.trend`), se= sd/sqrt(n))
testPC2

testwoody <- result_woody |> group_by(contrast, significant) |> summarise(n=n(), diff= mean(estimate), sd= sd(estimate), se= sd/sqrt(n))
testwoody
```




## 5.3 PLOT
```{r}
test <- ggpredict(MDL, terms= c("PC1","Neophyte"))

test$group <- gsub("native","Native",test$group)
test$group <- gsub("intra","Intra-European",test$group)
test$group <- gsub("extra","Extra-European",test$group)

general2$Neophyte <- gsub("native","Native",general2$Neophyte)
general2$Neophyte <- gsub("intra","Intra-European",general2$Neophyte)
general2$Neophyte <- gsub("extra","Extra-European",general2$Neophyte)

PC1<- ggplot(test, aes(x, predicted, colour= group))+
  geom_point(data= general2, aes(x= PC1, y= impact, colour=Neophyte, alpha= ifelse(Neophyte=="Native", 0.1, 0.3), shape = Neophyte))+
  geom_line(aes(linetype= group), size=1.5)+
  geom_ribbon(aes(ymin= conf.low, ymax= conf.high, colour= group, fill=group), alpha=0.25, linetype=0)+  
  theme_pubr()+
  xlab(expression(PC[size]))+
  ylab("Overall impact")+
  geom_hline(yintercept=0, linetype='dotted', alpha=0.5)+
  scale_colour_manual(values=c( "#004D40", "#FFC107","#1E88E5"))+
  scale_fill_manual(values = c( "#004D40", "#FFC107","#1E88E5"))+ 
  scale_linetype_manual(values=c("dashed","dashed","dashed" )) +
  guides(linetype = "none") + 
  labs(fill = "Origin", colour= "Origin")+ 
  scale_alpha_identity() + 
  theme(legend.position="none")+
  scale_y_continuous(limits=c(-1,1))+ 
  labs(fill = "Origin", colour= "Origin", shape= "Origin")
PC1

test <- ggpredict(MDL, terms= c("PC2","Neophyte"))

test$group <- gsub("native","Native",test$group)
test$group <- gsub("intra","Intra-European",test$group)
test$group <- gsub("extra","Extra-European",test$group)


PC2<- ggplot(test, aes(x, predicted, colour= group))+
  geom_point(data= general2, aes(x= PC2, y= impact, colour=Neophyte, alpha= ifelse(Neophyte=="Native", 0.1, 0.3), shape = Neophyte))+
  geom_line(aes(linetype= group),size=1.5)+
  geom_ribbon(aes(ymin= conf.low, ymax= conf.high, colour= group, fill=group), alpha=0.25, linetype=0)+  
  theme_pubr()+
  xlab(expression(PC[economics]))+
  ylab("Overall impact")+
  geom_hline(yintercept=0, linetype='dotted', alpha=0.5)+
  scale_colour_manual(values=c( "#004D40", "#FFC107","#1E88E5"))+
  scale_fill_manual(values = c( "#004D40", "#FFC107","#1E88E5"))+ 
  scale_linetype_manual(values=c("dashed","dashed","solid" )) +
  guides(linetype = "none") + 
  labs(fill = "Origin", colour= "Origin")+ 
  scale_alpha_identity() + 
  theme(legend.position="bottom")+
  scale_y_continuous(limits=c(-1,1))+ 
  labs(fill = "Origin", colour= "Origin", shape= "Origin")
PC2

growth = F
if(growth){
growth_neophyte=F
if(growth_neophyte){

test <- ggpredict(MDL, terms= c("growth", "Neophyte"))
test <- test[!is.na(test$group),]
test$group <- gsub("native","Native",test$group)
test$group <- gsub("intra","Intra-European",test$group)
test$group <- gsub("extra","Extra-European",test$group)
  
  
growth<- ggplot(test, aes(x, predicted, colour= group))+
  geom_violin(data= general2[!is.na(general2$growth),], aes(x= growth, y= impact, colour=Neophyte, fill=Neophyte, alpha= 0.3),  
              position = position_dodge(width=0.8))+
  geom_point(size=2, position = position_dodge2(width = 0.8, preserve = "single")) +
  geom_errorbar(aes(ymin= conf.low, ymax= conf.high, colour= group), position = position_dodge(width = 0.8))+
  theme_pubr()+
  xlab("Life form")+
  ylab("Overall impact")+
  geom_hline(yintercept=0, linetype='dotted', alpha=0.5)+
  scale_colour_manual(values=c( "#004D40", "#FFC107","#1E88E5"))+
  scale_fill_manual(values = c( "#004D40", "#FFC107","#1E88E5"))+ 
  labs(fill = "Origin", colour= "Origin")+ 
  scale_alpha_identity() + 
  theme(legend.position="none")
} else {

test <- ggpredict(MDL, terms= c("growth"))
test <- test[!is.na(test$group),]  
test$x <- as.factor(test$x)

growth<- ggplot(test, aes(x= factor(x, level= c("herb","graminoid","tree","shrub","other")), predicted), show.legend = FALSE)+
  geom_violin(data= general2[!is.na(general2$growth),], aes(x= factor(growth, level= c("herb","graminoid","tree","shrub","other")), 
                                                            y= impact, alpha= 0.3,
                                                            fill= "lightgrey"),
              position = position_dodge(width=0.8))+
  geom_point(size=2, position = position_dodge2(width = 0.8, preserve = "single")) +
  geom_errorbar(aes(ymin= conf.low, ymax= conf.high), position = position_dodge(width = 0.8))+
  theme_pubr()+
  xlab("Life form")+
  ylab("Overall impact")+
  geom_hline(yintercept=0, linetype='dotted', alpha=0.5)+
  scale_alpha_identity() + 
  theme(legend.position="none")+
  theme(axis.title.x=element_blank())+
  #scale_y_continuous(name= "Overall impact", limits= c(-1, max(general2$impact)+0.01))+
  # Significance letters (lower position)
  annotate("text", x=1, y=max(general2$impact), label="a", size=4, hjust=0.5, alpha=1, vjust= -0.3) +
  annotate("text", x=2, y=max(general2$impact), label="ab", size=4, hjust=0.5, alpha=1, vjust= -0.3) +
  annotate("text", x=3, y=max(general2$impact), label="ab", size=4, hjust=0.5, alpha=1, vjust= -0.3)+
  annotate("text", x=4, y=max(general2$impact), label="b", size=4, hjust=0.5, alpha=1, vjust= -0.3)+
  annotate("text", x=5, y=max(general2$impact), label="b", size=4, hjust=0.5, alpha=1, vjust= -0.3)

growth
}
} else {
  
test <- ggpredict(MDL, terms= c("woody"))
test <- test[!is.na(test$group),]  
test$x <- as.factor(test$x)
test$x
growth<- ggplot(test, aes(x= factor(x, level= c("non-woody","semi-woody","woody")), predicted), show.legend = FALSE)+
  geom_violin(data= general2[!is.na(general2$woody),], aes(x= factor(woody, level= c("non-woody","semi-woody","woody")), 
                                                            y= impact, alpha= 0.8,
                                                            fill= "#D81B60", color= "#D81B60"),
              position = position_dodge(width=0.8))+
  geom_point(size=2, position = position_dodge2(width = 0.8, preserve = "single")) +
  geom_errorbar(aes(ymin= conf.low, ymax= conf.high), position = position_dodge(width = 0.5), width= 0.5)+
  theme_pubr()+
  xlab("Woodiness")+
  ylab("Overall impact")+
  geom_hline(yintercept=0, linetype='dotted', alpha=0.5)+
  scale_alpha_identity() + 
  theme(legend.position="none")+
  theme(axis.title.x=element_blank())+
  #scale_y_continuous(name= "Overall impact", limits= c(-1, max(general2$impact)+0.01))+
  # Significance letters (lower position)
  annotate("text", x=1, y=max(general2$impact), label="a", size=4, hjust=0.5, alpha=1, vjust= -0.3) +
  annotate("text", x=2, y=max(general2$impact), label="ab", size=4, hjust=0.5, alpha=1, vjust= -0.3) +
  annotate("text", x=3, y=max(general2$impact), label="b", size=4, hjust=0.5, alpha=1, vjust= -0.3)+ 
  guides(fill = "none", color = "none", linetype = "none", shape = "none")

growth
}

test <- ggpredict(MDL, terms= c("Neophyte"))
test$x <- as.factor(test$x)
general2$Neophyte <- as.factor(general2$Neophyte)

# plot
p<-ggplot(general2, aes(x=Neophyte, y= impact, group = Neophyte,
               color = Neophyte, fill = Neophyte))+
    geom_violin(alpha=0.5, scale="width")+
    geom_boxplot(width= 0.25, alpha=0.8, fill="white")+
    theme_pubr()+
    stat_summary(fun= "mean",
                 geom = "point", aes(group= Neophyte), size=3)+
    scale_colour_manual(values=c("#1E88E5", "#FFC107", "#004D40"))+
    scale_fill_manual(values = c("#1E88E5", "#FFC107", "#004D40")) +
    theme(legend.position = "none")+
    theme(legend.text= element_text(size=12),legend.title =element_text(size=14))+
    annotate("text", x = 1, y = max(general2$impact)+0.1,label = "b", size = 4, vjust = 0.5, hjust = 0.5, alpha=1) +
    annotate("text", x = 2, y = max(general2$impact)+0.1, label = "b", size = 4, vjust = 0.5, hjust = 0.5, alpha=1)+
    annotate("text", x = 3, y = max(general2$impact)+0.1, label = "a", size = 4, vjust = 0.5, hjust = 0.5, alpha=1)+
    ylab("Overall impact")+
    xlab(NULL)+
    geom_hline(yintercept=0, linetype='dotted', alpha=0.5)+
    theme(plot.title= element_text(hjust=0.01),
    plot.subtitle = element_text( hjust = 0.01))
p

PC2

plot <- ggarrange(PC1, PC2, growth, labels= c("a","b","c"), common.legend = TRUE, legend="bottom", align="hv", nrow=1, ncol=3)

plot

#ggsave("../Images/Traits.svg", plot, width=12, height=4)

```



```{r}
plot_all <- ggarrange( PCA_plot, growth, PC1, PC2, labels= c("a","b","c","d","e"), common.legend = TRUE, 
                       legend="bottom", align="hv", nrow=2, ncol=2)

plot_all


ggsave("../Images/Drivers.svg", plot_all, width=8, height=8)
```

```{r}

legend_PC <- ggpubr::get_legend(PC2)
as_ggplot(legend_PC)

plot_all <- ggarrange( ggarrange(PCA_plot, plot_daisie, labels= c("a","b"), ncol=2, widths = c(1,2),common.legend = TRUE, align="hv", legend=F), 
                       ggarrange(growth, PC1, PC2, labels= c("c","d","e"),ncol=3,common.legend = TRUE, align="hv", legend = "none"),
                       common.legend = TRUE,  legend="bottom", align="hv", nrow=2, legend.grob = legend_PC)
plot_all

ggsave("../Images/Drivers_daisie.png", plot_all, width=12, height=8)


legend_PC <- ggpubr::get_legend(PC2)
as_ggplot(legend_PC)

plot_all <- ggarrange(ggarrange( PCA_plot, growth, PC1, PC2, labels= c("a","b","c","d"), common.legend = TRUE, 
                      legend="none", align="hv", nrow=2, ncol=2),
                      ggarrange(plot_daisie, nrow=1, ncol=1, legend= "none", labels= c("e")),
                      common.legend = TRUE,  legend="bottom", align="hv", nrow=2, legend.grob = legend_PC, heights= c(1.5,0.5))
plot_all

ggsave("../Images/Drivers_daisie.png", plot_all, width=8, height=10)
```
